Refactor glass catalog data structure
