+++ src/optics/glassCatalogAliases.ts
@@ -1,52 +1,131 @@
 /**
- * Informal glass-name aliases resolved by glassCatalog.ts.
+ * Informal and historical glass-name aliases resolved by glassCatalog.ts.
  */
 
+export type GlassAliasKind = "informal" | "legacy-name" | "vendor-equivalent" | "catalog-variant" | "class-surrogate";
+
+export interface GlassAliasRecord {
+  readonly alias: string;
+  readonly target: string;
+  readonly kind: GlassAliasKind;
+  readonly note: string;
+}
+
+export const ALIAS_RECORDS: readonly GlassAliasRecord[] = [
+  { alias: "BK7", target: "N-BK7", kind: "informal", note: "Common shorthand for the Schott N-BK7 catalog row." },
+  {
+    alias: "BSC7",
+    target: "S-BSL7",
+    kind: "vendor-equivalent",
+    note: "Hoya crown shorthand used as an S-BSL7-class match.",
+  },
+  { alias: "FLUORITE", target: "CaF2", kind: "informal", note: "Material name for the cataloged CaF2 crystal entry." },
+  { alias: "CAF2", target: "CaF2", kind: "informal", note: "Uppercase token form of the CaF2 crystal entry." },
+  {
+    alias: "SF57",
+    target: "S-TIH53",
+    kind: "vendor-equivalent",
+    note: "Schott high-dispersion family name sharing the 847/238 code class with S-TIH53.",
+  },
+  {
+    alias: "N-SF57",
+    target: "S-TIH53",
+    kind: "vendor-equivalent",
+    note: "Modern Schott high-dispersion family name sharing the 847/238 code class with S-TIH53.",
+  },
+  {
+    alias: "H-ZF88A",
+    target: "S-TIH53",
+    kind: "class-surrogate",
+    note: "Laowa annotations use the 847/238 dense-flint class, not CDGM's distinct H-ZF88 row.",
+  },
+  { alias: "SK16", target: "N-SK16", kind: "legacy-name", note: "Legacy Schott name superseded by N-SK16." },
+  {
+    alias: "BACD5",
+    target: "N-SK16",
+    kind: "vendor-equivalent",
+    note: "Hoya trade name for N-SK16-equivalent optical constants.",
+  },
+  {
+    alias: "L-TIM28",
+    target: "S-TIM28",
+    kind: "catalog-variant",
+    note: "Ohara large-format designation with the same published optical constants as S-TIM28.",
+  },
+  {
+    alias: "L-BAL42",
+    target: "S-BAL42",
+    kind: "catalog-variant",
+    note: "Ohara low-softening / PGM variant sharing the S-BAL42 optical constants.",
+  },
+  {
+    alias: "L-BSL7",
+    target: "S-BSL7",
+    kind: "catalog-variant",
+    note: "Ohara low-softening / PGM variant sharing the S-BSL7 optical constants.",
+  },
+  {
+    alias: "L-BAL35",
+    target: "S-BAL35",
+    kind: "catalog-variant",
+    note: "Ohara low-softening / PGM variant sharing the S-BAL35 optical constants.",
+  },
+  {
+    alias: "L-LAH53",
+    target: "S-LAH53",
+    kind: "catalog-variant",
+    note: "Ohara low-softening / PGM variant sharing the S-LAH53 optical constants.",
+  },
+  {
+    alias: "L-LAM60",
+    target: "S-LAM60",
+    kind: "catalog-variant",
+    note: "Ohara low-softening / PGM variant sharing the S-LAM60 optical constants.",
+  },
+  {
+    alias: "L-LAL13",
+    target: "S-LAL13",
+    kind: "catalog-variant",
+    note: "Ohara low-softening / PGM variant sharing the S-LAL13 optical constants.",
+  },
+  {
+    alias: "L-LAL14",
+    target: "S-LAL14",
+    kind: "catalog-variant",
+    note: "Ohara low-softening / PGM variant sharing the S-LAL14 optical constants.",
+  },
+  {
+    alias: "L-PHM52",
+    target: "S-PHM52",
+    kind: "catalog-variant",
+    note: "Ohara low-softening / PGM variant sharing the S-PHM52 optical constants.",
+  },
+  { alias: "SK5", target: "N-SK5", kind: "legacy-name", note: "Legacy Schott name superseded by N-SK5." },
+  { alias: "SK10", target: "N-SK10", kind: "legacy-name", note: "Legacy Schott name superseded by N-SK10." },
+  { alias: "SK14", target: "N-SK14", kind: "legacy-name", note: "Legacy Schott name superseded by N-SK14." },
+  { alias: "N-F2", target: "F2", kind: "legacy-name", note: "Route N-F2 annotations to the cataloged F2 row." },
+  { alias: "N-SF1", target: "SF1", kind: "legacy-name", note: "Route N-SF1 annotations to the cataloged SF1 row." },
+  { alias: "N-SF6", target: "SF6", kind: "legacy-name", note: "Route N-SF6 annotations to the cataloged SF6 row." },
+  { alias: "SF8", target: "N-SF8", kind: "legacy-name", note: "Legacy Schott name superseded by N-SF8." },
+  { alias: "FK5", target: "N-FK5", kind: "legacy-name", note: "Legacy Schott name superseded by N-FK5." },
+  { alias: "LAF21", target: "N-LAF21", kind: "legacy-name", note: "Legacy Schott name superseded by N-LAF21." },
+  { alias: "LA-F21", target: "N-LAF21", kind: "legacy-name", note: "Alternate legacy token for N-LAF21." },
+  { alias: "SK2", target: "N-SK2", kind: "legacy-name", note: "Legacy Schott name superseded by N-SK2." },
+  {
+    alias: "TAFD35L",
+    target: "TAFD35",
+    kind: "catalog-variant",
+    note: "Hoya PGM/pressed-glass row sharing TAFD35 911/353 optical constants.",
+  },
+  {
+    alias: "BSC3",
+    target: "E-C3",
+    kind: "vendor-equivalent",
+    note: "Historical Hoya crown annotations use the 518/590 E-C3 row; incompatible barium rows stay unmatched.",
+  },
+];
+
 /** Aliases that route to a canonical entry. Common informal names. */
-export const ALIASES: ReadonlyMap<string, string> = new Map([
-  ["BK7", "N-BK7"],
-  ["BSC7", "S-BSL7"],
-  ["FLUORITE", "CaF2"],
-  ["CAF2", "CaF2"],
-  // SF57 and N-SF57 are Schott names for the same high-dispersion glass family;
-  // S-TIH53 is the Ohara equivalent sharing the same 6-digit code (847/238 ≈ 847/239).
-  ["SF57", "S-TIH53"],
-  ["N-SF57", "S-TIH53"],
-  // CDGM H-ZF88A annotations in Laowa Probe data use the 847/238 dense-flint
-  // class; CDGM's cataloged H-ZF88 is a different 946/179 glass.
-  ["H-ZF88A", "S-TIH53"],
-  // SK16 is the legacy Schott name for N-SK16 (same glass, modernized designation).
-  ["SK16", "N-SK16"],
-  // BACD5 is the Hoya trade name for N-SK16-equivalent glass (same nd/vd).
-  ["BACD5", "N-SK16"],
-  // L-TIM28 is the Ohara large-format designation for S-TIM28 (identical Sellmeier).
-  ["L-TIM28", "S-TIM28"],
-  // Low-softening / PGM Ohara variants that share the same optical constants in
-  // the public catalog data used by the lens files.
-  ["L-BAL42", "S-BAL42"],
-  ["L-BSL7", "S-BSL7"],
-  ["L-BAL35", "S-BAL35"],
-  ["L-LAH53", "S-LAH53"],
-  ["L-LAM60", "S-LAM60"],
-  ["L-LAL13", "S-LAL13"],
-  ["L-LAL14", "S-LAL14"],
-  ["L-PHM52", "S-PHM52"],
-  // SK10 and SK14 are legacy Schott names superseded by N-SK10 and N-SK14.
-  ["SK5", "N-SK5"],
-  ["SK10", "N-SK10"],
-  ["SK14", "N-SK14"],
-  // Legacy Schott names with modern N- prefixed equivalents.
-  ["N-F2", "F2"],
-  ["N-SF1", "SF1"],
-  ["N-SF6", "SF6"],
-  ["SF8", "N-SF8"],
-  ["FK5", "N-FK5"],
-  ["LAF21", "N-LAF21"],
-  ["LA-F21", "N-LAF21"],
-  ["SK2", "N-SK2"],
-  // HOYA's TAFD35L PGM/pressed-glass row shares the TAFD35 911/353 optical constants.
-  ["TAFD35L", "TAFD35"],
-  // Historical Hoya BSC3 annotations in the lens data refer to the 518/590
-  // E-C3 crown row; 583/594 barium-crown rows must not use this alias.
-  ["BSC3", "E-C3"],
-]);
+export const ALIASES: ReadonlyMap<string, string> = new Map(
+  ALIAS_RECORDS.map(({ alias, target }) => [alias.toUpperCase(), target]),
+);
