+++ src/optics/glassCatalogData.ts
@@ -1,3664 +1,399 @@
 /**
  * Vendor-published optical glass dispersion entries used by glassCatalog.ts.
+ *
+ * The catalog is sharded by vendor under glassCatalogEntries/ for maintainability,
+ * then reconstructed in the original source order so public iteration and code
+ * lookup behavior stay stable.
  */
 
-/** A single entry in the glass catalog. */
-export interface GlassEntry {
-  /** Vendor-canonical glass name, e.g. "N-BK7", "S-FPL51". */
-  readonly name: string;
-  /** Vendor (Schott, Ohara, Hoya, Hikari, Sumita, CDGM, NHG, special). */
-  readonly vendor: "Schott" | "Ohara" | "Hoya" | "Hikari" | "Sumita" | "CDGM" | "NHG" | "Special";
-  /** Sellmeier B1, B2, B3 coefficients. */
-  readonly B?: readonly [number, number, number];
-  /** Sellmeier C1, C2, C3 coefficients (μm²). */
-  readonly C?: readonly [number, number, number];
-  /**
-   * Zemax formula 3 polynomial coefficients:
-   * n² = a0 + a1·λ² + a2·λ⁻² + a3·λ⁻⁴ + a4·λ⁻⁶ + a5·λ⁻⁸.
-   */
-  readonly polynomial?: readonly [number, number, number, number, number, number];
-  /**
-   * Explicit power-series terms for formula-3 entries whose published exponents
-   * do not fit the fixed six-slot `polynomial` shorthand.
-   */
-  readonly powerSeries?: readonly (readonly [coefficient: number, exponent: number])[];
-  /** Reference d-line index (587.5618 nm). For sanity-check / display. */
-  readonly nd: number;
-  /** Reference Abbe number νd. For sanity-check / display. */
-  readonly vd: number;
-  /** Reference partial dispersion P_g,F (optional, when vendor publishes it). */
-  readonly PgF?: number;
-  /** Optional six-digit Schott-style code "nnn vvv" without space, e.g. "517642" for N-BK7. */
-  readonly code6?: string;
-  /** Source citation for the Sellmeier coefficients. */
-  readonly source: string;
-}
+import type { GlassEntry } from "./glassCatalogTypes.js";
+import { SCHOTT_GLASS_ENTRIES } from "./glassCatalogEntries/schott.js";
+import { OHARA_GLASS_ENTRIES } from "./glassCatalogEntries/ohara.js";
+import { HOYA_GLASS_ENTRIES } from "./glassCatalogEntries/hoya.js";
+import { HIKARI_GLASS_ENTRIES } from "./glassCatalogEntries/hikari.js";
+import { SUMITA_GLASS_ENTRIES } from "./glassCatalogEntries/sumita.js";
+import { CDGM_GLASS_ENTRIES } from "./glassCatalogEntries/cdgm.js";
+import { NHG_GLASS_ENTRIES } from "./glassCatalogEntries/nhg.js";
+import { SPECIAL_GLASS_ENTRIES } from "./glassCatalogEntries/special.js";
 
-/* ──────────────────────────────────────────────────────────────────────────
- * GLASS CATALOG — 302 vendor-verified entries (current source count, June 2026)
+export type { GlassEntry } from "./glassCatalogTypes.js";
+
+const GLASS_CATALOG_SOURCE_ORDER = [
+  "N-BK7",
+  "S-BSL7",
+  "CaF2",
+  "S-FPL51",
+  "S-FPL53",
+  "S-FSL5",
+  "S-LAL18",
+  "S-LAH58",
+  "S-LAH55V",
+  "S-LAH65V",
+  "S-LAH79",
+  "S-NBH5",
+  "S-NPH2",
+  "S-TIH3",
+  "S-TIH6",
+  "S-TIH53",
+  "S-PHM52",
+  "S-FPM2",
+  "S-LAL14",
+  "S-LAH66",
+  "S-BAL42",
+  "S-BSM22",
+  "S-LAH63",
+  "S-LAM66",
+  "S-LAH55",
+  "S-LAH95",
+  "S-LAH59",
+  "S-LAH53",
+  "S-LAH92",
+  "S-LAH88",
+  "S-PHM53",
+  "S-TIH14",
+  "S-TIM35",
+  "S-FPL52",
+  "S-TIM25",
+  "S-FPM3",
+  "S-BAL35",
+  "S-TIM22",
+  "S-NPH1",
+  "S-TIM2",
+  "S-TIM3",
+  "S-LAH65",
+  "S-LAH52",
+  "S-NSL3",
+  "S-LAM54",
+  "S-NPH53",
+  "S-LAH51",
+  "S-TIM28",
+  "S-BSM14",
+  "SF6",
+  "SF4",
+  "SF1",
+  "N-LAK8",
+  "N-KZFS5",
+  "SF2",
+  "N-SK16",
+  "N-KZFS4",
+  "N-FK51A",
+  "N-SK11",
+  "F2",
+  "N-LAK22",
+  "FCD1",
+  "FCD505",
+  "TAFD30",
+  "TAFD37A",
+  "TAFD37",
+  "TAFD33",
+  "K-GFK68",
+  "K-SK7",
+  "K-SK18",
+  "S-TIH18",
+  "S-TIH4",
+  "S-TIM5",
+  "S-TIM27",
+  "S-NBH8",
+  "S-NBH51",
+  "S-NBH52",
+  "S-NBH55",
+  "S-NBH56",
+  "S-LAH55VS",
+  "S-LAH60",
+  "S-LAH60V",
+  "S-LAH64",
+  "S-LAH89",
+  "S-LAH97",
+  "S-LAM2",
+  "S-LAM51",
+  "S-LAL8",
+  "S-LAL9",
+  "S-BAM4",
+  "S-BAL14",
+  "S-BSM18",
+  "S-BSM81",
+  "S-NSL5",
+  "S-NPH4",
+  "N-SK10",
+  "N-SK14",
+  "S-FPL55",
+  "S-LAH98",
+  "S-LAH99",
+  "L-LAH85V",
+  "S-LAM60",
+  "S-NPH3",
+  "S-NPH5",
+  "S-LAH63Q",
+  "S-LAH65VS",
+  "S-NBM51",
+  "S-FTM16",
+  "S-NSL36",
+  "N-SF66",
+  "N-LAF21",
+  "SF11",
+  "N-SK2",
+  "E-FDS1",
+  "FCD100",
+  "FCD515",
+  "FCD705",
+  "TAFD45",
+  "TAFD55",
+  "E-CF6",
+  "MC-TAF101-100",
+  "NBFD15",
+  "NBFD25",
+  "NBFD29",
+  "FDS24",
+  "BACD15",
+  "TAFD40",
+  "S-LAH93",
+  "S-TIH11",
+  "S-TIL25",
+  "S-BAH27",
+  "S-NBH52V",
+  "S-TIH53W",
+  "S-BAH28",
+  "S-LAL59",
+  "S-TIL27",
+  "S-TIL6",
+  "N-FK5",
+  "FC5",
+  "E-FD15",
+  "TAFD5F",
+  "S-BSM16",
+  "S-TIH13",
+  "S-NBH53V",
+  "S-TIH23",
+  "SF56A",
+  "N-BALF4",
+  "N-SK5",
+  "N-BAF10",
+  "N-LAF34",
+  "N-LAK14",
+  "E-FD2",
+  "E-FD8",
+  "FF8",
+  "TAFD35",
+  "H-ZLaF68L",
+  "S-LAH96",
+  "H-ZF88",
+  "H-ZLaF76",
+  "LAC13",
+  "J-PSKH1",
+  "J-PSKH4",
+  "J-KZFH9",
+  "BAF3",
+  "J-LAK10",
+  "J-LASFH15",
+  "J-LASFH9",
+  "FDS18",
+  "NBFD3",
+  "NBFD11",
+  "TAFD25",
+  "TAF5",
+  "E-FD4",
+  "E-FD5",
+  "E-FL5",
+  "S-BAL2",
+  "S-LAM55",
+  "S-TIH10",
+  "S-BSM28",
+  "S-BSM71",
+  "S-LAL13",
+  "S-LAM3",
+  "S-LAM52",
+  "J-PKH1",
+  "Q-SK52S",
+  "J-LASKH2",
+  "J-LASFH22",
+  "J-LASF014",
+  "TAFD65",
+  "F5",
+  "E-FDS2",
+  "M-FCD500",
+  "PBL25",
+  "E-F1",
+  "LAC12",
+  "S-LAM59",
+  "N-KZFS8",
+  "S-LAL10",
+  "SF10",
+  "K-VC89",
+  "H-FK61",
+  "S-TIL2",
+  "S-TIL26",
+  "N-LAK33B",
+  "TAFD32",
+  "H-QK3L",
+  "H-ZF6",
+  "H-ZLAF50D",
+  "H-ZF4A",
+  "H-FK71",
+  "H-ZF13",
+  "H-ZF52",
+  "H-ZF72A",
+  "H-ZF7LA",
+  "H-ZLAF52A",
+  "H-ZLAF68C",
+  "H-LAK6A",
+  "H-LAF3B",
+  "S-PHM52Q",
+  "S-FPM4",
+  "S-NBH58",
+  "S-BAH11",
+  "S-BAL41",
+  "S-LAH52Q",
+  "S-LAH71",
+  "S-LAL7",
+  "S-LAM61",
+  "S-TIH1",
+  "S-TIL1",
+  "S-TIM8",
+  "E-F2",
+  "NBFD13",
+  "E-F3",
+  "N-BALF5",
+  "FK3",
+  "BPH50",
+  "H-LAK12",
+  "S-LAM58",
+  "D-ZLAF81-25",
+  "K-SKLD200",
+  "J-FK5",
+  "J-FK01A",
+  "J-FKH1",
+  "J-BASF2",
+  "J-BASF6",
+  "J-SFH1",
+  "J-SFH5",
+  "N-SSK5",
+  "J-KZFH6",
+  "J-LAF01",
+  "Q-LASFPH2S",
+  "J-LAF016",
+  "J-LASF017",
+  "Q-LASFPH3S",
+  "J-LASF021",
+  "J-LASFH23",
+  "PCD40",
+  "M-LAC130",
+  "M-BACD5N",
+  "M-TAFD305",
+  "BAFD7",
+  "SF9",
+  "S-LAL58",
+  "MC-NBFD135",
+  "J-BK7",
+  "BK7G18",
+  "P-SK57Q1",
+  "M-PCD51",
+  "SK9",
+  "E-BAF8",
+  "M-TAC80",
+  "M-LAF81",
+  "TAC2",
+  "LAK011",
+  "E-LAF11",
+  "E-LASF016",
+  "M-TAF401",
+  "M-TAF31",
+  "E-LAFH2",
+  "M-NBFD10",
+  "FDS16-W",
+  "S-LAL12",
+  "S-BSM10",
+  "S-LAM7",
+  "L-LAM69",
+  "N-SF8",
+  "H-LAF4",
+  "E-FD13",
+  "E-FD10",
+  "BACD4",
+  "E-C3",
+  "S-BSM15",
+  "SF5",
+  "N-SF5",
+  "N-LASF44",
+  "N-LAK9",
+  "N-PSK53A",
+  "N-LAF2",
+  "N-BAK4",
+  "N-LAK7",
+  "N-BAF4",
+  "N-SSK2",
+] as const;
+
+const GLASS_CATALOG_ENTRIES_BY_NAME: ReadonlyMap<string, GlassEntry> = new Map(
+  [
+    ...SCHOTT_GLASS_ENTRIES,
+    ...OHARA_GLASS_ENTRIES,
+    ...HOYA_GLASS_ENTRIES,
+    ...HIKARI_GLASS_ENTRIES,
+    ...SUMITA_GLASS_ENTRIES,
+    ...CDGM_GLASS_ENTRIES,
+    ...NHG_GLASS_ENTRIES,
+    ...SPECIAL_GLASS_ENTRIES,
+  ].map((entry) => [entry.name, entry] as const),
+);
+
+/**
+ * Duplicate six-digit glass-code winners from the original monolithic catalog order.
  *
- * Coefficients are taken from authoritative public vendor catalogs. Each
- * entry's `source` field cites the document or database used. To verify a
- * new entry, compute n at LINE_NM.d and compare to the listed nd; agreement
- * to ~1e-5 is the expected accuracy of a transcribed Sellmeier fit.
- *
- * For the prioritized addition list and sourcing playbook, see
- * agent_docs/glass-catalog-buildout.md.
- * ────────────────────────────────────────────────────────────────────────── */
+ * Several equivalent vendor entries intentionally share a Schott-style code. The
+ * resolver uses these explicit winners before adding non-duplicate code lookups.
+ */
+export const DUPLICATE_CODE6_PRECEDENCE: ReadonlyMap<string, string> = new Map([
+  ["497816", "S-FPL51"],
+  ["883408", "S-LAH58"],
+  ["654397", "S-NBH5"],
+  ["923189", "S-NPH2"],
+  ["805254", "S-TIH6"],
+  ["847238", "S-TIH53"],
+  ["618634", "S-PHM52"],
+  ["773496", "S-LAH66"],
+  ["804396", "S-LAH63"],
+  ["835427", "S-LAH55"],
+  ["904313", "S-LAH95"],
+  ["816466", "S-LAH59"],
+  ["699301", "S-TIM35"],
+  ["648338", "S-TIM22"],
+  ["620363", "S-TIM2"],
+  ["613370", "S-TIM3"],
+  ["804466", "S-LAH65"],
+  ["800422", "S-LAH52"],
+  ["518590", "S-NSL3"],
+  ["717295", "SF1"],
+  ["620603", "N-SK16"],
+  ["593686", "FCD505"],
+  ["954323", "S-LAH98"],
+  ["923209", "N-SF66"],
+  ["785257", "S-TIH11"],
+  ["702412", "S-BAH27"],
+  ["487704", "N-FK5"],
+  ["741278", "S-TIH13"],
+  ["589613", "N-SK5"],
+  ["755275", "E-FD4"],
+  ["673322", "E-FD5"],
+  ["694532", "S-LAL13"],
+  ["755523", "J-LASKH2"],
+  ["728283", "H-ZF4A"],
+  ["744449", "H-LAF3B"],
+  ["652585", "S-LAL7"],
+  ["795453", "J-LASF017"],
+  ["731405", "M-LAF81"],
+]);
+
+function entryByName(name: string): GlassEntry {
+  const entry = GLASS_CATALOG_ENTRIES_BY_NAME.get(name);
+  if (!entry) throw new Error(`Glass catalog source-order entry ${name} has no vendor shard entry.`);
+  return entry;
+}
 
 /**
  * Raw vendor glass catalog entries consumed by glassCatalog.ts.
  *
  * Entries are immutable data: Sellmeier wavelengths use micrometers internally,
- * reference `nd` is the d line, and `vd` is the vendor Abbe number used for
- * sanity checks and fallback dispersion.
+ * reference nd is the d line, and vd is the vendor Abbe number used for sanity
+ * checks and fallback dispersion.
  */
-export const RAW_CATALOG: readonly GlassEntry[] = [
-  /* ────── Universal crowns (BK7-class) ──────
-   * N-BK7 is the universal reference crown; S-BSL7 is Ohara's near-equivalent.
-   * Both are present in nearly every lens design somewhere in the rear group.
-   */
-  {
-    name: "N-BK7",
-    vendor: "Schott",
-    B: [1.03961212, 0.231792344, 1.01046945],
-    C: [0.00600069867, 0.0200179144, 103.560653],
-    nd: 1.5168,
-    vd: 64.17,
-    PgF: 0.5349,
-    code6: "517642",
-    source: "Schott TIE-29 / N-BK7 datasheet (Schott AG, public).",
-  },
-  {
-    name: "S-BSL7",
-    vendor: "Ohara",
-    B: [1.03773418, 0.231473863, 1.01072958],
-    C: [0.00611873959, 0.0202423191, 103.677302],
-    nd: 1.51633,
-    vd: 64.14,
-    PgF: 0.5349,
-    code6: "517641",
-    source: "Ohara optical glass catalog (S-BSL7 datasheet).",
-  },
-
-  /* ────── Special: fluorite (CaF2) ──────
-   * The fundamental ultra-low-dispersion crystal. Used in Canon "Fluorite",
-   * many Nikkor and Fujinon long teles, and as the reference end-point on
-   * the Abbe diagram (νd ≈ 95).
-   */
-  {
-    name: "CaF2",
-    vendor: "Special",
-    B: [0.5675888, 0.4710914, 3.8484723],
-    C: [0.0025264, 0.0100783, 1200.5557],
-    nd: 1.43385,
-    vd: 95.1,
-    PgF: 0.5348,
-    source: "Daimon & Masumura, Appl. Opt. 41 (2002), CaF2 room-temperature fit.",
-  },
-
-  /* ────── Ohara — primary catalog vendor for this codebase ──────
-   * S-FPL51 alone accounts for 74 declarations across the lens library; the
-   * remaining Ohara entries together cover several hundred more. Coefficients
-   * transcribed verbatim from the vendor-published Sellmeier-1 form (AGF
-   * dispersion form 2). Round-trip at d-line is well inside 1e-4.
-   */
-  {
-    name: "S-FPL51",
-    vendor: "Ohara",
-    B: [1.17010505, 0.0475710783, 0.763832445],
-    C: [0.00616203924, 0.0263372876, 141.882642],
-    nd: 1.496999,
-    vd: 81.5459,
-    PgF: 0.5346,
-    code6: "497816",
-    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
-  },
-  {
-    name: "S-FPL53",
-    vendor: "Ohara",
-    B: [0.983532327, 0.069568814, 1.11409238],
-    C: [0.00492234955, 0.0193581091, 264.275294],
-    nd: 1.43875,
-    vd: 94.946,
-    PgF: 0.5302,
-    code6: "439950",
-    source:
-      "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data. Super-ED defining glass for many APO triplets.",
-  },
-  {
-    name: "S-FSL5",
-    vendor: "Ohara",
-    B: [1.17447043, 0.0140056154, 1.19272435],
-    C: [0.00841855181, -0.0581790767, 129.599726],
-    nd: 1.48749,
-    vd: 70.2363,
-    PgF: 0.5279,
-    code6: "487702",
-    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
-  },
-  {
-    name: "S-LAL18",
-    vendor: "Ohara",
-    B: [1.50276318, 0.430224497, 1.3472606],
-    C: [0.0145462356, -0.00332784153, 93.3508342],
-    nd: 1.729157,
-    vd: 54.68,
-    PgF: 0.5432,
-    code6: "729547",
-    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
-  },
-  {
-    name: "S-LAH58",
-    vendor: "Ohara",
-    B: [1.78764964, 0.6526356, 1.79914564],
-    C: [0.00847378536, 0.0313126408, 132.788001],
-    nd: 1.882997,
-    vd: 40.7651,
-    PgF: 0.5664,
-    code6: "883408",
-    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
-  },
-  {
-    name: "S-LAH55V",
-    vendor: "Ohara",
-    B: [1.97025325, 0.30489414, 1.39214665],
-    C: [0.00991088134, 0.0383202295, 97.7785249],
-    nd: 1.834807,
-    vd: 42.7253,
-    PgF: 0.5644,
-    source:
-      "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data. Vacuum-melt variant; distinct from non-V.",
-  },
-  {
-    name: "S-LAH65V",
-    vendor: "Ohara",
-    B: [1.81419034, 0.361376301, 1.32729484],
-    C: [0.00874935029, 0.0318352836, 91.3406898],
-    nd: 1.803999,
-    vd: 46.5834,
-    PgF: 0.5566,
-    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
-  },
-  {
-    name: "S-LAH79",
-    vendor: "Ohara",
-    B: [2.32557148, 0.507967133, 2.43087198],
-    C: [0.0132895208, 0.0528335449, 161.122408],
-    nd: 2.0033,
-    vd: 28.2733,
-    PgF: 0.5985,
-    code6: "003283",
-    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
-  },
-  {
-    name: "S-NBH5",
-    vendor: "Ohara",
-    B: [1.47544521, 0.193060095, 1.5093901],
-    C: [0.0095583674, 0.0460430483, 126.422746],
-    nd: 1.654115,
-    vd: 39.6828,
-    PgF: 0.5735,
-    code6: "654397",
-    source:
-      "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data. KZFS-class with negative ΔPgF; APO-relevant.",
-  },
-  {
-    name: "S-NPH2",
-    vendor: "Ohara",
-    B: [2.0386951, 0.437269641, 2.96711461],
-    C: [0.0170796224, 0.0749254813, 174.155354],
-    nd: 1.92286,
-    vd: 18.8969,
-    PgF: 0.6506,
-    code6: "923189",
-    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
-  },
-  {
-    name: "S-TIH3",
-    vendor: "Ohara",
-    B: [1.64797648, 0.267261917, 2.19772845],
-    C: [0.0121917693, 0.0597893039, 192.15834],
-    nd: 1.739998,
-    vd: 28.296005,
-    code6: "740283",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIH3 page.",
-  },
-  {
-    name: "S-TIH6",
-    vendor: "Ohara",
-    B: [1.77227611, 0.34569125, 2.40788501],
-    C: [0.0131182633, 0.0614479619, 200.753254],
-    nd: 1.805181,
-    vd: 25.4254,
-    PgF: 0.6168,
-    code6: "805254",
-    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
-  },
-  {
-    name: "S-TIH53",
-    vendor: "Ohara",
-    B: [1.87904886, 0.369719775, 2.33730863],
-    C: [0.014412177, 0.063881799, 182.66818],
-    nd: 1.84666,
-    vd: 23.7779,
-    PgF: 0.6213,
-    code6: "847238",
-    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
-  },
-
-  /* ────── Ohara expansion (Phase 1 buildout, Apr 2026) ──────
-   * Eight high-frequency Ohara glasses each with ≥10 lens-element occurrences
-   * across the library. Sourced from the same OHARA Zemax catalog
-   * (ohara_2017-11-30.agf) as the entries above; coefficients pulled via
-   * refractiveindex.info's mirror of that AGF (form 2 = Sellmeier-1).
-   * PgF computed from Schott's normal-line baseline plus ΔPgF.
-   */
-  {
-    name: "S-PHM52",
-    vendor: "Ohara",
-    B: [1.0996655, 0.478125422, 1.13214074],
-    C: [0.0132718559, -0.000601649685, 130.595472],
-    nd: 1.618,
-    vd: 63.3335,
-    PgF: 0.5424,
-    code6: "618634",
-    source:
-      "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Phosphate crown with notable +ΔPgF; APO-relevant.",
-  },
-  {
-    name: "S-FPM2",
-    vendor: "Ohara",
-    B: [0.761242785, 0.747033375, 0.938928947],
-    C: [0.00321174095, 0.0140234423, 139.52353],
-    nd: 1.59522,
-    vd: 67.7357,
-    PgF: 0.5422,
-    code6: "595677",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Fluorophosphate ED with significant +ΔPgF.",
-  },
-  {
-    name: "S-LAL14",
-    vendor: "Ohara",
-    B: [1.2372097, 0.589722623, 1.3192188],
-    C: [0.015355132, -0.00030789625, 93.7202947],
-    nd: 1.696797,
-    vd: 55.5322,
-    PgF: 0.5422,
-    code6: "697555",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Lanthanum crown.",
-  },
-  {
-    name: "S-LAH66",
-    vendor: "Ohara",
-    B: [1.39280586, 0.679577094, 1.38702069],
-    C: [0.00608475118, 0.0233925351, 95.8354094],
-    nd: 1.772499,
-    vd: 49.5984,
-    PgF: 0.5512,
-    code6: "773496",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum.",
-  },
-  {
-    name: "S-BAL42",
-    vendor: "Ohara",
-    B: [1.39570615, 0.071850507, 1.27129267],
-    C: [0.0112218843, -0.0252117422, 134.49786],
-    nd: 1.583126,
-    vd: 59.3747,
-    PgF: 0.5419,
-    code6: "583594",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Barium crown.",
-  },
-  {
-    name: "S-BSM22",
-    vendor: "Ohara",
-    B: [1.44305741, 0.140786358, 1.26093951],
-    C: [0.0081920891, 0.0356911455, 131.959337],
-    nd: 1.622296,
-    vd: 53.169067,
-    code6: "622532",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BSM22 page.",
-  },
-  {
-    name: "S-LAH63",
-    vendor: "Ohara",
-    B: [1.89458276, 0.268702978, 1.45705526],
-    C: [0.0102277048, 0.0442801243, 104.874927],
-    nd: 1.804398,
-    vd: 39.5862,
-    PgF: 0.5727,
-    code6: "804396",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum.",
-  },
-  {
-    name: "S-LAM66",
-    vendor: "Ohara",
-    B: [1.92094221, 0.219901208, 1.72705231],
-    C: [0.0115075241, 0.0547993543, 120.133674],
-    nd: 1.800999,
-    vd: 34.9674,
-    PgF: 0.5865,
-    code6: "801350",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Lanthanum medium-dispersion.",
-  },
-  {
-    name: "S-LAH55",
-    vendor: "Ohara",
-    B: [1.95615766, 0.319216215, 1.39173189],
-    C: [0.00979338965, 0.0376836296, 94.8775271],
-    nd: 1.834807,
-    vd: 42.7137,
-    PgF: 0.5638,
-    code6: "835427",
-    source:
-      "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Standard melt; distinct from S-LAH55V (vacuum-melt variant) — different Sellmeier coefficients despite shared nominal nd.",
-  },
-
-  /* ────── Ohara expansion (Phase 2 buildout, Apr 2026) ──────
-   * Six more OHARA glasses targeted at the largest "no candidate" clusters
-   * surfaced by the verified mismatch sweep (see agent_docs/glass-relabel-followup.md).
-   * Each fills a real gap: S-LAH95 covers the ~10 surfaces of "S-LAH79 type (903/313)"
-   * mislabels; S-LAH59 covers the (816/466) cluster; S-PHM53 / S-LAH53 / S-LAH92 /
-   * S-LAH88 fill smaller but distinct clusters. Same AGF source as the entries above.
-   */
-  {
-    name: "S-LAH95",
-    vendor: "Ohara",
-    B: [2.15636617, 0.329558178, 1.72178935],
-    C: [0.012288051, 0.0555507835, 124.43934],
-    nd: 1.903659,
-    vd: 31.3431,
-    PgF: 0.5966,
-    code6: "904313",
-    source:
-      "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum (904/313); resolves the largest cluster of mislabeled S-LAH79 surfaces in the lens library.",
-  },
-  {
-    name: "S-LAH59",
-    vendor: "Ohara",
-    B: [1.51372967, 0.702462343, 1.33600982],
-    C: [0.00705246901, 0.0249488689, 100.085908],
-    nd: 1.816,
-    vd: 46.6206,
-    PgF: 0.5654,
-    code6: "816466",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Lanthanum dense crown.",
-  },
-  {
-    name: "S-LAH53",
-    vendor: "Ohara",
-    B: [1.91811619, 0.253724399, 1.39473885],
-    C: [0.0102147684, 0.0433176011, 101.938021],
-    nd: 1.806098,
-    vd: 40.926,
-    PgF: 0.575,
-    code6: "806409",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum.",
-  },
-  {
-    name: "S-LAH92",
-    vendor: "Ohara",
-    B: [2.10440311, 0.358346161, 1.63010064],
-    C: [0.0108531811, 0.044340592, 123.2498],
-    nd: 1.891899,
-    vd: 37.134,
-    PgF: 0.578,
-    code6: "892371",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum (892/371).",
-  },
-  {
-    name: "S-LAH88",
-    vendor: "Ohara",
-    B: [2.1284434, 0.405082139, 1.67918461],
-    C: [0.0117309815, 0.0508706599, 107.091456],
-    nd: 1.916499,
-    vd: 31.6041,
-    PgF: 0.5915,
-    code6: "917316",
-    source:
-      "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info; Ohara GmbH S-LAH88 datasheet gives d-code 917316.",
-  },
-  {
-    name: "S-PHM53",
-    vendor: "Ohara",
-    B: [1.09775423, 0.434816432, 1.13894976],
-    C: [0.01233694, -0.000372522903, 124.276984],
-    nd: 1.603001,
-    vd: 65.4436,
-    PgF: 0.5337,
-    code6: "603655",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Phosphate crown (603/655).",
-  },
-
-  /* ────── Ohara expansion (Phase 3 buildout, Apr 2026) ──────
-   * Fifteen additional Ohara glasses covering the next tier of frequency:
-   * dense TIM/TIH flints (S-TIH14, S-TIM35, S-TIM25, S-TIM22, S-TIM2),
-   * ED and fluorophosphate crowns (S-FPL52, S-FPM3),
-   * barium and borosilicate crowns (S-BAL35, S-NSL3),
-   * high-index lanthanum variants (S-LAH51, S-LAH52, S-LAH65, S-LAM54),
-   * and ultra-high-index flints (S-NPH1, S-NPH53).
-   * All sourced from ohara_2017-11-30.agf via refractiveindex.info (form 2).
-   */
-  {
-    name: "S-TIH14",
-    vendor: "Ohara",
-    B: [1.68915108, 0.290462024, 2.37971516],
-    C: [0.0128202514, 0.0618090841, 201.094352],
-    nd: 1.761821,
-    vd: 26.517885,
-    PgF: 0.6142,
-    code6: "762265",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Dense flint (762/265).",
-  },
-  {
-    name: "S-TIM35",
-    vendor: "Ohara",
-    B: [1.55849775, 0.230767007, 1.84436099],
-    C: [0.0115367235, 0.0586095947, 162.981888],
-    nd: 1.698947,
-    vd: 30.127865,
-    PgF: 0.6034,
-    code6: "699301",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Dense flint (699/301).",
-  },
-  {
-    name: "S-FPL52",
-    vendor: "Ohara",
-    B: [1.06785857, 0.0335857718, 1.10219763],
-    C: [0.00699227302, -0.0207608925, 226.496541],
-    nd: 1.455999,
-    vd: 90.288026,
-    PgF: 0.5305,
-    code6: "456903",
-    source:
-      "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. ED fluorophosphate crown (456/903); sister to S-FPL51. C2 is negative — this is valid for fluorophosphate glasses with anomalous UV resonances.",
-  },
-  {
-    name: "S-TIM25",
-    vendor: "Ohara",
-    B: [1.50659233, 0.204786135, 1.92036668],
-    C: [0.0109501562, 0.0574980285, 178.128535],
-    nd: 1.6727,
-    vd: 32.099206,
-    PgF: 0.5991,
-    code6: "673321",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Dense flint (673/321).",
-  },
-  {
-    name: "S-FPM3",
-    vendor: "Ohara",
-    B: [0.809407286, 0.527007033, 0.909127704],
-    C: [0.00376072389, 0.0135654895, 142.503612],
-    nd: 1.53775,
-    vd: 74.703119,
-    PgF: 0.5368,
-    code6: "538747",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Fluorophosphate crown with +ΔPgF.",
-  },
-  {
-    name: "S-BAL35",
-    vendor: "Ohara",
-    B: [0.941357273, 0.546174895, 1.16168917],
-    C: [0.0140333996, 0.000906635683, 114.163758],
-    nd: 1.58913,
-    vd: 61.135024,
-    PgF: 0.5392,
-    code6: "589612",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Barium crown (589/612).",
-  },
-  {
-    name: "S-TIM22",
-    vendor: "Ohara",
-    B: [1.44222294, 0.194432265, 1.74092482],
-    C: [0.0104249404, 0.0550235257, 169.710769],
-    nd: 1.647689,
-    vd: 33.792803,
-    PgF: 0.594,
-    code6: "648338",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Dense flint (648/338).",
-  },
-  {
-    name: "S-NPH1",
-    vendor: "Ohara",
-    B: [1.75156623, 0.364006304, 2.47874141],
-    C: [0.0135004681, 0.0668245147, 170.756006],
-    nd: 1.808095,
-    vd: 22.760817,
-    PgF: 0.6316,
-    code6: "808228",
-    source:
-      "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Ultra-high-index niobophosphate flint (808/228).",
-  },
-  {
-    name: "S-TIM2",
-    vendor: "Ohara",
-    B: [1.42193846, 0.133827968, 1.45060574],
-    C: [0.0107291511, 0.0572587546, 145.381805],
-    nd: 1.620041,
-    vd: 36.263378,
-    PgF: 0.5879,
-    code6: "620363",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Dense flint (620/363).",
-  },
-  {
-    name: "S-TIM3",
-    vendor: "Ohara",
-    B: [1.40691144, 0.128369745, 1.51826191],
-    C: [0.0105633641, 0.0568483105, 152.107924],
-    nd: 1.612929,
-    vd: 37.005342,
-    PgF: 0.586,
-    code6: "613370",
-    source: "Ohara S-TIM/S-FTM table and Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIM3 page.",
-  },
-  {
-    name: "S-LAH65",
-    vendor: "Ohara",
-    B: [1.68191258, 0.493779818, 1.45682822],
-    C: [0.0077668425, 0.0288916181, 99.2574356],
-    nd: 1.804,
-    vd: 46.570373,
-    PgF: 0.5565,
-    code6: "804466",
-    source:
-      "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum (804/466); standard melt — distinct from S-LAH65V (vacuum-melt variant).",
-  },
-  {
-    name: "S-LAH52",
-    vendor: "Ohara",
-    B: [1.85390925, 0.297925555, 1.39382086],
-    C: [0.00955320687, 0.039381685, 102.706848],
-    nd: 1.799516,
-    vd: 42.225007,
-    PgF: 0.5668,
-    code6: "800422",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum (800/422).",
-  },
-  {
-    name: "S-NSL3",
-    vendor: "Ohara",
-    B: [0.882514764, 0.389271907, 1.10693448],
-    C: [0.00464504582, 0.0200551397, 136.234339],
-    nd: 1.518229,
-    vd: 58.902057,
-    PgF: 0.5442,
-    code6: "518590",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Borosilicate crown (518/590).",
-  },
-  {
-    name: "S-LAM54",
-    vendor: "Ohara",
-    B: [1.84213306, 0.175468631, 1.25750878],
-    C: [0.0094399322, 0.0395281122, 86.5463013],
-    nd: 1.756998,
-    vd: 47.82317,
-    PgF: 0.5558,
-    code6: "757478",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Lanthanum medium-dispersion (757/478).",
-  },
-  {
-    name: "S-NPH53",
-    vendor: "Ohara",
-    B: [1.85484904, 0.396194484, 2.43512461],
-    C: [0.0134621486, 0.0631945361, 170.864886],
-    nd: 1.84666,
-    vd: 23.88395,
-    PgF: 0.6226,
-    code6: "847239",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Niobophosphate flint (847/239).",
-  },
-  {
-    name: "S-LAH51",
-    vendor: "Ohara",
-    B: [1.82586991, 0.283023349, 1.35964319],
-    C: [0.00935297152, 0.0373803057, 100.655798],
-    nd: 1.785896,
-    vd: 44.202637,
-    PgF: 0.5626,
-    code6: "786442",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum (786/442).",
-  },
-
-  /* ────── Ohara expansion (Phase 4 buildout, Apr 2026) ──────
-   * S-TIM28: dense TIM flint used in Canon RF, Nikon L35AF/PCE19, Fuji XF zooms.
-   * S-BSM14: borosilicate medium crown used in Canon EF/RF and other normal lenses.
-   * Coefficients from OHARA Zemax catalog (ohara_2017-11-30.agf) via refractiveindex.info.
-   */
-  {
-    name: "S-TIM28",
-    vendor: "Ohara",
-    B: [1.5427081, 0.217113891, 1.81904459],
-    C: [0.0113925005, 0.0579224572, 167.697189],
-    nd: 1.688931,
-    vd: 31.07545,
-    PgF: 0.6008,
-    code6: "689311",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Dense flint (689/311).",
-  },
-  {
-    name: "S-BSM14",
-    vendor: "Ohara",
-    B: [1.2828627, 0.247647429, 1.10383999],
-    C: [0.0122902399, -0.00613142361, 106.883378],
-    nd: 1.603112,
-    vd: 60.64108,
-    PgF: 0.5399,
-    code6: "603607",
-    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Borosilicate medium crown (603/607).",
-  },
-
-  /* ────── Schott — high-dispersion flints and lanthanum crowns ──────
-   * SF6/SF4/SF1 are the legacy lead flints used widely in vintage designs.
-   * N-LAK8 is a high-frequency lanthanum crown. N-KZFS5 carries notable
-   * negative ΔPgF and is APO-relevant. Coefficients transcribed verbatim
-   * from the vendor-published Sellmeier-1 form (AGF dispersion form 2).
-   */
-  {
-    name: "SF6",
-    vendor: "Schott",
-    B: [1.72448482, 0.390104889, 1.04572858],
-    C: [0.0134871947, 0.0569318095, 118.557185],
-    nd: 1.80518,
-    vd: 25.43,
-    PgF: 0.6102,
-    code6: "805254",
-    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data.",
-  },
-  {
-    name: "SF4",
-    vendor: "Schott",
-    B: [1.61957826, 0.339493189, 1.02566931],
-    C: [0.0125502104, 0.0544559822, 117.652222],
-    nd: 1.7552,
-    vd: 27.58,
-    PgF: 0.6036,
-    code6: "755276",
-    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data.",
-  },
-  {
-    name: "SF1",
-    vendor: "Schott",
-    B: [1.55912923, 0.284246288, 0.968842926],
-    C: [0.0121481001, 0.0534549042, 112.174809],
-    nd: 1.71736,
-    vd: 29.51,
-    PgF: 0.5984,
-    code6: "717295",
-    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data.",
-  },
-  {
-    name: "N-LAK8",
-    vendor: "Schott",
-    B: [1.33183167, 0.546623206, 1.19084015],
-    C: [0.00620023871, 0.0216465439, 82.5827736],
-    nd: 1.713,
-    vd: 53.83,
-    PgF: 0.545,
-    code6: "713538",
-    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data.",
-  },
-  {
-    name: "N-KZFS5",
-    vendor: "Schott",
-    B: [1.47460789, 0.193584488, 1.26589974],
-    C: [0.00986143816, 0.0445477583, 106.436258],
-    nd: 1.65412,
-    vd: 39.7,
-    PgF: 0.571,
-    code6: "654397",
-    source:
-      "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data. KZFS family — negative ΔPgF, APO-relevant.",
-  },
-
-  /* ────── Schott expansion (Phase 4 buildout, Apr 2026) ──────
-   * SF2: legacy dense flint (648/339), fills the gap between SF1 and N-BK7 in
-   *   pre-1990 Schott-dominant double-Gauss and Sonnar designs.
-   * N-SK16: dense lanthanum crown (620/603) ubiquitous in Zeiss, classic Nikon,
-   *   Leica Elmarit, and other pre-1990 normal/standard designs (also sold as
-   *   Hoya BACD5 — same glass, different vendor name).
-   * N-KZFS4: KZFS family with negative ΔPgF; APO paired glass for ED crowns.
-   * N-FK51A: fluorcrown with strong positive ΔPgF (+0.034); APO-relevant in
-   *   Zeiss and Nikon telephoto designs.
-   * N-SK11: barium crown used in Carl Zeiss zoom rear groups.
-   * F2: canonical Schott flint (620/364); dense lead flint in vintage designs.
-   * N-LAK22: lanthanum crown (651/559); common in compact zoom rear elements.
-   * All coefficients from SCHOTT Zemax catalog (schott_2017-01-20b.agf) via
-   * refractiveindex.info.
-   */
-  {
-    name: "SF2",
-    vendor: "Schott",
-    B: [1.40301821, 0.231767504, 0.939056586],
-    C: [0.0105795466, 0.0493226978, 112.405955],
-    nd: 1.64769,
-    vd: 33.85,
-    PgF: 0.5886,
-    code6: "648339",
-    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data.",
-  },
-  {
-    name: "N-SK16",
-    vendor: "Schott",
-    B: [1.34317774, 0.241144399, 0.994317969],
-    C: [0.00704687339, 0.0229005, 92.7508526],
-    nd: 1.62041,
-    vd: 60.32,
-    PgF: 0.5412,
-    code6: "620603",
-    source:
-      "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data. Dense lanthanum crown; Hoya BACD5-equivalent.",
-  },
-  {
-    name: "N-KZFS4",
-    vendor: "Schott",
-    B: [1.35055424, 0.197575506, 1.09962992],
-    C: [0.0087628207, 0.0371767201, 90.3866994],
-    nd: 1.61336,
-    vd: 44.49,
-    PgF: 0.559,
-    code6: "613445",
-    source:
-      "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data. KZFS family — negative ΔPgF, APO-relevant.",
-  },
-  {
-    name: "N-FK51A",
-    vendor: "Schott",
-    B: [0.971247817, 0.216901417, 0.904651666],
-    C: [0.00472301995, 0.0153575612, 168.68133],
-    nd: 1.48656,
-    vd: 84.47,
-    PgF: 0.5359,
-    code6: "487845",
-    source:
-      "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data. Fluorcrown with strong +ΔPgF (+0.034); APO-relevant.",
-  },
-  {
-    name: "N-SK11",
-    vendor: "Schott",
-    B: [1.17963631, 0.229817295, 0.935789652],
-    C: [0.00680282081, 0.0219737205, 101.513232],
-    nd: 1.56384,
-    vd: 60.8,
-    PgF: 0.5412,
-    code6: "564608",
-    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data. Barium crown.",
-  },
-  {
-    name: "F2",
-    vendor: "Schott",
-    B: [1.34533359, 0.209073176, 0.937357162],
-    C: [0.00997743871, 0.0470450767, 111.886764],
-    nd: 1.62004,
-    vd: 36.37,
-    PgF: 0.5829,
-    code6: "620364",
-    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data. Legacy lead flint.",
-  },
-  {
-    name: "N-LAK22",
-    vendor: "Schott",
-    B: [1.14229781, 0.535138441, 1.04088385],
-    C: [0.00585778594, 0.0198546147, 100.834017],
-    nd: 1.65113,
-    vd: 55.89,
-    PgF: 0.5467,
-    code6: "651559",
-    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data. Lanthanum crown.",
-  },
-
-  /* ────── Hoya — ED crowns and dense lanthanum flints ──────
-   * Hoya AGF entries use the polynomial Schott formula (form 1), not
-   * Sellmeier-1. The B/C below are a least-squares Sellmeier-1 fit to
-   * the polynomial sampled densely (80 points across 365–1014 nm). Max
-   * absolute index error across the spectrum is below 1.4e-5.
-   */
-  {
-    name: "FCD1",
-    vendor: "Hoya",
-    B: [0.708299951, 0.5095943307, 0.5656963028],
-    C: [0.002809999603, 0.01255522353, 101.5038807],
-    nd: 1.497,
-    vd: 81.61,
-    PgF: 0.544,
-    code6: "497816",
-    source:
-      "Hoya AGF (hoya_2017-04-01.agf) form-1 polynomial; Sellmeier-1 fit, max abs index error 1.4e-5 across 365–1014 nm. ED, S-FPL51-equivalent.",
-  },
-  {
-    name: "FCD505",
-    vendor: "Hoya",
-    B: [1.034776218, 0.4665786919, 0.6985808477],
-    C: [0.004826471098, 0.01646358686, 105.4578263],
-    nd: 1.59282,
-    vd: 68.63,
-    PgF: 0.5478,
-    code6: "593686",
-    source:
-      "Hoya AGF (hoya_2017-04-01.agf) form-1 polynomial; Sellmeier-1 fit, max abs index error 4.3e-6 across 365–1014 nm.",
-  },
-  {
-    name: "TAFD30",
-    vendor: "Hoya",
-    B: [1.660453863, 0.7793549404, 1.03002609],
-    C: [0.007957073706, 0.02889879311, 79.74654506],
-    nd: 1.883,
-    vd: 40.8,
-    PgF: 0.5658,
-    code6: "883408",
-    source:
-      "Hoya AGF (hoya_2017-04-01.agf) form-1 polynomial; Sellmeier-1 fit, max abs index error 7.1e-6 across 365–1014 nm.",
-  },
-  {
-    name: "TAFD37A",
-    vendor: "Hoya",
-    B: [2.188879041, 0.3049704899, 1.232253827],
-    C: [0.01151042519, 0.04599567769, 93.14898114],
-    nd: 1.90043,
-    vd: 37.37,
-    PgF: 0.5767,
-    code6: "900374",
-    source:
-      "Hoya Zemax catalog (hoya_2017-04-01, formula 3) via refractiveindex.info; Sellmeier-1 fit, max abs index error 5.1e-6 across 365–1014 nm. Ultra-high-index lanthanum; used in Voigtländer Nokton X 50mm f/1.2.",
-  },
-  {
-    name: "TAFD37",
-    vendor: "Hoya",
-    B: [2.259360214, 0.2343652859, 1.481042614],
-    C: [0.01221915793, 0.04989586099, 111.9367914],
-    nd: 1.90043,
-    vd: 37.37,
-    PgF: 0.5772,
-    source:
-      "Hoya Zemax catalog (hoya_2017-04-01, formula 3) via refractiveindex.info; Sellmeier-1 fit, max abs index error 3.7e-6 across 365–1014 nm. Distinct Sellmeier from TAFD37A.",
-  },
-  {
-    name: "TAFD33",
-    vendor: "Hoya",
-    B: [2.212737897, 0.2183360248, 1.27909893],
-    C: [0.01179423107, 0.04560326145, 100.6882193],
-    nd: 1.881,
-    vd: 40.14,
-    PgF: 0.5704,
-    code6: "881401",
-    source:
-      "Hoya Zemax catalog (hoya_2017-04-01, formula 3) via refractiveindex.info; Sellmeier-1 fit, max abs index error 1.0e-6 across 365–1014 nm. Dense lanthanum flint used in Ricoh GR III 28mm.",
-  },
-
-  /* ────── Sumita — APO-relevant K-prefix crowns ──────
-   * K-GFK68 appears in the Voigtländer APO-Lanthar 50/2 prescription. Sumita
-   * AGF entries use the polynomial Schott formula (form 1); B/C below are a
-   * least-squares Sellmeier-1 fit, max abs index error 6.7e-6 across 365–1014 nm.
-   */
-  {
-    name: "K-GFK68",
-    vendor: "Sumita",
-    B: [1.029009037, 0.4702661225, 0.6990302054],
-    C: [0.005049695326, 0.01621519052, 121.8498943],
-    nd: 1.5924,
-    vd: 68.3,
-    PgF: 0.5468,
-    code6: "592683",
-    source:
-      "Sumita AGF (sumita_2017-02-02.agf) form-1 polynomial; Sellmeier-1 fit, max abs index error 6.7e-6 across 365–1014 nm. APO-relevant.",
-  },
-  {
-    name: "K-SK7",
-    vendor: "Sumita",
-    polynomial: [2.5429277, -0.01031719, 0.014356944, 0.00031433872, -0.000010824341, 0.00000068100892],
-    nd: 1.60729,
-    vd: 59.5,
-    PgF: 0.5431,
-    code6: "607595",
-    source: "Sumita Zemax catalog 2017-02-02 via refractiveindex.info; K-SK7 page (formula 3 polynomial).",
-  },
-  {
-    name: "K-SK18",
-    vendor: "Sumita",
-    polynomial: [2.6383229, -0.010806956, 0.016153414, 0.00046673954, -0.000025414619, 0.0000019703165],
-    nd: 1.63854,
-    vd: 55.5,
-    PgF: 0.5497,
-    code6: "639555",
-    source: "Sumita Zemax catalog 2017-02-02 via refractiveindex.info; K-SK18 page (formula 3 polynomial).",
-  },
-
-  /* ────── Ohara Phase 5 additions (May 2026) ──────
-   * Sourced from OHARA Zemax catalog (ohara_2017-11-30.agf) via
-   * refractiveindex.info. All entries use formula 2 (standard Sellmeier-1,
-   * K=0). Covers dense flints, NBH niobophosphate family, LAH/LAM/LAL/BAL/BSM
-   * lanthanum and barium families, and NPH ultra-high-index flints not in the
-   * earlier phases.
-   */
-
-  /* Dense flints — TIH / TIM additions */
-  {
-    name: "S-TIH18",
-    vendor: "Ohara",
-    B: [1.59921608, 0.259532164, 2.12454543],
-    C: [0.0116469304, 0.0584824883, 186.927779],
-    nd: 1.721507,
-    vd: 29.23,
-    code6: "722292",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-  {
-    name: "S-TIH4",
-    vendor: "Ohara",
-    B: [1.66755531, 0.294411865, 2.49422119],
-    C: [0.0122052137, 0.0597775329, 214.869618],
-    nd: 1.755199,
-    vd: 27.51,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-  {
-    name: "S-TIM5",
-    vendor: "Ohara",
-    B: [1.38531342, 0.122372945, 1.40508326],
-    C: [0.0104074567, 0.0557440088, 144.878733],
-    nd: 1.60342,
-    vd: 38.03,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-  {
-    name: "S-TIM27",
-    vendor: "Ohara",
-    B: [1.4168047, 0.196785057, 1.68001322],
-    C: [0.0100732158, 0.0537616908, 164.672436],
-    nd: 1.639799,
-    vd: 34.47,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-  {
-    name: "S-NBH8",
-    vendor: "Ohara",
-    B: [1.61344136, 0.257295888, 1.98364455],
-    C: [0.0106386752, 0.0487071624, 159.784404],
-    nd: 1.720467,
-    vd: 34.71,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-
-  /* NBH niobophosphate family — anomalous partial dispersion dense flints */
-  {
-    name: "S-NBH51",
-    vendor: "Ohara",
-    B: [1.71203689, 0.255989588, 1.81456998],
-    C: [0.0107724134, 0.0488593504, 136.359013],
-    nd: 1.749504,
-    vd: 35.33,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-  {
-    name: "S-NBH52",
-    vendor: "Ohara",
-    B: [1.50305799, 0.221715926, 1.84496391],
-    C: [0.00999021738, 0.0450327698, 163.722302],
-    nd: 1.672999,
-    vd: 38.15,
-    code6: "673382",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-  {
-    name: "S-NBH55",
-    vendor: "Ohara",
-    B: [1.83145156, 0.287818024, 2.152083],
-    C: [0.0122443139, 0.057387731, 186.099124],
-    nd: 1.799999,
-    vd: 29.84,
-    code6: "800298",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-  {
-    name: "S-NBH56",
-    vendor: "Ohara",
-    B: [1.85191438, 0.431102852, 3.45278284],
-    C: [0.013273262, 0.0585944644, 239.357089],
-    nd: 1.854779,
-    vd: 24.8,
-    code6: "855248",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-
-  /* LAH lanthanum dense flint additions */
-  {
-    name: "S-LAH55VS",
-    vendor: "Ohara",
-    B: [1.92591095, 0.34895346, 1.42230744],
-    C: [0.0096115249, 0.036513298, 103.36409],
-    nd: 1.834809,
-    vd: 42.74,
-    source:
-      "Ohara Zemax catalog 2017-11-30 via refractiveindex.info. Vacuum-melt stabilized variant of S-LAH55V; distinct Sellmeier.",
-  },
-  {
-    name: "S-LAH60",
-    vendor: "Ohara",
-    B: [1.95243469, 0.30710021, 1.56578094],
-    C: [0.0106442437, 0.0456735302, 110.28141],
-    nd: 1.834,
-    vd: 37.16,
-    code6: "834372",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-  {
-    name: "S-LAH60V",
-    vendor: "Ohara",
-    B: [2.05081962, 0.208475257, 1.31486394],
-    C: [0.0116035991, 0.0526489359, 99.38065],
-    nd: 1.833999,
-    vd: 37.21,
-    source:
-      "Ohara Zemax catalog 2017-11-30 via refractiveindex.info. Vacuum-melt variant; distinct Sellmeier from S-LAH60.",
-  },
-  {
-    name: "S-LAH64",
-    vendor: "Ohara",
-    B: [1.83021453, 0.29156359, 1.28544024],
-    C: [0.0090482329, 0.0330756689, 89.3675501],
-    nd: 1.788001,
-    vd: 47.37,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-  {
-    name: "S-LAH89",
-    vendor: "Ohara",
-    B: [1.95118827, 0.377607223, 1.47757262],
-    C: [0.00976560799, 0.0382232043, 112.23672],
-    nd: 1.8515,
-    vd: 40.78,
-    code6: "852408",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-  {
-    name: "S-LAH97",
-    vendor: "Ohara",
-    B: [1.0273018, 0.989293564, 1.25781057],
-    C: [0.0183406129, 0.00371264195, 87.85105],
-    nd: 1.755,
-    vd: 52.32,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-
-  /* LAM / LAL lanthanum crown additions */
-  {
-    name: "S-LAM2",
-    vendor: "Ohara",
-    B: [1.7713, 0.19581423, 1.19487834],
-    C: [0.00976652444, 0.0412718628, 110.458122],
-    nd: 1.743997,
-    vd: 44.79,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-  {
-    name: "S-LAM51",
-    vendor: "Ohara",
-    B: [1.638472, 0.188330533, 1.47502357],
-    C: [0.00904853452, 0.0372740173, 137.77005],
-    nd: 1.699998,
-    vd: 48.08,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-  {
-    name: "S-LAL8",
-    vendor: "Ohara",
-    B: [1.30663291, 0.571377253, 1.24303605],
-    C: [0.00611862448, 0.021272147, 90.6285686],
-    nd: 1.712995,
-    vd: 53.87,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-  {
-    name: "S-LAL9",
-    vendor: "Ohara",
-    B: [1.16195687, 0.644860099, 1.25062221],
-    C: [0.0159659509, 0.000505502467, 93.8284169],
-    nd: 1.691002,
-    vd: 54.82,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-
-  /* BAM / BAL barium additions */
-  {
-    name: "S-BAM4",
-    vendor: "Ohara",
-    B: [1.41059317, 0.111201306, 1.34148939],
-    C: [0.00963312192, 0.049877821, 152.237696],
-    nd: 1.60562,
-    vd: 43.71,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-  {
-    name: "S-BAL14",
-    vendor: "Ohara",
-    B: [1.27553696, 0.146083393, 1.16754699],
-    C: [0.00749692359, 0.031042153, 128.947092],
-    nd: 1.568832,
-    vd: 56.36,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-
-  /* BSM barium crown additions */
-  {
-    name: "S-BSM18",
-    vendor: "Ohara",
-    B: [0.927886025, 0.708858526, 1.18610897],
-    C: [0.00417549199, 0.0184691838, 122.210416],
-    nd: 1.638539,
-    vd: 55.38,
-    code6: "639554",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-  {
-    name: "S-BSM81",
-    vendor: "Ohara",
-    B: [0.996356844, 0.651392837, 1.22432622],
-    C: [0.0144821587, 0.00154826389, 89.9818604],
-    nd: 1.639999,
-    vd: 60.08,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-
-  /* NSL light crown addition */
-  {
-    name: "S-NSL5",
-    vendor: "Ohara",
-    B: [1.04574577, 0.239613026, 1.1590685],
-    C: [0.0058523228, 0.0236858752, 131.329061],
-    nd: 1.522494,
-    vd: 59.84,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-
-  /* NPH ultra-high-index dense flint addition */
-  {
-    name: "S-NPH4",
-    vendor: "Ohara",
-    B: [1.93563931, 0.449596478, 2.71828573],
-    C: [0.0152585289, 0.0696815778, 170.327149],
-    nd: 1.89286,
-    vd: 20.36,
-    code6: "893204",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
-  },
-
-  /* ────── Schott Phase 5 additions (May 2026) ──────
-   * N-SK10 and N-SK14 are barium / lanthanum crowns found in legacy Schott-
-   * dominant designs; SK10 and SK14 are the deprecated pre-2000 names.
-   * Coefficients from Schott Zemax catalog (schott_2017-01-20b.agf) via
-   * refractiveindex.info.
-   */
-  {
-    name: "N-SK10",
-    vendor: "Schott",
-    B: [1.34972093, 0.238587973, 0.9667336],
-    C: [0.00736272269, 0.0253765327, 103.502909],
-    nd: 1.62278,
-    vd: 56.98,
-    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data.",
-  },
-  {
-    name: "N-SK14",
-    vendor: "Schott",
-    B: [0.936155374, 0.594052018, 1.04374583],
-    C: [0.00461716525, 0.016885927, 103.736265],
-    nd: 1.60311,
-    vd: 60.6,
-    code6: "603606",
-    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data.",
-  },
-
-  /* ────── Phase 6 priority additions (May 2026) ──────
-   * High-impact unresolved glasses from the catalog-priority sweep. Formula-2
-   * entries are standard Sellmeier. Formula-3 Hoya/Sumita entries store the
-   * vendor polynomial exactly rather than using a fitted Sellmeier surrogate.
-   */
-
-  /* Ohara priority additions */
-  {
-    name: "S-FPL55",
-    vendor: "Ohara",
-    B: [0.839067682, 0.214083503, 1.47914677],
-    C: [0.00795286639, -0.00232581717, 340.0437],
-    nd: 1.43875,
-    vd: 94.661059,
-    PgF: 0.5303,
-    code6: "439948",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-FPL55 page.",
-  },
-  {
-    name: "S-LAH98",
-    vendor: "Ohara",
-    B: [2.28510629, 0.385532264, 2.0655112],
-    C: [0.0122178962, 0.0514752342, 145.92087],
-    nd: 1.95375,
-    vd: 32.32,
-    PgF: 0.5907,
-    code6: "954323",
-    source: "Ohara S-LAH98 datasheet (Ohara public PDF, OHARA 24-01).",
-  },
-  {
-    name: "S-LAH99",
-    vendor: "Ohara",
-    B: [2.39140662, 0.439219228, 2.38358467],
-    C: [0.01314675, 0.0553226042, 161.2599],
-    nd: 2.001,
-    vd: 29.14,
-    PgF: 0.5997,
-    code6: "001291",
-    source: "Ohara S-LAH99 datasheet (Ohara public PDF, 2026 page crawl).",
-  },
-  {
-    name: "L-LAH85V",
-    vendor: "Ohara",
-    B: [1.97460503, 0.361903592, 1.38433241],
-    C: [0.0101854258, 0.0384829118, 103.713627],
-    nd: 1.854,
-    vd: 40.378368,
-    PgF: 0.5688,
-    code6: "854404",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; L-LAH85V page.",
-  },
-  {
-    name: "S-LAM60",
-    vendor: "Ohara",
-    B: [1.60673056, 0.36641564, 1.31761804],
-    C: [0.0077504614, 0.0289967611, 93.0720709],
-    nd: 1.743198,
-    vd: 49.339444,
-    PgF: 0.5523,
-    code6: "743493",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAM60 page.",
-  },
-  {
-    name: "S-NPH3",
-    vendor: "Ohara",
-    B: [2.09834903, 0.489088388, 2.94009268],
-    C: [0.0179123869, 0.0776653353, 160.930428],
-    nd: 1.95906,
-    vd: 17.471252,
-    PgF: 0.661,
-    code6: "959175",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-NPH3 page.",
-  },
-  {
-    name: "S-NPH5",
-    vendor: "Ohara",
-    B: [1.89108996, 0.395220126, 2.20492127],
-    C: [0.0141164499, 0.0662834445, 148.6807],
-    nd: 1.858955,
-    vd: 22.728736,
-    PgF: 0.6293,
-    code6: "859227",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-NPH5 page.",
-  },
-  {
-    name: "S-LAH63Q",
-    vendor: "Ohara",
-    B: [1.96723017, 0.194953915, 1.25386282],
-    C: [0.0110456086, 0.0497137061, 104.84352],
-    nd: 1.8044,
-    vd: 39.580489,
-    PgF: 0.576,
-    code6: "804396",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAH63Q page.",
-  },
-  {
-    name: "S-LAH65VS",
-    vendor: "Ohara",
-    B: [1.76068422, 0.414128906, 1.33415439],
-    C: [0.00853607198, 0.0301826383, 98.09421],
-    nd: 1.804,
-    vd: 46.527532,
-    PgF: 0.557,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAH65VS page.",
-  },
-  {
-    name: "S-NBM51",
-    vendor: "Ohara",
-    B: [1.37023101, 0.177665568, 1.30515471],
-    C: [0.00871920342, 0.0405725552, 112.703058],
-    nd: 1.613397,
-    vd: 44.267706,
-    PgF: 0.5628,
-    code6: "613443",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-NBM51 page.",
-  },
-  {
-    name: "S-FTM16",
-    vendor: "Ohara",
-    B: [1.32940907, 0.141512125, 1.44299068],
-    C: [0.0102377287, 0.0578081956, 150.597139],
-    nd: 1.592701,
-    vd: 35.310075,
-    PgF: 0.5934,
-    code6: "593353",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-FTM16 page.",
-  },
-  {
-    name: "S-NSL36",
-    vendor: "Ohara",
-    B: [1.09666153, 0.168990073, 1.20580827],
-    C: [0.00667491123, 0.033609545, 141.668738],
-    nd: 1.517417,
-    vd: 52.430903,
-    PgF: 0.5554,
-    code6: "517524",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-NSL36 page.",
-  },
-
-  /* Schott priority additions */
-  {
-    name: "N-SF66",
-    vendor: "Schott",
-    B: [2.0245976, 0.470187196, 2.59970433],
-    C: [0.0147053225, 0.0692998276, 161.817601],
-    nd: 1.92286,
-    vd: 20.88,
-    PgF: 0.6394,
-    code6: "923209",
-    source: "Schott AGF (schott_2017-01-20b.agf) via refractiveindex.info; N-SF66 page.",
-  },
-  {
-    name: "N-LAF21",
-    vendor: "Schott",
-    B: [1.87134529, 0.25078301, 1.22048639],
-    C: [0.0093332228, 0.0345637762, 83.2404866],
-    nd: 1.788,
-    vd: 47.49,
-    PgF: 0.5555,
-    code6: "788475",
-    source: "Schott AGF (schott_2017-01-20b.agf) via refractiveindex.info; N-LAF21 page.",
-  },
-  {
-    name: "SF11",
-    vendor: "Schott",
-    B: [1.73848403, 0.311168974, 1.17490871],
-    C: [0.0136068604, 0.0615960463, 121.922711],
-    nd: 1.78472,
-    vd: 25.76,
-    PgF: 0.6147,
-    code6: "785258",
-    source: "Schott AGF (schott_2017-01-20b.agf) via refractiveindex.info; SF11 page.",
-  },
-  {
-    name: "N-SK2",
-    vendor: "Schott",
-    B: [1.28189012, 0.257738258, 0.96818604],
-    C: [0.0072719164, 0.0242823527, 110.377773],
-    nd: 1.60738,
-    vd: 56.65,
-    PgF: 0.5477,
-    code6: "607567",
-    source: "Schott AGF (schott_2017-01-20b.agf) via refractiveindex.info; N-SK2 page.",
-  },
-
-  /* Hoya priority additions */
-  {
-    name: "E-FDS1",
-    vendor: "Hoya",
-    polynomial: [3.5044639, -0.020333833, 0.0538027, 0.0062874192, -0.00056922176, 0.000073559387],
-    nd: 1.92286,
-    vd: 20.88,
-    PgF: 0.6369,
-    code6: "923209",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FDS1 page (formula 3 polynomial).",
-  },
-  {
-    name: "FCD100",
-    vendor: "Hoya",
-    polynomial: [2.0482157, -0.0043211222, 0.0061826755, 0.00003141148, 0.0000035370793, -0.00000016694497],
-    nd: 1.437,
-    vd: 95.1,
-    PgF: 0.54,
-    code6: "437951",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FCD100 page (formula 3 polynomial).",
-  },
-  {
-    name: "FCD515",
-    vendor: "Hoya",
-    polynomial: [2.502355, -0.0072080042, 0.012012743, 0.00034920842, -0.000024729676, 0.0000013801567],
-    nd: 1.59282,
-    vd: 68.63,
-    PgF: 0.5478,
-    code6: "593686",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FCD515 page (formula 3 polynomial).",
-  },
-  {
-    name: "FCD705",
-    vendor: "Hoya",
-    polynomial: [2.3742832, -0.0062014418, 0.010582967, 0.000063100703, 0.000008196715, -0.00000034424446],
-    nd: 1.55032,
-    vd: 75.5,
-    PgF: 0.5445,
-    code6: "550755",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FCD705 page (formula 3 polynomial).",
-  },
-  {
-    name: "TAFD45",
-    vendor: "Hoya",
-    polynomial: [3.672962, -0.0151805, 0.0457187, 0.002186221, -0.0000993742, 0.00001490466],
-    nd: 1.95375,
-    vd: 32.32,
-    code6: "954323",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD45 page (formula 3 polynomial).",
-  },
-  {
-    name: "TAFD55",
-    vendor: "Hoya",
-    polynomial: [3.8317578, -0.015266397, 0.054573843, 0.0023109346, -0.000044476201, 0.000016143887],
-    nd: 2.001,
-    vd: 29.13,
-    PgF: 0.5986,
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD55 page (formula 3 polynomial).",
-  },
-  {
-    name: "E-CF6",
-    vendor: "Hoya",
-    polynomial: [2.2668912, -0.0093866212, 0.012174195, 0.0005076903, -0.000033573414, 0.0000029059013],
-    nd: 1.51742,
-    vd: 52.15,
-    PgF: 0.5561,
-    code6: "517522",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-CF6 page (formula 3 polynomial).",
-  },
-  {
-    name: "MC-TAF101-100",
-    vendor: "Hoya",
-    polynomial: [3.0591566, -0.013604717, 0.024624272, 0.00041834242, 0.0000025361908, 0.0000007998082],
-    nd: 1.76902,
-    vd: 49.29,
-    PgF: 0.5627,
-    code6: "769493",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; MC-TAF101-100 page (formula 3 polynomial).",
-  },
-  {
-    name: "NBFD15",
-    vendor: "Hoya",
-    polynomial: [3.1540011, -0.013256753, 0.034103312, 0.0018472167, -0.00010807332, 0.000012975901],
-    nd: 1.8061,
-    vd: 33.27,
-    code6: "806333",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; NBFD15 page (formula 3 polynomial).",
-  },
-  {
-    name: "NBFD25",
-    vendor: "Hoya",
-    polynomial: [3.2879389, -0.015856356, 0.045638807, 0.0033027661, -0.00021690686, 0.000029625863],
-    nd: 1.85451,
-    vd: 25.15,
-    PgF: 0.6103,
-    code6: "855252",
-    source: "HOYA Optical Glass Catalogue NBFD25 PDF, HOYA20250701M.xlsm; formula-3 A0-A5 constants, d-code 855-252.",
-  },
-  {
-    name: "NBFD29",
-    vendor: "Hoya",
-    polynomial: [3.023484, -0.014446333, 0.034673303, 0.0020637437, -0.00011126222, 0.000014468973],
-    nd: 1.77047,
-    vd: 29.74,
-    PgF: 0.5951,
-    code6: "770297",
-    source: "HOYA Optical Glass Catalogue NBFD29 PDF, HOYA20250701M.xlsm; formula-3 A0-A5 constants, d-code 770-297.",
-  },
-  {
-    name: "FDS24",
-    vendor: "Hoya",
-    polynomial: [3.5126327, -0.0153015, 0.053936061, 0.0034949369, -0.00017950295, 0.000034403181],
-    nd: 1.92119,
-    vd: 23.96,
-    PgF: 0.6465,
-    code6: "921240",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FDS24 page (formula 3 polynomial).",
-  },
-  {
-    name: "BACD15",
-    vendor: "Hoya",
-    polynomial: [2.5864832, -0.0080357829, 0.018080573, -0.00045974172, 0.000088484321, -0.0000037544765],
-    nd: 1.62299,
-    vd: 58.12,
-    PgF: 0.5439,
-    code6: "623581",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; BACD15 page (formula 3 polynomial).",
-  },
-  {
-    name: "TAFD40",
-    vendor: "Hoya",
-    polynomial: [3.8106232, -0.016351565, 0.058801097, 0.0034919026, -0.0001694811, 0.000032569333],
-    nd: 2.00069,
-    vd: 25.46,
-    PgF: 0.6121,
-    code6: "001255",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD40 page (formula 3 polynomial).",
-  },
-
-  /* ────── Phase 8 unresolved-token additions (May 2026) ──────
-   * Clean, sourceable names from the unresolved-glass report. Ambiguous
-   * high-frequency labels such as NBFD3, S-NPH7, and TAFD25 are intentionally
-   * left for per-lens patent relabeling rather than broad catalog resolution.
-   */
-
-  /* Ohara unresolved-token additions */
-  {
-    name: "S-LAH93",
-    vendor: "Ohara",
-    B: [2.17884378, 0.325508683, 1.70608646],
-    C: [0.0115760879, 0.0495982503, 128.91358],
-    nd: 1.90525,
-    vd: 35.036826,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAH93 page.",
-  },
-  {
-    name: "S-TIH11",
-    vendor: "Ohara",
-    B: [1.72677471, 0.324568628, 2.65816809],
-    C: [0.0129369958, 0.0618255245, 221.904637],
-    nd: 1.784723,
-    vd: 25.683446,
-    PgF: 0.6168,
-    code6: "785257",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIH11 page.",
-  },
-  {
-    name: "S-TIL25",
-    vendor: "Ohara",
-    B: [1.32122534, 0.123824976, 1.43685254],
-    C: [0.00952091436, 0.0516062665, 149.064883],
-    nd: 1.581439,
-    vd: 40.747566,
-    PgF: 0.5772,
-    code6: "581407",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIL25 page.",
-  },
-  {
-    name: "S-BAH27",
-    vendor: "Ohara",
-    B: [1.68939052, 0.133081013, 1.41165515],
-    C: [0.0103598193, 0.0533982239, 126.515503],
-    nd: 1.701536,
-    vd: 41.23944,
-    PgF: 0.5762,
-    code6: "702412",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BAH27 page.",
-  },
-  {
-    name: "S-NBH52V",
-    vendor: "Ohara",
-    B: [1.51336868, 0.212341478, 1.54149143],
-    C: [0.00987077827, 0.0462843662, 126.97851],
-    nd: 1.673,
-    vd: 38.256077,
-    PgF: 0.5756,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-NBH52V page.",
-  },
-  {
-    name: "S-TIH53W",
-    vendor: "Ohara",
-    B: [1.87904886, 0.369719775, 2.33730863],
-    C: [0.014412177, 0.063881799, 182.66818],
-    nd: 1.84666,
-    vd: 23.77794,
-    PgF: 0.6213,
-    code6: "847238",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIH53W page. Improved-transmittance variant.",
-  },
-  {
-    name: "S-BAH28",
-    vendor: "Ohara",
-    B: [1.69493484, 0.192890298, 1.56385948],
-    C: [0.010272319, 0.052118764, 137.818035],
-    nd: 1.72342,
-    vd: 37.955602,
-    PgF: 0.5835,
-    code6: "723380",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BAH28 page.",
-  },
-  {
-    name: "S-LAL59",
-    vendor: "Ohara",
-    B: [1.13962742, 0.805227838, 1.29488061],
-    C: [0.00493294862, 0.020247996, 93.4746507],
-    nd: 1.733997,
-    vd: 51.47057,
-    PgF: 0.5476,
-    code6: "734515",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAL59 page.",
-  },
-  {
-    name: "S-TIL27",
-    vendor: "Ohara",
-    B: [1.31433154, 0.112300168, 1.413901],
-    C: [0.00950404477, 0.0524112772, 148.429972],
-    nd: 1.575006,
-    vd: 41.505207,
-    PgF: 0.5764,
-    code6: "575415",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIL27 page.",
-  },
-  {
-    name: "S-TIL6",
-    vendor: "Ohara",
-    B: [1.17701777, 0.12795803, 1.34740124],
-    C: [0.00771087686, 0.0411325328, 154.531692],
-    nd: 1.531717,
-    vd: 48.840718,
-    PgF: 0.5623,
-    code6: "532489",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIL6 page.",
-  },
-
-  /* Schott unresolved-token addition */
-  {
-    name: "N-FK5",
-    vendor: "Schott",
-    B: [0.844309338, 0.344147824, 0.910790213],
-    C: [0.00475111955, 0.0149814849, 97.8600293],
-    nd: 1.48749,
-    vd: 70.41,
-    PgF: 0.529,
-    code6: "487704",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-FK5 page.",
-  },
-
-  /* Hoya unresolved-token additions */
-  {
-    name: "FC5",
-    vendor: "Hoya",
-    polynomial: [2.1894054, -0.0099044908, 0.008640337, 0.00022263067, -0.000012291942, 0.00000059386349],
-    nd: 1.48749,
-    vd: 70.45,
-    PgF: 0.5345,
-    code6: "487704",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FC5 page (formula 3 polynomial).",
-  },
-  {
-    name: "E-FD15",
-    vendor: "Hoya",
-    polynomial: [2.7919943, -0.012311206, 0.028802741, 0.0021430059, -0.00016932119, 0.000019741766],
-    nd: 1.69895,
-    vd: 30.05,
-    PgF: 0.6019,
-    code6: "699301",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FD15 page (formula 3 polynomial).",
-  },
-  {
-    name: "TAFD5F",
-    vendor: "Hoya",
-    polynomial: [3.275546, -0.014470354, 0.030922062, 0.00073289981, 0.0000046573307, 0.0000018242746],
-    nd: 1.83481,
-    vd: 42.72,
-    PgF: 0.5657,
-    code6: "835427",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD5F page (formula 3 polynomial).",
-  },
-
-  /* ────── Phase 9 named-token additions (May 2026) ──────
-   * Directly sourceable 2+ occurrence names whose catalog constants match at
-   * least one current lens-data declaration. Ambiguous names from the same
-   * survey remain deferred to per-lens patent relabeling.
-   */
-
-  /* Ohara Phase 9 additions */
-  {
-    name: "S-BSM16",
-    vendor: "Ohara",
-    B: [1.14490383, 0.439563911, 1.27688079],
-    C: [0.0137034916, -0.00186514205, 119.535585],
-    nd: 1.620411,
-    vd: 60.289582,
-    PgF: 0.5412,
-    code6: "620603",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BSM16 page.",
-  },
-  {
-    name: "S-TIH13",
-    vendor: "Ohara",
-    B: [1.62224674, 0.293844589, 1.99225164],
-    C: [0.0118368386, 0.0590208025, 171.959976],
-    nd: 1.740769,
-    vd: 27.789359,
-    PgF: 0.6101,
-    code6: "741278",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIH13 page.",
-  },
-  {
-    name: "S-NBH53V",
-    vendor: "Ohara",
-    B: [1.65444141, 0.267453927, 2.14530347],
-    C: [0.0112485533, 0.052027274, 167.3661],
-    nd: 1.737999,
-    vd: 32.32659,
-    PgF: 0.5902,
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-NBH53V page.",
-  },
-  {
-    name: "S-TIH23",
-    vendor: "Ohara",
-    B: [1.73986485, 0.313894918, 2.31093206],
-    C: [0.01294413, 0.0612116868, 197.420482],
-    nd: 1.784696,
-    vd: 26.291221,
-    PgF: 0.6142,
-    code6: "785263",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIH23 page.",
-  },
-
-  /* Schott current datasheet additions */
-  {
-    name: "SF56A",
-    vendor: "Schott",
-    B: [1.70579259, 0.344223052, 1.09601828],
-    C: [0.01338747, 0.0579561608, 121.616024],
-    nd: 1.7847,
-    vd: 26.08,
-    PgF: 0.6098,
-    code6: "785261",
-    source: "Schott optical glass datasheet SF56A, public PDF (as of 2014-02-01; media.schott.com content 7428154c).",
-  },
-  {
-    name: "N-BALF4",
-    vendor: "Schott",
-    B: [1.31004128, 0.142038259, 0.964929351],
-    C: [0.007965965, 0.0330672072, 109.19732],
-    nd: 1.57956,
-    vd: 53.87,
-    PgF: 0.552,
-    code6: "580539",
-    source: "Schott optical glass datasheet N-BALF4, public PDF (as of 2014-02-01; media.schott.com content 710baa54).",
-  },
-
-  /* Schott Phase 9 additions */
-  {
-    name: "N-SK5",
-    vendor: "Schott",
-    B: [0.991463823, 0.495982121, 0.987393925],
-    C: [0.00522730467, 0.0172733646, 98.3594579],
-    nd: 1.58913,
-    vd: 61.27,
-    PgF: 0.54,
-    code6: "589613",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-SK5 page.",
-  },
-  {
-    name: "N-BAF10",
-    vendor: "Schott",
-    B: [1.5851495, 0.143559385, 1.08521269],
-    C: [0.00926681282, 0.0424489805, 105.613573],
-    nd: 1.67003,
-    vd: 47.11,
-    PgF: 0.563,
-    code6: "670471",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-BAF10 page.",
-  },
-  {
-    name: "N-LAF34",
-    vendor: "Schott",
-    B: [1.75836958, 0.313537785, 1.18925231],
-    C: [0.00872810026, 0.0293020832, 85.1780644],
-    nd: 1.7725,
-    vd: 49.62,
-    PgF: 0.5518,
-    code6: "773496",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-LAF34 page.",
-  },
-  {
-    name: "N-LAK14",
-    vendor: "Schott",
-    B: [1.50781212, 0.318866829, 1.14287213],
-    C: [0.00746098727, 0.0242024834, 80.9565165],
-    nd: 1.6968,
-    vd: 55.41,
-    PgF: 0.5427,
-    code6: "697554",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-LAK14 page.",
-  },
-
-  /* Hoya Phase 9 additions */
-  {
-    name: "E-FD2",
-    vendor: "Hoya",
-    polynomial: [2.6399644, -0.011645718, 0.023061733, 0.0017348219, -0.00013986059, 0.00001391667],
-    nd: 1.64769,
-    vd: 33.84,
-    PgF: 0.5918,
-    code6: "648338",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FD2 page (formula 3 polynomial).",
-  },
-  {
-    name: "E-FD8",
-    vendor: "Hoya",
-    polynomial: [2.7617586, -0.011254039, 0.028221737, 0.0017540315, -0.00012127053, 0.000015522384],
-    nd: 1.68893,
-    vd: 31.16,
-    PgF: 0.5981,
-    code6: "689312",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FD8 page (formula 3 polynomial).",
-  },
-  {
-    name: "FF8",
-    vendor: "Hoya",
-    polynomial: [2.9564361, -0.019762195, 0.029878823, 0.0051830852, -0.00054202389, 0.000048463245],
-    nd: 1.75211,
-    vd: 25.05,
-    code6: "752251",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FF8 page (formula 3 polynomial).",
-  },
-  {
-    name: "TAFD35",
-    vendor: "Hoya",
-    polynomial: [3.5274747, -0.014839948, 0.039745402, 0.001803472, -0.000089279334, 0.000011377783],
-    nd: 1.91082,
-    vd: 35.25,
-    PgF: 0.5819,
-    code6: "911353",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD35 page (formula 3 polynomial).",
-  },
-  {
-    name: "H-ZLaF68L",
-    vendor: "NHG",
-    polynomial: [3.43867274, -0.0143248055, 0.0349115425, 0.0014530167, -0.0000762581625, 0.0000074766879],
-    nd: 1.883,
-    vd: 39.22,
-    PgF: 0.5698,
-    code6: "883392",
-    source: "NHG/Hubei New Huaguang H-ZLaF68L data sheet (d-code 883392, formula constants A0-A5).",
-  },
-
-  /* Phase 9 six-digit code-family additions */
-  {
-    name: "S-LAH96",
-    vendor: "Ohara",
-    B: [1.85078519, 0.189204854, 1.19763137],
-    C: [0.00940657541, 0.0380345187, 101.426835],
-    nd: 1.76385,
-    vd: 48.488692,
-    PgF: 0.5581,
-    code6: "764485",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAH96 page.",
-  },
-  {
-    name: "H-ZF88",
-    vendor: "CDGM",
-    B: [2.04275349, 1.98276153, 0.504267829],
-    C: [0.0169742976, 109.068151, 0.075171071],
-    nd: 1.94595,
-    vd: 17.941964,
-    PgF: 0.6537,
-    code6: "946179",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZF88 page.",
-  },
-  {
-    name: "H-ZLaF76",
-    vendor: "CDGM",
-    B: [0.279518569, 2.01429741, 1.5507336],
-    C: [0.0593856526, 0.012625744, 100.032291],
-    nd: 1.85013,
-    vd: 30.06,
-    PgF: 0.6001,
-    code6: "850301",
-    source: "CDGM H-ZLaF76 public datasheet, 2026 access; Sellmeier constants K1-L3, d-code 850301.",
-  },
-  {
-    name: "LAC13",
-    vendor: "Hoya",
-    polynomial: [2.8125316, -0.012915271, 0.019960339, 0.00019877888, 0.000017647061, -0.00000044767632],
-    nd: 1.6935,
-    vd: 53.34,
-    PgF: 0.5475,
-    code6: "694533",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; LAC13 page (formula 3 polynomial).",
-  },
-  {
-    name: "J-PSKH1",
-    vendor: "Hikari",
-    powerSeries: [
-      [2.50208083, 0],
-      [-0.00672143907, 2],
-      [-0.0000534313751, 4],
-      [0.01282644, -2],
-      [0.000156205388, -4],
-      [0.00000121593549, -6],
-      [0.0000000959550869, -8],
-    ],
-    nd: 1.59319,
-    vd: 67.900555,
-    PgF: 0.544,
-    code6: "593679",
-    source:
-      "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; Hikari Optical Glass Catalog 2023, J-PSKH1 page (formula 3 power series, d-code 593679).",
-  },
-  {
-    name: "J-PSKH4",
-    vendor: "Hikari",
-    powerSeries: [
-      [2.50453078, 0],
-      [-0.0101597822, 2],
-      [-0.000108653142, 4],
-      [0.0127723327, -2],
-      [0.000133860625, -4],
-      [0.00000337285381, -6],
-      [-0.0000000256491019, -8],
-    ],
-    nd: 1.59349,
-    vd: 67.001281,
-    PgF: 0.5358,
-    code6: "593670",
-    source:
-      "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; Hikari Optical Glass Catalog 2023, J-PSKH4 page (formula 3 power series, d-code 593670).",
-  },
-  {
-    name: "J-KZFH9",
-    vendor: "Hikari",
-    powerSeries: [
-      [2.92190512, 0],
-      [-0.0131913454, 2],
-      [-0.0000794286252, 4],
-      [0.0327997049, -2],
-      [0.000700950165, -4],
-      [0.00012416909, -6],
-      [-0.0000112359582, -8],
-      [0.000000911052912, -10],
-    ],
-    nd: 1.738,
-    vd: 32.26,
-    PgF: 0.5896,
-    code6: "738323",
-    source: "Hikari Optical Glass Catalog 2023, J-KZFH9 page (formula 3 power series, d-code 738323).",
-  },
-  {
-    name: "BAF3",
-    vendor: "Hikari",
-    polynomial: [2.438682, 0.001609961, 0.02688829, -0.002446579, 0.0003965661, -0.00001973046],
-    nd: 1.58267,
-    vd: 46.476929,
-    PgF: 0.5673,
-    code6: "583465",
-    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; BAF3 page (formula 3 polynomial).",
-  },
-  {
-    name: "J-LAK10",
-    vendor: "Hikari",
-    powerSeries: [
-      [2.89571408, 0],
-      [-0.0120013315, 2],
-      [-0.000136916169, 4],
-      [0.0219522159, -2],
-      [0.000357973143, -4],
-      [0.00000826304425, -6],
-      [0.00000027388172, -8],
-    ],
-    nd: 1.71999,
-    vd: 50.27,
-    PgF: 0.5527,
-    code6: "720503",
-    source:
-      "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; Hikari Optical Glass Catalog 2023, J-LAK10 page (formula 3 power series, d-code 720503).",
-  },
-  {
-    name: "J-LASFH15",
-    vendor: "Hikari",
-    powerSeries: [
-      [3.64640666, 0],
-      [-0.0151039558, 2],
-      [0.0480157444, -2],
-      [0.00349072452, -4],
-      [-0.000573639028, -6],
-      [0.000130249514, -8],
-      [-0.0000129576789, -10],
-      [0.000000592144355, -12],
-    ],
-    nd: 1.95,
-    vd: 29.37,
-    PgF: 0.6001,
-    code6: "950294",
-    source:
-      "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; Hikari Optical Glass Catalog 2023, J-LASFH15 page (formula 3 power series, d-code 950294).",
-  },
-  {
-    name: "J-LASFH9",
-    vendor: "Hikari",
-    powerSeries: [
-      [3.49709032, 0],
-      [-0.0138248635, 2],
-      [-0.000132935974, 4],
-      [0.0409278939, -2],
-      [0.000880758278, -4],
-      [0.0000893285504, -6],
-      [-0.0000066176262, -8],
-      [0.000000624399823, -10],
-    ],
-    nd: 1.90265,
-    vd: 35.725382,
-    PgF: 0.5805,
-    code6: "903357",
-    source:
-      "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; Hikari Optical Glass Catalog 2012, J-LASFH9 page (formula 3 power series, d-code 903357).",
-  },
-
-  /* ────── Phase 10 research additions (May 2026) ──────
-   * Batch sourced from the post-Phase-9 unresolved and relabel reports. Hoya
-   * formula-3 entries and Hikari power-series entries keep the vendor
-   * published form; Ohara entries use standard Sellmeier-1 coefficients.
-   */
-
-  /* Hoya formula-3 additions */
-  {
-    name: "FDS18",
-    vendor: "Hoya",
-    polynomial: [3.563784, -0.025449644, 0.058401659, 0.0092310983, -0.00095499727, 0.00011777456],
-    nd: 1.94595,
-    vd: 17.98,
-    PgF: 0.6522,
-    code6: "946180",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FDS18 page (formula 3 polynomial).",
-  },
-  {
-    name: "NBFD3",
-    vendor: "Hoya",
-    polynomial: [3.1633746, -0.01327628, 0.031133309, 0.0008352338, 0.0000029980499, 0.0000025568853],
-    nd: 1.8045,
-    vd: 39.63,
-    PgF: 0.5715,
-    code6: "805396",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; NBFD3 page (formula 3 polynomial).",
-  },
-  {
-    name: "NBFD11",
-    vendor: "Hoya",
-    polynomial: [3.1010267, -0.010022108, 0.032631687, -0.00081029268, 0.0001935011, -0.0000080279799],
-    nd: 1.7859,
-    vd: 43.93,
-    PgF: 0.5619,
-    code6: "786439",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; NBFD11 page (formula 3 polynomial).",
-  },
-  {
-    name: "TAFD25",
-    vendor: "Hoya",
-    polynomial: [3.4905528, -0.016325771, 0.041146322, 0.0027169124, -0.00018260518, 0.000020641862],
-    nd: 1.90366,
-    vd: 31.32,
-    PgF: 0.5939,
-    code6: "904313",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD25 page (formula 3 polynomial).",
-  },
-  {
-    name: "TAF5",
-    vendor: "Hoya",
-    polynomial: [3.2191571, -0.015310562, 0.02636257, 0.0010534452, -0.000061660412, 0.0000039733897],
-    nd: 1.816,
-    vd: 46.57,
-    PgF: 0.5572,
-    code6: "816466",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAF5 page (formula 3 polynomial).",
-  },
-  {
-    name: "E-FD4",
-    vendor: "Hoya",
-    polynomial: [2.9673173, -0.013989565, 0.03345668, 0.0030717804, -0.00026682435, 0.000028950241],
-    nd: 1.7552,
-    vd: 27.53,
-    PgF: 0.6078,
-    code6: "755275",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FD4 page (formula 3 polynomial).",
-  },
-  {
-    name: "E-FD5",
-    vendor: "Hoya",
-    polynomial: [2.7135839, -0.011419719, 0.026133708, 0.0017576479, -0.00013307737, 0.000015203351],
-    nd: 1.6727,
-    vd: 32.17,
-    PgF: 0.5956,
-    code6: "673322",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FD5 page (formula 3 polynomial).",
-  },
-  {
-    name: "E-FL5",
-    vendor: "Hoya",
-    polynomial: [2.4465754, -0.010107041, 0.017759132, 0.00087970259, -0.000057014929, 0.0000060924663],
-    nd: 1.58144,
-    vd: 40.89,
-    PgF: 0.577,
-    code6: "581409",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FL5 page (formula 3 polynomial).",
-  },
-
-  /* Ohara unresolved-token additions */
-  {
-    name: "S-BAL2",
-    vendor: "Ohara",
-    B: [1.30923813, 0.114137353, 1.17882259],
-    C: [0.00838873953, 0.0399436485, 140.257892],
-    nd: 1.570989,
-    vd: 50.799858,
-    PgF: 0.5579,
-    code6: "571508",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BAL2 page.",
-  },
-  {
-    name: "S-LAM55",
-    vendor: "Ohara",
-    B: [1.85412979, 0.165450323, 1.27255422],
-    C: [0.0108438152, 0.051405098, 109.986837],
-    nd: 1.762001,
-    vd: 40.100176,
-    PgF: 0.5763,
-    code6: "762401",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAM55 page.",
-  },
-  {
-    name: "S-TIH10",
-    vendor: "Ohara",
-    B: [1.61549392, 0.262433239, 2.09426189],
-    C: [0.0119830897, 0.059651024, 181.657554],
-    nd: 1.72825,
-    vd: 28.460599,
-    PgF: 0.6082,
-    code6: "728285",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIH10 page.",
-  },
-  {
-    name: "S-BSM28",
-    vendor: "Ohara",
-    B: [1.43822841, 0.128100017, 1.3435553],
-    C: [0.0085977975, 0.0408617854, 143.70989],
-    nd: 1.617722,
-    vd: 49.815052,
-    PgF: 0.5594,
-    code6: "618498",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BSM28 page.",
-  },
-  {
-    name: "S-BSM71",
-    vendor: "Ohara",
-    B: [1.50847885, 0.158099826, 1.36815368],
-    C: [0.00812769076, 0.0354200898, 136.110038],
-    nd: 1.648498,
-    vd: 53.023881,
-    PgF: 0.5536,
-    code6: "649530",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BSM71 page.",
-  },
-  {
-    name: "S-LAL13",
-    vendor: "Ohara",
-    B: [0.980071267, 0.832904776, 1.28111995],
-    C: [0.00389123698, 0.0189164592, 98.9052676],
-    nd: 1.693501,
-    vd: 53.205327,
-    PgF: 0.5462,
-    code6: "694532",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAL13 page.",
-  },
-  {
-    name: "S-LAM3",
-    vendor: "Ohara",
-    B: [1.64258713, 0.23963461, 1.22483026],
-    C: [0.0086824602, 0.0351226242, 116.604369],
-    nd: 1.717004,
-    vd: 47.927969,
-    PgF: 0.5598,
-    code6: "717479",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAM3 page.",
-  },
-  {
-    name: "S-LAM52",
-    vendor: "Ohara",
-    B: [1.73442942, 0.15155391, 1.46225433],
-    C: [0.0100690928, 0.0470634701, 140.084396],
-    nd: 1.72,
-    vd: 43.689812,
-    PgF: 0.5694,
-    code6: "720437",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAM52 page.",
-  },
-
-  /* Hikari six-digit code-family additions */
-  {
-    name: "J-PKH1",
-    vendor: "Hikari",
-    powerSeries: [
-      [2.27892705, 0],
-      [-0.00904327622, 2],
-      [-0.000110679206, 4],
-      [0.0101870033, -2],
-      [0.0000931149884, -4],
-      [0.0000027225654, -6],
-      [-0.0000000182952398, -8],
-    ],
-    nd: 1.5186,
-    vd: 69.89,
-    PgF: 0.5318,
-    code6: "519699",
-    source: "Hikari Optical Glass Catalog 2023, J-PKH1 page (formula 3 power series, d-code 519699).",
-  },
-  {
-    name: "Q-SK52S",
-    vendor: "Hikari",
-    powerSeries: [
-      [2.46682919, 0],
-      [-0.0100692159, 2],
-      [-0.000110027728, 4],
-      [0.0140726622, -2],
-      [0.000136654094, -4],
-      [0.00000857488472, -6],
-      [-0.000000175474987, -8],
-    ],
-    nd: 1.58286,
-    vd: 59.51,
-    PgF: 0.5408,
-    code6: "583595",
-    source: "Hikari Optical Glass Catalog 2023, Q-SK52S page (formula 3 power series, d-code 583595).",
-  },
-  {
-    name: "J-LASKH2",
-    vendor: "Hikari",
-    powerSeries: [
-      [3.01618042, 0],
-      [-0.0139280117, 2],
-      [-0.000204284446, 4],
-      [0.0227027519, -2],
-      [0.000317846393, -4],
-      [0.0000101400049, -6],
-      [0.0000000363521536, -8],
-    ],
-    nd: 1.755,
-    vd: 52.34,
-    PgF: 0.5476,
-    code6: "755523",
-    source: "Hikari Optical Glass Catalog 2023, J-LASKH2 page (formula 3 power series, d-code 755523).",
-  },
-  {
-    name: "J-LASFH22",
-    vendor: "Hikari",
-    powerSeries: [
-      [3.32390176, 0],
-      [-0.0124810607, 2],
-      [-0.000116698235, 4],
-      [0.0316622125, -2],
-      [0.000639837554, -4],
-      [0.000008855505, -6],
-      [0.00000107749628, -8],
-    ],
-    nd: 1.8485,
-    vd: 43.79,
-    PgF: 0.562,
-    code6: "849438",
-    source: "Hikari Optical Glass Catalog 2023, J-LASFH22 page (formula 3 power series, d-code 849438).",
-  },
-  {
-    name: "J-LASF014",
-    vendor: "Hikari",
-    powerSeries: [
-      [3.1211948, 0],
-      [-0.0137527216, 2],
-      [-0.000186592469, 4],
-      [0.0263764849, -2],
-      [0.00045369866, -4],
-      [0.0000109722341, -6],
-      [0.000000611243363, -8],
-    ],
-    nd: 1.788,
-    vd: 47.35,
-    PgF: 0.5559,
-    code6: "788474",
-    source: "Hikari Optical Glass Catalog 2023, J-LASF014 page (formula 3 power series, d-code 788474).",
-  },
-
-  /* ────── Phase 11 code-family additions (May 2026) ──────
-   * Exact six-digit patent-code targets from the remaining unresolved report.
-   * These entries turn code-only labels into catalog-backed Sellmeier surfaces
-   * without requiring speculative vendor-name relabeling in the lens data.
-   */
-  {
-    name: "TAFD65",
-    vendor: "Hoya",
-    polynomial: [4.0113202, -0.017752009, 0.05856933, 0.0043200932, -0.00031070842, 0.000037485638],
-    nd: 2.0509,
-    vd: 26.94,
-    PgF: 0.604,
-    code6: "051269",
-    source:
-      "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD65 page (formula 3 polynomial, d-code 051269 from AGF).",
-  },
-  {
-    name: "F5",
-    vendor: "Schott",
-    B: [1.3104463, 0.19603426, 0.96612977],
-    C: [0.00958633048, 0.0457627627, 115.011883],
-    nd: 1.60342,
-    vd: 38.03,
-    PgF: 0.5795,
-    code6: "603380",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; F5 page.",
-  },
-  {
-    name: "E-FDS2",
-    vendor: "Hoya",
-    polynomial: [3.7672603, -0.017001227, 0.069849634, 0.0063862892, -0.00050812539, 0.000084455529],
-    nd: 2.00272,
-    vd: 19.32,
-    PgF: 0.6428,
-    code6: "003193",
-    source:
-      "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FDS2 page (formula 3 polynomial, d-code 003193 from AGF).",
-  },
-  {
-    name: "M-FCD500",
-    vendor: "Hoya",
-    polynomial: [2.3841554, -0.0078084423, 0.0097130883, 0.00049959889, -0.000047303224, 0.0000023205441],
-    nd: 1.55332,
-    vd: 71.68,
-    PgF: 0.5443,
-    code6: "553717",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-FCD500 page (formula 3 polynomial).",
-  },
-  {
-    name: "PBL25",
-    vendor: "Ohara",
-    polynomial: [2.4465509, -0.0097768464, 0.01753112, 0.00095250889, -0.000054837761, 0.0000047303899],
-    nd: 1.581439,
-    vd: 40.753351,
-    PgF: 0.5748,
-    code6: "581408",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; PBL25 page (formula 3 polynomial).",
-  },
-  {
-    name: "E-F1",
-    vendor: "Hoya",
-    polynomial: [2.5730565, -0.010203605, 0.023095784, 0.00087114252, -0.000030730079, 0.0000069457986],
-    nd: 1.62588,
-    vd: 35.74,
-    PgF: 0.5875,
-    code6: "626357",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-F1 page (formula 3 polynomial).",
-  },
-  {
-    name: "LAC12",
-    vendor: "Hoya",
-    polynomial: [2.7634844, -0.011068339, 0.018246442, 0.00037697356, -0.000017788655, 0.0000015314262],
-    nd: 1.6779,
-    vd: 55.52,
-    PgF: 0.5473,
-    code6: "678555",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; LAC12 page (formula 3 polynomial).",
-  },
-  {
-    name: "S-LAM59",
-    vendor: "Ohara",
-    B: [1.63056133, 0.186994897, 1.30014289],
-    C: [0.00899690705, 0.0368011993, 122.239544],
-    nd: 1.697002,
-    vd: 48.516066,
-    PgF: 0.5581,
-    code6: "697485",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAM59 page.",
-  },
-  {
-    name: "N-KZFS8",
-    vendor: "Schott",
-    B: [1.62693651, 0.24369876, 1.62007141],
-    C: [0.010880863, 0.0494207753, 131.009163],
-    nd: 1.72047,
-    vd: 34.7,
-    PgF: 0.5833,
-    code6: "720347",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-KZFS8 page.",
-  },
-  {
-    name: "S-LAL10",
-    vendor: "Ohara",
-    B: [1.52812575, 0.367965267, 1.11751784],
-    C: [0.00776817644, 0.0272026548, 88.86974],
-    nd: 1.719995,
-    vd: 50.230564,
-    PgF: 0.5512,
-    code6: "720502",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAL10 page.",
-  },
-  {
-    name: "SF10",
-    vendor: "Schott",
-    B: [1.61625977, 0.259229334, 1.07762317],
-    C: [0.0127534559, 0.0581983954, 116.60768],
-    nd: 1.72825,
-    vd: 28.41,
-    PgF: 0.6045,
-    code6: "728284",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; SF10 page.",
-  },
-
-  /* Sumita priority addition */
-  {
-    name: "K-VC89",
-    vendor: "Sumita",
-    polynomial: [3.1860388, -0.013756822, 0.029614017, 0.0012383727, -0.000080134175, 0.0000072330635],
-    nd: 1.81,
-    vd: 41,
-    PgF: 0.5664,
-    code6: "810410",
-    source: "Sumita Zemax catalog 2017-02-02 via refractiveindex.info; K-VC89 page (formula 3 polynomial).",
-  },
-
-  /* ────── Phase 12 high-frequency unresolved-token additions (May 2026) ──────
-   * Direct coefficient-backed names from the unresolved-glass report. Entries
-   * without public dispersion coefficients remain unresolved until a vendor
-   * coefficient source can be checked.
-   */
-  {
-    name: "H-FK61",
-    vendor: "CDGM",
-    B: [0.502504076, 0.715240076, 0.836628986],
-    C: [0.00132743798, 0.0108378966, 155.933755],
-    nd: 1.497,
-    vd: 81.612838,
-    code6: "497816",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-FK61 page.",
-  },
-  {
-    name: "S-TIL2",
-    vendor: "Ohara",
-    B: [1.23401499, 0.0959796833, 1.20503991],
-    C: [0.00869507801, 0.0465611429, 137.953301],
-    nd: 1.54072,
-    vd: 47.227127,
-    code6: "541472",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIL2 page.",
-  },
-  {
-    name: "S-TIL26",
-    vendor: "Ohara",
-    B: [1.31066488, 0.0941903094, 1.23292644],
-    C: [0.00968897812, 0.0527763106, 133.296422],
-    nd: 1.567322,
-    vd: 42.817076,
-    code6: "567428",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIL26 page.",
-  },
-  {
-    name: "N-LAK33B",
-    vendor: "Schott",
-    B: [1.42288601, 0.593661336, 1.1613526],
-    C: [0.00670283452, 0.021941621, 80.7407701],
-    nd: 1.755,
-    vd: 52.3,
-    code6: "755523",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-LAK33B page.",
-  },
-  {
-    name: "TAFD32",
-    vendor: "Hoya",
-    polynomial: [3.396108, -0.013087065, 0.034853516, 0.00079083177, 0.0000076995212, 0.0000023516742],
-    nd: 1.8707,
-    vd: 40.73,
-    code6: "871407",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD32 page (formula 3 polynomial).",
-  },
-
-  /* Phase 14 catalog expansion: CDGM, current Ohara, Hoya, and Schott unresolved-token pass. */
-  {
-    name: "H-QK3L",
-    vendor: "CDGM",
-    B: [0.601737491, 0.858737004, 0.586761501],
-    C: [0.012574974, 92.3325663, 0.002687509],
-    nd: 1.48749,
-    vd: 70.440487,
-    PgF: 0.5288,
-    code6: "487704",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-QK3L page.",
-  },
-  {
-    name: "H-ZF6",
-    vendor: "CDGM",
-    B: [1.64768548, 0.314478021, 2.03864924],
-    C: [0.0118179362, 0.0585408205, 173.964477],
-    nd: 1.7552,
-    vd: 27.530047,
-    PgF: 0.6086,
-    code6: "755275",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZF6 page.",
-  },
-  {
-    name: "H-ZLAF50D",
-    vendor: "CDGM",
-    B: [0.497440659, 1.67780264, 1.32327423],
-    C: [0.0285446966, 0.00788379288, 94.0514342],
-    nd: 1.804,
-    vd: 46.575016,
-    PgF: 0.5565,
-    code6: "804466",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZLAF50D page.",
-  },
-  {
-    name: "H-ZF4A",
-    vendor: "CDGM",
-    B: [1.67238093, 1.34134437, 0.204305471],
-    C: [0.0133653851, 122.464022, 0.0641192325],
-    nd: 1.72825,
-    vd: 28.318797,
-    PgF: 0.6069,
-    code6: "728283",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZF4A page.",
-  },
-  {
-    name: "H-FK71",
-    vendor: "CDGM",
-    B: [0.22226661, 0.880444178, 0.797848915],
-    C: [0.014065122, 0.004204574, 171.029818],
-    nd: 1.4565,
-    vd: 90.273821,
-    PgF: 0.5349,
-    code6: "457903",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-FK71 page.",
-  },
-  {
-    name: "H-ZF13",
-    vendor: "CDGM",
-    B: [1.73521591, 0.316277446, 2.16384634],
-    C: [0.0131087904, 0.0621421663, 178.845558],
-    nd: 1.78472,
-    vd: 25.720045,
-    PgF: 0.6143,
-    code6: "785257",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZF13 page.",
-  },
-  {
-    name: "H-ZF52",
-    vendor: "CDGM",
-    B: [0.409982615, 2.37517176, 1.83913582],
-    C: [0.0621421199, 185.055134, 0.0136093459],
-    nd: 1.84666,
-    vd: 23.784497,
-    PgF: 0.6193,
-    code6: "847238",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZF52 page.",
-  },
-  {
-    name: "H-ZF72A",
-    vendor: "CDGM",
-    B: [0.470478653, 2.55040233, 2.00593441],
-    C: [0.0736024509, 149.306621, 0.0162999226],
-    nd: 1.92286,
-    vd: 18.895887,
-    PgF: 0.6483,
-    code6: "923189",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZF72A page.",
-  },
-  {
-    name: "H-ZF7LA",
-    vendor: "CDGM",
-    B: [1.65846092, 0.298796445, 1.8184834],
-    C: [144.874699, 0.0635424599, 0.0141908346],
-    nd: 1.80518,
-    vd: 25.456256,
-    PgF: 0.6137,
-    code6: "805255",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZF7LA page.",
-  },
-  {
-    name: "H-ZLAF52A",
-    vendor: "CDGM",
-    B: [1.85313688, 0.317938591, 1.16396318],
-    C: [0.00985534639, 0.0392611678, 93.1032763],
-    nd: 1.8061,
-    vd: 41.016887,
-    PgF: 0.5681,
-    code6: "806410",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZLAF52A page.",
-  },
-  {
-    name: "H-ZLAF68C",
-    vendor: "CDGM",
-    B: [1.58833884, 0.851761035, 1.58277416],
-    C: [0.00689365045, 0.0288290435, 125.035659],
-    nd: 1.883,
-    vd: 40.847465,
-    PgF: 0.5667,
-    code6: "883409",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZLAF68C page.",
-  },
-  {
-    name: "H-LAK6A",
-    vendor: "CDGM",
-    B: [1.13018444, 1.60589, 0.207868982],
-    C: [81.4541493, 0.00817222723, 0.030380314],
-    nd: 1.6935,
-    vd: 53.380441,
-    PgF: 0.5458,
-    code6: "694534",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-LAK6A page.",
-  },
-  {
-    name: "H-LAF3B",
-    vendor: "CDGM",
-    B: [1.66486969, 0.30162248, 1.1973888],
-    C: [0.00895646712, 0.0350299695, 123.334438],
-    nd: 1.744,
-    vd: 44.900027,
-    PgF: 0.5642,
-    code6: "744449",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-LAF3B page.",
-  },
-  {
-    name: "S-PHM52Q",
-    vendor: "Ohara",
-    B: [1.2696875, 0.308418136, 1.04078976],
-    C: [0.00621622949, 0.0207785265, 112.05125],
-    nd: 1.617999,
-    vd: 63.321449,
-    code6: "618633",
-    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
-  },
-  {
-    name: "S-FPM4",
-    vendor: "Ohara",
-    B: [0.685585084, 0.623380215, 0.914178386],
-    C: [0.0028817201, 0.0124701707, 153.5772],
-    nd: 1.52841,
-    vd: 76.452824,
-    code6: "528765",
-    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
-  },
-  {
-    name: "S-NBH58",
-    vendor: "Ohara",
-    B: [1.71408219, 0.362966167, 2.56486266],
-    C: [0.0116812775, 0.0540847347, 188.7857],
-    nd: 1.7888,
-    vd: 28.42865,
-    code6: "789284",
-    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
-  },
-  {
-    name: "S-BAH11",
-    vendor: "Ohara",
-    B: [1.5713886, 0.147869313, 1.28092846],
-    C: [0.00910807936, 0.0402401684, 130.399367],
-    nd: 1.666718,
-    vd: 48.325247,
-    code6: "667483",
-    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
-  },
-  {
-    name: "S-BAL41",
-    vendor: "Ohara",
-    B: [1.243442, 0.166301104, 1.10586114],
-    C: [0.0116396708, -0.00890464938, 114.11122],
-    nd: 1.563839,
-    vd: 60.666575,
-    code6: "564607",
-    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
-  },
-  {
-    name: "S-LAH52Q",
-    vendor: "Ohara",
-    B: [1.91082318, 0.239854589, 1.16159733],
-    C: [0.0103565352, 0.0413805081, 96.60373],
-    nd: 1.79952,
-    vd: 42.239506,
-    code6: "800422",
-    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
-  },
-  {
-    name: "S-LAH71",
-    vendor: "Ohara",
-    B: [1.98280031, 0.31675845, 2.44472646],
-    C: [0.0118987456, 0.0527156001, 213.220697],
-    nd: 1.850259,
-    vd: 32.27012,
-    code6: "850323",
-    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
-  },
-  {
-    name: "S-LAL7",
-    vendor: "Ohara",
-    B: [0.916121247, 0.765948319, 1.27745023],
-    C: [0.00395889743, 0.0167547425, 110.762706],
-    nd: 1.651597,
-    vd: 58.550655,
-    code6: "652585",
-    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
-  },
-  {
-    name: "S-LAM61",
-    vendor: "Ohara",
-    B: [1.7388333, 0.15093743, 1.12118445],
-    C: [0.00980244105, 0.0433179685, 101.214625],
-    nd: 1.720002,
-    vd: 46.024571,
-    code6: "720460",
-    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
-  },
-  {
-    name: "S-TIH1",
-    vendor: "Ohara",
-    B: [1.60326759, 0.242980935, 1.81313592],
-    C: [0.0118019139, 0.0591363658, 161.218747],
-    nd: 1.717362,
-    vd: 29.518091,
-    code6: "717295",
-    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
-  },
-  {
-    name: "S-TIL1",
-    vendor: "Ohara",
-    B: [1.25088944, 0.0997973327, 1.20583504],
-    C: [0.00883921279, 0.0482685052, 137.414953],
-    nd: 1.548141,
-    vd: 45.784277,
-    code6: "548458",
-    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
-  },
-  {
-    name: "S-TIM8",
-    vendor: "Ohara",
-    B: [1.37262713, 0.112636276, 1.39786421],
-    C: [0.0103220068, 0.0550195044, 147.735609],
-    nd: 1.595509,
-    vd: 39.242273,
-    code6: "596392",
-    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
-  },
-  {
-    name: "E-F2",
-    vendor: "Hoya",
-    polynomial: [2.5561822, -0.010237038, 0.022470348, 0.00084580679, -0.000033087089, 0.0000071054919],
-    nd: 1.62004,
-    vd: 36.3,
-    code6: "620363",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; formula 3 polynomial.",
-  },
-  {
-    name: "NBFD13",
-    vendor: "Hoya",
-    polynomial: [3.1729223, -0.014248534, 0.029168114, 0.001342876, -0.000096178753, 0.0000084042556],
-    nd: 1.8061,
-    vd: 40.73,
-    code6: "806407",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; formula 3 polynomial.",
-  },
-  {
-    name: "E-F3",
-    vendor: "Hoya",
-    polynomial: [2.5362697, -0.010610424, 0.021233821, 0.00099012105, -0.000059984348, 0.0000082354366],
-    nd: 1.61293,
-    vd: 36.96,
-    code6: "613370",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; formula 3 polynomial.",
-  },
-  {
-    name: "N-BALF5",
-    vendor: "Schott",
-    B: [1.28385965, 0.0719300942, 1.05048927],
-    C: [0.00825815975, 0.0441920027, 107.097324],
-    nd: 1.54739,
-    vd: 53.63,
-    code6: "547536",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info.",
-  },
-  {
-    name: "FK3",
-    vendor: "Schott",
-    B: [0.973346627, 0.146642231, 0.679304225],
-    C: [0.00640795469, 0.020565293, 80.4965389],
-    nd: 1.4645,
-    vd: 65.77,
-    code6: "464658",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info.",
-  },
-
-  /* ────── Phase 17 additions (May 2026) ──────
-   * Six-digit missing-Sellmeier audit over Hasselblad, Laowa, Leica, Minolta,
-   * and Nikon rows. Entries below are coefficient-backed public matches or
-   * documented catalog equivalents that round-trip through the d-line guard.
-   */
-  {
-    name: "BPH50",
-    vendor: "Ohara",
-    polynomial: [2.9271459, -0.011970157, 0.031922538, 0.0015457537, -0.000063507309, 0.0000096835139],
-    nd: 1.739997,
-    vd: 31.710001,
-    code6: "740317",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; BPH50 page (formula 3 polynomial).",
-  },
-  {
-    name: "H-LAK12",
-    vendor: "CDGM",
-    B: [0.786703112, 1.04086924, 1.21770392],
-    C: [0.00202073208, 0.0163643393, 86.703479],
-    nd: 1.6968,
-    vd: 56.175305,
-    code6: "697562",
-    source:
-      "CDGM Zemax catalog 2022-06 via refractiveindex.info; public CDGM cross-reference lists old Ohara LAL64 as the 697565 equivalent.",
-  },
-  {
-    name: "S-LAM58",
-    vendor: "Ohara",
-    B: [1.70984856, 0.173342897, 1.64833565],
-    C: [0.0100852127, 0.0470890831, 157.46852],
-    nd: 1.72,
-    vd: 41.98,
-    PgF: 0.5729,
-    code6: "720420",
-    source: "Ohara GmbH S-LAM58 datasheet, April 2018; Sellmeier constants A1-A3/B1-B3.",
-  },
-  {
-    name: "D-ZLAF81-25",
-    vendor: "CDGM",
-    B: [1.90958888, 0.266893034, 1.26585975],
-    C: [0.0102182926, 0.0421328894, 97.0733228],
-    nd: 1.80757,
-    vd: 40.97334,
-    code6: "808410",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; D-ZLAF81-25 page.",
-  },
-  {
-    name: "K-SKLD200",
-    vendor: "Sumita",
-    polynomial: [2.4786815, -0.010928711, 0.014104052, 0.00015365594, 0.00001194296, -0.00000053372519],
-    nd: 1.5866,
-    vd: 59.0,
-    PgF: 0.541,
-    code6: "587590",
-    source: "Sumita Optical Glass K-SKLD200 datasheet, catalog Ver. 14.01.00; formula-3 coefficients A0-A5.",
-  },
-
-  /* ────── Phase 18 additions (May 2026) ──────
-   * Nikon six-digit missing-Sellmeier audit. Hikari entries use the vendor's
-   * published formula-3 power-series coefficients from the 2023 Nikon/Hikari
-   * catalog; N-SSK5 uses Schott's published Sellmeier constants.
-   */
-  {
-    name: "J-FK5",
-    vendor: "Hikari",
-    powerSeries: [
-      [2.18826855, 0],
-      [-0.00919044724, 2],
-      [-0.000111621071, 4],
-      [0.00926372815, -2],
-      [0.0000734900733, -4],
-      [0.00000419724242, -6],
-      [-0.000000115412203, -8],
-    ],
-    nd: 1.48749,
-    vd: 70.31,
-    PgF: 0.5291,
-    code6: "487703",
-    source: "Hikari Optical Glass Catalog 2023, J-FK5 page (formula 3 power series, d-code 487703).",
-  },
-  {
-    name: "J-FK01A",
-    vendor: "Hikari",
-    powerSeries: [
-      [2.21785004, 0],
-      [-0.00552619544, 2],
-      [-0.0000404219098, 4],
-      [0.00839820345, -2],
-      [0.000088019088, -4],
-      [0.000000115723877, -6],
-      [0.0000000538178618, -8],
-    ],
-    nd: 1.497,
-    vd: 81.65,
-    PgF: 0.5372,
-    code6: "497817",
-    source: "Hikari Optical Glass Catalog 2023, J-FK01A page (formula 3 power series, d-code 497817).",
-  },
-  {
-    name: "J-FKH1",
-    vendor: "Hikari",
-    powerSeries: [
-      [2.22016073, 0],
-      [-0.00500725473, 2],
-      [-0.0000355507111, 4],
-      [0.00842088796, -2],
-      [0.0000702327459, -4],
-      [0.000002470079, -6],
-      [-0.0000000650002003, -8],
-    ],
-    nd: 1.49782,
-    vd: 82.57,
-    PgF: 0.5386,
-    code6: "498826",
-    source: "Hikari Optical Glass Catalog 2023, J-FKH1 page (formula 3 power series, d-code 498826).",
-  },
-  {
-    name: "J-BASF2",
-    vendor: "Hikari",
-    powerSeries: [
-      [2.69527127, 0],
-      [-0.0109541476, 2],
-      [0, 4],
-      [0.0234490053, -2],
-      [0.0019016356, -4],
-      [-0.000368276327, -6],
-      [0.0000749770823, -8],
-      [-0.0000070943249, -10],
-      [0.000000301058679, -12],
-    ],
-    nd: 1.66446,
-    vd: 35.87,
-    PgF: 0.589,
-    code6: "664359",
-    source: "Hikari Optical Glass Catalog 2023, J-BASF2 page (formula 3 power series, d-code 664359).",
-  },
-  {
-    name: "J-BASF6",
-    vendor: "Hikari",
-    powerSeries: [
-      [2.71408053, 0],
-      [-0.011443869, 2],
-      [-0.000185062065, 4],
-      [0.0232439131, -2],
-      [0.000130291556, -4],
-      [0.00011814346, -6],
-      [-0.0000118586652, -8],
-      [0.000000768022789, -10],
-      [0, -12],
-    ],
-    nd: 1.66755,
-    vd: 41.87,
-    PgF: 0.5751,
-    code6: "668419",
-    source: "Hikari Optical Glass Catalog 2023, J-BASF6 page (formula 3 power series, d-code 668419).",
-  },
-  {
-    name: "J-SFH1",
-    vendor: "Hikari",
-    powerSeries: [
-      [3.11637039, 0],
-      [-0.0140103252, 2],
-      [0.0455295459, -2],
-      [0.00379129507, -4],
-      [-0.000576203793, -6],
-      [0.000153793977, -8],
-      [-0.0000165780029, -10],
-      [0.0000008814805, -12],
-    ],
-    nd: 1.80809,
-    vd: 22.74,
-    PgF: 0.6288,
-    code6: "808227",
-    source: "Hikari Optical Glass Catalog 2023, J-SFH1 page (formula 3 power series, d-code 808227).",
-  },
-  {
-    name: "J-SFH5",
-    vendor: "Hikari",
-    powerSeries: [
-      [2.95136928, 0],
-      [-0.012221704, 2],
-      [-0.000229797089, 4],
-      [0.0439763055, -2],
-      [-0.000594321969, -4],
-      [0.000766493369, -6],
-      [-0.000098263351, -8],
-      [0.00000700260276, -10],
-    ],
-    nd: 1.75575,
-    vd: 24.71,
-    PgF: 0.6291,
-    code6: "756247",
-    source: "Hikari Optical Glass Catalog 2023, J-SFH5 page (formula 3 power series, d-code 756247).",
-  },
-  {
-    name: "N-SSK5",
-    vendor: "Schott",
-    B: [1.59222659, 0.103520774, 1.05174016],
-    C: [0.00920284626, 0.0423530072, 106.927374],
-    nd: 1.65844,
-    vd: 50.88,
-    PgF: 0.5575,
-    code6: "658509",
-    source: "Schott N-SSK5 datasheet / optical glass collection, Sellmeier constants B1-B3/C1-C3, code 658509.371.",
-  },
-  {
-    name: "J-KZFH6",
-    vendor: "Hikari",
-    powerSeries: [
-      [2.76011543, 0],
-      [-0.0123775063, 2],
-      [0.0244484473, -2],
-      [0.00115656997, -4],
-      [-0.0000974000173, -6],
-      [0.0000185226731, -8],
-      [-0.00000141521524, -10],
-      [0.0000000592951085, -12],
-    ],
-    nd: 1.68376,
-    vd: 37.64,
-    PgF: 0.5782,
-    code6: "684376",
-    source: "Hikari Optical Glass Catalog 2023, J-KZFH6 page (formula 3 power series, d-code 684376).",
-  },
-  {
-    name: "J-LAF01",
-    vendor: "Hikari",
-    powerSeries: [
-      [2.82684251, 0],
-      [-0.0103793494, 2],
-      [-0.0000759583803, 4],
-      [0.021662046, -2],
-      [0.000472548975, -4],
-      [-0.00000258637501, -6],
-      [0.00000143317138, -8],
-    ],
-    nd: 1.7,
-    vd: 48.11,
-    PgF: 0.5604,
-    code6: "700481",
-    source: "Hikari Optical Glass Catalog 2023, J-LAF01 page (formula 3 power series, d-code 700481).",
-  },
-  {
-    name: "Q-LASFPH2S",
-    vendor: "Hikari",
-    powerSeries: [
-      [3.0430437, 0],
-      [-0.0130169861, 2],
-      [-0.000175012562, 4],
-      [0.025603733, -2],
-      [0.000448879136, -4],
-      [0.00001131389, -6],
-      [0.000000469755261, -8],
-    ],
-    nd: 1.765437,
-    vd: 46.75,
-    PgF: 0.5565,
-    code6: "765468",
-    source: "Nikon Q-series optical glass catalog, Q-LASFPH2S page (formula 3 power series, d-code 765468).",
-  },
-  {
-    name: "J-LAF016",
-    vendor: "Hikari",
-    powerSeries: [
-      [3.13977744, 0],
-      [-0.0136971027, 2],
-      [-0.000162026634, 4],
-      [0.0348771551, -2],
-      [0.000587991861, -4],
-      [0.000135087453, -6],
-      [-0.0000123042975, -8],
-      [0.000000955959268, -10],
-    ],
-    nd: 1.801,
-    vd: 34.92,
-    PgF: 0.5853,
-    code6: "801349",
-    source: "Hikari Optical Glass Catalog 2023, J-LAF016 page (formula 3 power series, d-code 801349).",
-  },
-  {
-    name: "J-LASF017",
-    vendor: "Hikari",
-    powerSeries: [
-      [3.14264424, 0],
-      [-0.0144687256, 2],
-      [-0.000159589924, 4],
-      [0.0273342436, -2],
-      [0.00060806842, -4],
-      [0.00000018005447, -6],
-      [0.00000158176253, -8],
-    ],
-    nd: 1.795,
-    vd: 45.31,
-    PgF: 0.5598,
-    code6: "795453",
-    source: "Hikari Optical Glass Catalog 2023, J-LASF017 page (formula 3 power series, d-code 795453).",
-  },
-  {
-    name: "Q-LASFPH3S",
-    vendor: "Hikari",
-    powerSeries: [
-      [3.14195165, 0],
-      [-0.0129889243, 2],
-      [-0.000159736617, 4],
-      [0.0279148154, -2],
-      [0.000509887221, -4],
-      [0.0000132658761, -6],
-      [0.00000051483904, -8],
-    ],
-    nd: 1.795256,
-    vd: 45.25,
-    PgF: 0.5589,
-    code6: "795453",
-    source: "Hikari Optical Glass Catalog 2023, Q-LASFPH3S page (formula 3 power series, d-code 795453).",
-  },
-  {
-    name: "J-LASF021",
-    vendor: "Hikari",
-    powerSeries: [
-      [3.30477446, 0],
-      [-0.0145925795, 2],
-      [0.0359916632, -2],
-      [0.00349591839, -4],
-      [-0.000743189237, -6],
-      [0.000155265866, -8],
-      [-0.0000152961037, -10],
-      [0.000000659677491, -12],
-    ],
-    nd: 1.85026,
-    vd: 32.35,
-    PgF: 0.5947,
-    code6: "850324",
-    source: "Hikari Optical Glass Catalog 2023, J-LASF021 page (formula 3 power series, d-code 850324).",
-  },
-  {
-    name: "J-LASFH23",
-    vendor: "Hikari",
-    powerSeries: [
-      [3.27710657, 0],
-      [-0.0120480934, 2],
-      [-0.000105066933, 4],
-      [0.0470173769, -2],
-      [0.00095791266, -4],
-      [0.000282409939, -6],
-      [-0.0000289931135, -8],
-      [0.00000251862401, -10],
-    ],
-    nd: 1.85,
-    vd: 27.03,
-    PgF: 0.6093,
-    code6: "850270",
-    source: "Hikari Optical Glass Catalog 2023, J-LASFH23 page (formula 3 power series, d-code 850270).",
-  },
-
-  /* ────── Phase 19 Sigma patent-audit additions (May 2026) ──────
-   * Coefficient-backed HOYA rows needed by the Sigma 105/1.4, 20/1.4,
-   * and 14-24/2.8 patent audit. All are Zemax formula-3 polynomial rows
-   * from HOYA's 2017 public AGF as mirrored by refractiveindex.info.
-   */
-  {
-    name: "PCD40",
-    vendor: "Hoya",
-    polynomial: [2.5844463, -0.0090521858, 0.014301987, 0.0001705077, 0.000004964536, -0.00000012927479],
-    nd: 1.61997,
-    vd: 63.88,
-    PgF: 0.5457,
-    code6: "620639",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; PCD40 page (formula 3 polynomial).",
-  },
-  {
-    name: "M-LAC130",
-    vendor: "Hoya",
-    polynomial: [2.813412, -0.01332448, 0.01927865, 0.0004124452, -0.000009940432, 0.0000009622024],
-    nd: 1.6935,
-    vd: 53.2,
-    PgF: 0.5484,
-    code6: "694532",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-LAC130 page (formula 3 polynomial).",
-  },
-  {
-    name: "M-BACD5N",
-    vendor: "Hoya",
-    polynomial: [2.48951, -0.01224496, 0.01294262, 0.0003490449, -0.00001777228, 0.0000009113719],
-    nd: 1.58913,
-    vd: 61.25,
-    PgF: 0.5401,
-    code6: "589613",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-BACD5N page (formula 3 polynomial).",
-  },
-  {
-    name: "M-TAFD305",
-    vendor: "Hoya",
-    polynomial: [3.327184, -0.01343242, 0.033398, 0.001006047, -0.00002136908, 0.000003972561],
-    nd: 1.85135,
-    vd: 40.1,
-    PgF: 0.5697,
-    code6: "851401",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-TAFD305 page (formula 3 polynomial).",
-  },
-  {
-    name: "BAFD7",
-    vendor: "Hoya",
-    polynomial: [2.8217211, -0.0097265483, 0.024345979, 0.00078406305, -0.000022431497, 0.0000046726029],
-    nd: 1.70154,
-    vd: 41.15,
-    PgF: 0.5774,
-    code6: "702412",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; BAFD7 page (formula 3 polynomial).",
-  },
-
-  /* ────── Phase 20 six-digit Sweep 2 additions (June 2026) ──────
-   * Exact public coefficient-backed matches for six-digit code rows in the
-   * missing-Sellmeier report. Entries below are vendor Zemax formulas mirrored
-   * by refractiveindex.info and selected only when the six-digit code and
-   * stored patent nd/vd values stayed inside the runtime nd guard.
-   */
-  {
-    name: "SF9",
-    vendor: "Hikari",
-    polynomial: [2.661342, -0.01187903, 0.02328061, 0.001777729, -0.0001298549, 0.00001138748],
-    nd: 1.65446,
-    vd: 33.855439,
-    code6: "654339",
-    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; SF9 page (formula 3 polynomial).",
-  },
-  {
-    name: "S-LAL58",
-    vendor: "Ohara",
-    B: [1.06368789, 0.744939067, 1.59178942],
-    C: [0.018519964, 0.00116295862, 156.636025],
-    nd: 1.693495,
-    vd: 50.810652,
-    code6: "694508",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAL58 page.",
-  },
-  {
-    name: "MC-NBFD135",
-    vendor: "Hoya",
-    polynomial: [3.1819698, -0.014926522, 0.028987649, 0.0012223654, -0.000052916688, 0.0000047807227],
-    nd: 1.80834,
-    vd: 40.92,
-    code6: "808409",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; MC-NBFD135 page (formula 3 polynomial).",
-  },
-  {
-    name: "J-BK7",
-    vendor: "Hikari",
-    powerSeries: [
-      [2.27109726, 0],
-      [-0.00947304881, 2],
-      [-0.000089187152, 4],
-      [0.0109352525, -2],
-      [0.000136527555, -4],
-      [0.00000168617824, -6],
-      [0.0000000585391298, -8],
-    ],
-    nd: 1.5168,
-    vd: 63.880667,
-    code6: "517639",
-    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; J-BK7 page (formula 3 power series).",
-  },
-  {
-    name: "BK7G18",
-    vendor: "Schott",
-    B: [1.26538542, 0.0144191073, 1.00323028],
-    C: [0.00813104078, 0.0543303226, 102.821166],
-    nd: 1.51975,
-    vd: 63.58,
-    code6: "520636",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; BK7G18 page.",
-  },
-  {
-    name: "P-SK57Q1",
-    vendor: "Schott",
-    B: [1.30536483, 0.171434328, 1.10117219],
-    C: [0.00736408831, 0.0255786047, 106.72606],
-    nd: 1.586,
-    vd: 59.5,
-    code6: "586595",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; P-SK57Q1 page.",
-  },
-  {
-    name: "M-PCD51",
-    vendor: "Hoya",
-    polynomial: [2.502026, -0.01108292, 0.0112772, 0.0005715408, -0.00005587455, 0.000002905292],
-    nd: 1.59201,
-    vd: 67.02,
-    code6: "592670",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-PCD51 page (formula 3 polynomial).",
-  },
-  {
-    name: "SK9",
-    vendor: "Hikari",
-    polynomial: [2.564248, -0.01337038, 0.01398205, 0.0007030773, -0.00004206054, 0.000002098576],
-    nd: 1.61405,
-    vd: 55.0612,
-    code6: "614551",
-    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; SK9 page (formula 3 polynomial).",
-  },
-  {
-    name: "E-BAF8",
-    vendor: "Hikari",
-    polynomial: [2.58072694, -0.00826682499, 0.0192773695, 0.000287391647, 0.0000154399962, 0.000000463747358],
-    nd: 1.62374,
-    vd: 47.044346,
-    code6: "624470",
-    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; E-BAF8 page (formula 3 polynomial).",
-  },
-  {
-    name: "M-TAC80",
-    vendor: "Hoya",
-    polynomial: [2.9351948, -0.016193458, 0.018392287, 0.0010086167, -0.000087565474, 0.0000046921333],
-    nd: 1.72903,
-    vd: 54.04,
-    code6: "729540",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-TAC80 page (formula 3 polynomial).",
-  },
-  {
-    name: "M-LAF81",
-    vendor: "Hoya",
-    polynomial: [2.9167886, -0.012398199, 0.026675237, 0.00065190504, 0.0000046956175, 0.000002691054],
-    nd: 1.73077,
-    vd: 40.5,
-    code6: "731405",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-LAF81 page (formula 3 polynomial).",
-  },
-  {
-    name: "TAC2",
-    vendor: "Hoya",
-    polynomial: [2.9717137, -0.014952593, 0.020162868, 0.00094072283, -0.000088614104, 0.0000053191242],
-    nd: 1.741,
-    vd: 52.6,
-    code6: "741526",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAC2 page (formula 3 polynomial).",
-  },
-  {
-    name: "LAK011",
-    vendor: "Hikari",
-    polynomial: [2.971993, -0.01524352, 0.02020612, 0.0008752522, -0.00007671807, 0.000004848072],
-    nd: 1.741,
-    vd: 52.672067,
-    code6: "741527",
-    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; LAK011 page (formula 3 polynomial).",
-  },
-  {
-    name: "E-LAF11",
-    vendor: "Hikari",
-    polynomial: [2.98504551, -0.012861773, 0.0312839577, 0.00217089493, -0.000162580251, 0.0000182003929],
-    nd: 1.75692,
-    vd: 31.591329,
-    code6: "757316",
-    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; E-LAF11 page (formula 3 polynomial).",
-  },
-  {
-    name: "E-LASF016",
-    vendor: "Hikari",
-    polynomial: [3.07241378, -0.0146849804, 0.0244285857, 0.000398183255, 0.0000122802438, 0.000000123709929],
-    nd: 1.772499,
-    vd: 49.60759,
-    code6: "772496",
-    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; E-LASF016 page (formula 3 polynomial).",
-  },
-  {
-    name: "M-TAF401",
-    vendor: "Hoya",
-    polynomial: [3.07279, -0.01366385, 0.02527346, 0.000612182, -0.00001147045, 0.000001543],
-    nd: 1.77377,
-    vd: 47.17,
-    code6: "774472",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-TAF401 page (formula 3 polynomial).",
-  },
-  {
-    name: "M-TAF31",
-    vendor: "Hoya",
-    polynomial: [3.164358, -0.01403321, 0.02767034, 0.0006403025, -0.000006349234, 0.000001615744],
-    nd: 1.80139,
-    vd: 45.45,
-    code6: "801455",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-TAF31 page (formula 3 polynomial).",
-  },
-  {
-    name: "E-LAFH2",
-    vendor: "Hikari",
-    polynomial: [3.1493362, -0.0148382096, 0.0335324542, 0.00159644183, -0.0000706244061, 0.0000115814159],
-    nd: 1.80384,
-    vd: 33.889217,
-    code6: "804339",
-    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; E-LAFH2 page (formula 3 polynomial).",
-  },
-  {
-    name: "M-NBFD10",
-    vendor: "Hoya",
-    polynomial: [3.264623, -0.01581977, 0.03225894, 0.001709312, -0.0001081957, 0.00001058323],
-    nd: 1.83441,
-    vd: 37.28,
-    code6: "834373",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-NBFD10 page (formula 3 polynomial).",
-  },
-  {
-    name: "FDS16-W",
-    vendor: "Hoya",
-    polynomial: [3.6906098, -0.027849024, 0.063139051, 0.012443698, -0.0014774085, 0.00017478701],
-    nd: 1.98612,
-    vd: 16.48,
-    code6: "986165",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FDS16-W page (formula 3 polynomial).",
-  },
-
-  /* ────── Phase 21 chromatic-coverage additions (June 2026) ──────
-   * Source-backed entries from the generated named-token opportunity queue.
-   * Each row round-trips through assertCatalogConsistent at the existing 1e-4
-   * tolerance. Broad E-LAF7 and S-NPH7 aliases remain intentionally omitted
-   * because current annotations span incompatible nd/vd regions.
-   */
-  {
-    name: "S-LAL12",
-    vendor: "Ohara",
-    B: [0.992053895, 0.771377731, 1.18296264],
-    C: [0.0167095063, 0.00236750156, 105.90108],
-    nd: 1.6779,
-    vd: 55.34,
-    PgF: 0.5472,
-    code6: "678553",
-    source: "Ohara S-LAL12 datasheet, OHARA 25-04 official optical-glass PDF.",
-  },
-  {
-    name: "S-BSM10",
-    vendor: "Ohara",
-    B: [0.945443081, 0.643237376, 1.17752968],
-    C: [0.0157263798, 0.00161924066, 121.361748],
-    nd: 1.6228,
-    vd: 57.05,
-    PgF: 0.5464,
-    code6: "623570",
-    source: "Ohara S-BSM10 datasheet, OHARA 25-04 official optical-glass PDF.",
-  },
-  {
-    name: "S-LAM7",
-    vendor: "Ohara",
-    B: [1.71014712, 0.256943292, 1.63986271],
-    C: [0.010516108, 0.0502809636, 146.181217],
-    nd: 1.7495,
-    vd: 35.28,
-    PgF: 0.5869,
-    code6: "750353",
-    source: "Ohara S-LAM7 datasheet, OHARA 25-04 official optical-glass PDF.",
-  },
-  {
-    name: "L-LAM69",
-    vendor: "Ohara",
-    B: [1.7403896, 0.176996917, 1.76775413],
-    C: [0.010339887, 0.0484822765, 136.671996],
-    nd: 1.73077,
-    vd: 40.51,
-    PgF: 0.5728,
-    code6: "731405",
-    source: "Ohara L-LAM69 datasheet, OHARA 25-04 official optical-glass PDF.",
-  },
-  {
-    name: "N-SF8",
-    vendor: "Schott",
-    B: [1.55075812, 0.209816918, 1.46205491],
-    C: [0.011433834, 0.0582725652, 133.24165],
-    nd: 1.68894,
-    vd: 31.31,
-    PgF: 0.5999,
-    code6: "689313",
-    source: "SCHOTT N-SF8 optical-glass datasheet, official SCHOTT Advanced Optics PDF.",
-  },
-  {
-    name: "H-LAF4",
-    vendor: "CDGM",
-    B: [0.210645313, 1.75547352, 1.28979005],
-    C: [0.0542759294, 0.0112370777, 119.852069],
-    nd: 1.7495,
-    vd: 34.989266,
-    PgF: 0.588,
-    code6: "750350",
-    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-LAF4 page.",
-  },
-
-  /* Phase 23 named-token additions (June 2026)
-   * Direct records from Hoya/Ohara Zemax data mirrored by refractiveindex.info.
-   * BSC3 is resolved through the legacy Hoya E-C3 optical constants; incompatible
-   * 583/594 annotations must use their own matching catalog row instead.
-   */
-  {
-    name: "E-FD13",
-    vendor: "Hoya",
-    polynomial: [2.9205906, -0.013681618, 0.03256776, 0.0028725519, -0.00024277107, 0.000026614514],
-    nd: 1.74077,
-    vd: 27.76,
-    PgF: 0.6077,
-    code6: "741278",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FD13 page (formula 3 polynomial).",
-  },
-  {
-    name: "E-FD10",
-    vendor: "Hoya",
-    polynomial: [2.881518, -0.013228312, 0.03145559, 0.0026851666, -0.00022577544, 0.000024693268],
-    nd: 1.72825,
-    vd: 28.32,
-    PgF: 0.6059,
-    code6: "728283",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FD10 page (formula 3 polynomial).",
-  },
-  {
-    name: "BACD4",
-    vendor: "Hoya",
-    polynomial: [2.5584621, -0.0097457283, 0.015110495, 0.00023273333, 0.0000011137249, 0.00000016424767],
-    nd: 1.61272,
-    vd: 58.58,
-    PgF: 0.5448,
-    code6: "613586",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; BACD4 page (formula 3 polynomial).",
-  },
-  {
-    name: "E-C3",
-    vendor: "Hoya",
-    polynomial: [2.2725452, -0.008410491, 0.011306843, 0.00037782297, -0.000026903423, 0.0000015976489],
-    nd: 1.51823,
-    vd: 58.96,
-    PgF: 0.5442,
-    code6: "518590",
-    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-C3 page (formula 3 polynomial).",
-  },
-  {
-    name: "S-BSM15",
-    vendor: "Ohara",
-    B: [0.953128328, 0.637613977, 1.65245647],
-    C: [0.00387638985, 0.0185094632, 159.442367],
-    nd: 1.622992,
-    vd: 58.165763,
-    PgF: 0.5459,
-    code6: "623582",
-    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BSM15 page.",
-  },
-
-  /* Phase 24 named-token additions (June 2026)
-   * Repeated Schott-name gaps from the coverage queue with source-backed
-   * formula-2 coefficients. E-line patent annotations that cite these families
-   * but preserve n_e values remain explicitly unmatched in lens data so the
-   * d-line Sellmeier catalog is not applied to a different reference line.
-   */
-  {
-    name: "SF5",
-    vendor: "Schott",
-    B: [1.46141885, 0.247713019, 0.949995832],
-    C: [0.0111826126, 0.0508594669, 112.041888],
-    nd: 1.6727,
-    vd: 32.21,
-    PgF: 0.5901,
-    code6: "673322",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; SF5 page.",
-  },
-  {
-    name: "N-SF5",
-    vendor: "Schott",
-    B: [1.52481889, 0.187085527, 1.42729015],
-    C: [0.011254756, 0.0588995392, 129.141675],
-    nd: 1.67271,
-    vd: 32.25,
-    PgF: 0.5969,
-    code6: "673323",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-SF5 page.",
-  },
-  {
-    name: "N-LASF44",
-    vendor: "Schott",
-    B: [1.78897105, 0.38675867, 1.30506243],
-    C: [0.00872506277, 0.0308085023, 92.7743824],
-    nd: 1.8042,
-    vd: 46.5,
-    PgF: 0.5573,
-    code6: "804465",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-LASF44 page.",
-  },
-  {
-    name: "N-LAK9",
-    vendor: "Schott",
-    B: [1.46231905, 0.344399589, 1.15508372],
-    C: [0.00724270156, 0.0243353131, 85.4686868],
-    nd: 1.691,
-    vd: 54.71,
-    PgF: 0.5368,
-    code6: "691547",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-LAK9 page.",
-  },
-  {
-    name: "N-PSK53A",
-    vendor: "Schott",
-    B: [1.38121836, 0.196745645, 0.886089205],
-    C: [0.00706416337, 0.0233251345, 97.4847345],
-    nd: 1.618,
-    vd: 63.39,
-    PgF: 0.5366,
-    code6: "618634",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-PSK53A page.",
-  },
-  {
-    name: "N-LAF2",
-    vendor: "Schott",
-    B: [1.80984227, 0.15729555, 1.0930037],
-    C: [0.0101711622, 0.0442431765, 100.687748],
-    nd: 1.74397,
-    vd: 44.85,
-    PgF: 0.5588,
-    code6: "744449",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-LAF2 page.",
-  },
-  {
-    name: "N-BAK4",
-    vendor: "Schott",
-    B: [1.28834642, 0.132817724, 0.945395373],
-    C: [0.00779980626, 0.0315631177, 105.965875],
-    nd: 1.56883,
-    vd: 55.98,
-    PgF: 0.5431,
-    code6: "569560",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-BAK4 page.",
-  },
-  {
-    name: "N-LAK7",
-    vendor: "Schott",
-    B: [1.23679889, 0.445051837, 1.01745888],
-    C: [0.00610105538, 0.0201388334, 90.638038],
-    nd: 1.6516,
-    vd: 58.52,
-    PgF: 0.5417,
-    code6: "652585",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-LAK7 page.",
-  },
-  {
-    name: "N-BAF4",
-    vendor: "Schott",
-    B: [1.42056328, 0.102721269, 1.14380976],
-    C: [0.00942015382, 0.0531087291, 110.278856],
-    nd: 1.60568,
-    vd: 43.72,
-    PgF: 0.5584,
-    code6: "606437",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-BAF4 page.",
-  },
-  {
-    name: "N-SSK2",
-    vendor: "Schott",
-    B: [1.4306027, 0.153150554, 1.01390904],
-    C: [0.00823982975, 0.0333736841, 106.870822],
-    nd: 1.62229,
-    vd: 53.27,
-    PgF: 0.5542,
-    code6: "622533",
-    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-SSK2 page.",
-  },
-];
+export const RAW_CATALOG: readonly GlassEntry[] = GLASS_CATALOG_SOURCE_ORDER.map(entryByName);
