+++ src/optics/glassCatalogEntries/cdgm.ts
@@ -1,0 +1,221 @@
+/**
+ * Vendor-published CDGM optical glass dispersion entries.
+ */
+
+import type { GlassEntry } from "../glassCatalogTypes.js";
+
+export const CDGM_GLASS_ENTRIES: readonly GlassEntry[] = [
+  {
+    name: "H-ZF88",
+    vendor: "CDGM",
+    B: [2.04275349, 1.98276153, 0.504267829],
+    C: [0.0169742976, 109.068151, 0.075171071],
+    nd: 1.94595,
+    vd: 17.941964,
+    PgF: 0.6537,
+    code6: "946179",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZF88 page.",
+  },
+  {
+    name: "H-ZLaF76",
+    vendor: "CDGM",
+    B: [0.279518569, 2.01429741, 1.5507336],
+    C: [0.0593856526, 0.012625744, 100.032291],
+    nd: 1.85013,
+    vd: 30.06,
+    PgF: 0.6001,
+    code6: "850301",
+    source: "CDGM H-ZLaF76 public datasheet, 2026 access; Sellmeier constants K1-L3, d-code 850301.",
+  },
+  /* ────── Phase 12 high-frequency unresolved-token additions (May 2026) ──────
+   * Direct coefficient-backed names from the unresolved-glass report. Entries
+   * without public dispersion coefficients remain unresolved until a vendor
+   * coefficient source can be checked.
+   */
+  {
+    name: "H-FK61",
+    vendor: "CDGM",
+    B: [0.502504076, 0.715240076, 0.836628986],
+    C: [0.00132743798, 0.0108378966, 155.933755],
+    nd: 1.497,
+    vd: 81.612838,
+    code6: "497816",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-FK61 page.",
+  },
+  /* Phase 14 catalog expansion: CDGM, current Ohara, Hoya, and Schott unresolved-token pass. */
+  {
+    name: "H-QK3L",
+    vendor: "CDGM",
+    B: [0.601737491, 0.858737004, 0.586761501],
+    C: [0.012574974, 92.3325663, 0.002687509],
+    nd: 1.48749,
+    vd: 70.440487,
+    PgF: 0.5288,
+    code6: "487704",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-QK3L page.",
+  },
+  {
+    name: "H-ZF6",
+    vendor: "CDGM",
+    B: [1.64768548, 0.314478021, 2.03864924],
+    C: [0.0118179362, 0.0585408205, 173.964477],
+    nd: 1.7552,
+    vd: 27.530047,
+    PgF: 0.6086,
+    code6: "755275",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZF6 page.",
+  },
+  {
+    name: "H-ZLAF50D",
+    vendor: "CDGM",
+    B: [0.497440659, 1.67780264, 1.32327423],
+    C: [0.0285446966, 0.00788379288, 94.0514342],
+    nd: 1.804,
+    vd: 46.575016,
+    PgF: 0.5565,
+    code6: "804466",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZLAF50D page.",
+  },
+  {
+    name: "H-ZF4A",
+    vendor: "CDGM",
+    B: [1.67238093, 1.34134437, 0.204305471],
+    C: [0.0133653851, 122.464022, 0.0641192325],
+    nd: 1.72825,
+    vd: 28.318797,
+    PgF: 0.6069,
+    code6: "728283",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZF4A page.",
+  },
+  {
+    name: "H-FK71",
+    vendor: "CDGM",
+    B: [0.22226661, 0.880444178, 0.797848915],
+    C: [0.014065122, 0.004204574, 171.029818],
+    nd: 1.4565,
+    vd: 90.273821,
+    PgF: 0.5349,
+    code6: "457903",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-FK71 page.",
+  },
+  {
+    name: "H-ZF13",
+    vendor: "CDGM",
+    B: [1.73521591, 0.316277446, 2.16384634],
+    C: [0.0131087904, 0.0621421663, 178.845558],
+    nd: 1.78472,
+    vd: 25.720045,
+    PgF: 0.6143,
+    code6: "785257",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZF13 page.",
+  },
+  {
+    name: "H-ZF52",
+    vendor: "CDGM",
+    B: [0.409982615, 2.37517176, 1.83913582],
+    C: [0.0621421199, 185.055134, 0.0136093459],
+    nd: 1.84666,
+    vd: 23.784497,
+    PgF: 0.6193,
+    code6: "847238",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZF52 page.",
+  },
+  {
+    name: "H-ZF72A",
+    vendor: "CDGM",
+    B: [0.470478653, 2.55040233, 2.00593441],
+    C: [0.0736024509, 149.306621, 0.0162999226],
+    nd: 1.92286,
+    vd: 18.895887,
+    PgF: 0.6483,
+    code6: "923189",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZF72A page.",
+  },
+  {
+    name: "H-ZF7LA",
+    vendor: "CDGM",
+    B: [1.65846092, 0.298796445, 1.8184834],
+    C: [144.874699, 0.0635424599, 0.0141908346],
+    nd: 1.80518,
+    vd: 25.456256,
+    PgF: 0.6137,
+    code6: "805255",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZF7LA page.",
+  },
+  {
+    name: "H-ZLAF52A",
+    vendor: "CDGM",
+    B: [1.85313688, 0.317938591, 1.16396318],
+    C: [0.00985534639, 0.0392611678, 93.1032763],
+    nd: 1.8061,
+    vd: 41.016887,
+    PgF: 0.5681,
+    code6: "806410",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZLAF52A page.",
+  },
+  {
+    name: "H-ZLAF68C",
+    vendor: "CDGM",
+    B: [1.58833884, 0.851761035, 1.58277416],
+    C: [0.00689365045, 0.0288290435, 125.035659],
+    nd: 1.883,
+    vd: 40.847465,
+    PgF: 0.5667,
+    code6: "883409",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-ZLAF68C page.",
+  },
+  {
+    name: "H-LAK6A",
+    vendor: "CDGM",
+    B: [1.13018444, 1.60589, 0.207868982],
+    C: [81.4541493, 0.00817222723, 0.030380314],
+    nd: 1.6935,
+    vd: 53.380441,
+    PgF: 0.5458,
+    code6: "694534",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-LAK6A page.",
+  },
+  {
+    name: "H-LAF3B",
+    vendor: "CDGM",
+    B: [1.66486969, 0.30162248, 1.1973888],
+    C: [0.00895646712, 0.0350299695, 123.334438],
+    nd: 1.744,
+    vd: 44.900027,
+    PgF: 0.5642,
+    code6: "744449",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-LAF3B page.",
+  },
+  {
+    name: "H-LAK12",
+    vendor: "CDGM",
+    B: [0.786703112, 1.04086924, 1.21770392],
+    C: [0.00202073208, 0.0163643393, 86.703479],
+    nd: 1.6968,
+    vd: 56.175305,
+    code6: "697562",
+    source:
+      "CDGM Zemax catalog 2022-06 via refractiveindex.info; public CDGM cross-reference lists old Ohara LAL64 as the 697565 equivalent.",
+  },
+  {
+    name: "D-ZLAF81-25",
+    vendor: "CDGM",
+    B: [1.90958888, 0.266893034, 1.26585975],
+    C: [0.0102182926, 0.0421328894, 97.0733228],
+    nd: 1.80757,
+    vd: 40.97334,
+    code6: "808410",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; D-ZLAF81-25 page.",
+  },
+  {
+    name: "H-LAF4",
+    vendor: "CDGM",
+    B: [0.210645313, 1.75547352, 1.28979005],
+    C: [0.0542759294, 0.0112370777, 119.852069],
+    nd: 1.7495,
+    vd: 34.989266,
+    PgF: 0.588,
+    code6: "750350",
+    source: "CDGM Zemax catalog 2022-06 via refractiveindex.info; H-LAF4 page.",
+  },
+];
