+++ src/optics/glassCatalogEntries/hikari.ts
@@ -1,0 +1,596 @@
+/**
+ * Vendor-published Hikari optical glass dispersion entries.
+ */
+
+import type { GlassEntry } from "../glassCatalogTypes.js";
+
+export const HIKARI_GLASS_ENTRIES: readonly GlassEntry[] = [
+  {
+    name: "J-PSKH1",
+    vendor: "Hikari",
+    powerSeries: [
+      [2.50208083, 0],
+      [-0.00672143907, 2],
+      [-0.0000534313751, 4],
+      [0.01282644, -2],
+      [0.000156205388, -4],
+      [0.00000121593549, -6],
+      [0.0000000959550869, -8],
+    ],
+    nd: 1.59319,
+    vd: 67.900555,
+    PgF: 0.544,
+    code6: "593679",
+    source:
+      "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; Hikari Optical Glass Catalog 2023, J-PSKH1 page (formula 3 power series, d-code 593679).",
+  },
+  {
+    name: "J-PSKH4",
+    vendor: "Hikari",
+    powerSeries: [
+      [2.50453078, 0],
+      [-0.0101597822, 2],
+      [-0.000108653142, 4],
+      [0.0127723327, -2],
+      [0.000133860625, -4],
+      [0.00000337285381, -6],
+      [-0.0000000256491019, -8],
+    ],
+    nd: 1.59349,
+    vd: 67.001281,
+    PgF: 0.5358,
+    code6: "593670",
+    source:
+      "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; Hikari Optical Glass Catalog 2023, J-PSKH4 page (formula 3 power series, d-code 593670).",
+  },
+  {
+    name: "J-KZFH9",
+    vendor: "Hikari",
+    powerSeries: [
+      [2.92190512, 0],
+      [-0.0131913454, 2],
+      [-0.0000794286252, 4],
+      [0.0327997049, -2],
+      [0.000700950165, -4],
+      [0.00012416909, -6],
+      [-0.0000112359582, -8],
+      [0.000000911052912, -10],
+    ],
+    nd: 1.738,
+    vd: 32.26,
+    PgF: 0.5896,
+    code6: "738323",
+    source: "Hikari Optical Glass Catalog 2023, J-KZFH9 page (formula 3 power series, d-code 738323).",
+  },
+  {
+    name: "BAF3",
+    vendor: "Hikari",
+    polynomial: [2.438682, 0.001609961, 0.02688829, -0.002446579, 0.0003965661, -0.00001973046],
+    nd: 1.58267,
+    vd: 46.476929,
+    PgF: 0.5673,
+    code6: "583465",
+    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; BAF3 page (formula 3 polynomial).",
+  },
+  {
+    name: "J-LAK10",
+    vendor: "Hikari",
+    powerSeries: [
+      [2.89571408, 0],
+      [-0.0120013315, 2],
+      [-0.000136916169, 4],
+      [0.0219522159, -2],
+      [0.000357973143, -4],
+      [0.00000826304425, -6],
+      [0.00000027388172, -8],
+    ],
+    nd: 1.71999,
+    vd: 50.27,
+    PgF: 0.5527,
+    code6: "720503",
+    source:
+      "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; Hikari Optical Glass Catalog 2023, J-LAK10 page (formula 3 power series, d-code 720503).",
+  },
+  {
+    name: "J-LASFH15",
+    vendor: "Hikari",
+    powerSeries: [
+      [3.64640666, 0],
+      [-0.0151039558, 2],
+      [0.0480157444, -2],
+      [0.00349072452, -4],
+      [-0.000573639028, -6],
+      [0.000130249514, -8],
+      [-0.0000129576789, -10],
+      [0.000000592144355, -12],
+    ],
+    nd: 1.95,
+    vd: 29.37,
+    PgF: 0.6001,
+    code6: "950294",
+    source:
+      "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; Hikari Optical Glass Catalog 2023, J-LASFH15 page (formula 3 power series, d-code 950294).",
+  },
+  {
+    name: "J-LASFH9",
+    vendor: "Hikari",
+    powerSeries: [
+      [3.49709032, 0],
+      [-0.0138248635, 2],
+      [-0.000132935974, 4],
+      [0.0409278939, -2],
+      [0.000880758278, -4],
+      [0.0000893285504, -6],
+      [-0.0000066176262, -8],
+      [0.000000624399823, -10],
+    ],
+    nd: 1.90265,
+    vd: 35.725382,
+    PgF: 0.5805,
+    code6: "903357",
+    source:
+      "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; Hikari Optical Glass Catalog 2012, J-LASFH9 page (formula 3 power series, d-code 903357).",
+  },
+  /* Hikari six-digit code-family additions */
+  {
+    name: "J-PKH1",
+    vendor: "Hikari",
+    powerSeries: [
+      [2.27892705, 0],
+      [-0.00904327622, 2],
+      [-0.000110679206, 4],
+      [0.0101870033, -2],
+      [0.0000931149884, -4],
+      [0.0000027225654, -6],
+      [-0.0000000182952398, -8],
+    ],
+    nd: 1.5186,
+    vd: 69.89,
+    PgF: 0.5318,
+    code6: "519699",
+    source: "Hikari Optical Glass Catalog 2023, J-PKH1 page (formula 3 power series, d-code 519699).",
+  },
+  {
+    name: "Q-SK52S",
+    vendor: "Hikari",
+    powerSeries: [
+      [2.46682919, 0],
+      [-0.0100692159, 2],
+      [-0.000110027728, 4],
+      [0.0140726622, -2],
+      [0.000136654094, -4],
+      [0.00000857488472, -6],
+      [-0.000000175474987, -8],
+    ],
+    nd: 1.58286,
+    vd: 59.51,
+    PgF: 0.5408,
+    code6: "583595",
+    source: "Hikari Optical Glass Catalog 2023, Q-SK52S page (formula 3 power series, d-code 583595).",
+  },
+  {
+    name: "J-LASKH2",
+    vendor: "Hikari",
+    powerSeries: [
+      [3.01618042, 0],
+      [-0.0139280117, 2],
+      [-0.000204284446, 4],
+      [0.0227027519, -2],
+      [0.000317846393, -4],
+      [0.0000101400049, -6],
+      [0.0000000363521536, -8],
+    ],
+    nd: 1.755,
+    vd: 52.34,
+    PgF: 0.5476,
+    code6: "755523",
+    source: "Hikari Optical Glass Catalog 2023, J-LASKH2 page (formula 3 power series, d-code 755523).",
+  },
+  {
+    name: "J-LASFH22",
+    vendor: "Hikari",
+    powerSeries: [
+      [3.32390176, 0],
+      [-0.0124810607, 2],
+      [-0.000116698235, 4],
+      [0.0316622125, -2],
+      [0.000639837554, -4],
+      [0.000008855505, -6],
+      [0.00000107749628, -8],
+    ],
+    nd: 1.8485,
+    vd: 43.79,
+    PgF: 0.562,
+    code6: "849438",
+    source: "Hikari Optical Glass Catalog 2023, J-LASFH22 page (formula 3 power series, d-code 849438).",
+  },
+  {
+    name: "J-LASF014",
+    vendor: "Hikari",
+    powerSeries: [
+      [3.1211948, 0],
+      [-0.0137527216, 2],
+      [-0.000186592469, 4],
+      [0.0263764849, -2],
+      [0.00045369866, -4],
+      [0.0000109722341, -6],
+      [0.000000611243363, -8],
+    ],
+    nd: 1.788,
+    vd: 47.35,
+    PgF: 0.5559,
+    code6: "788474",
+    source: "Hikari Optical Glass Catalog 2023, J-LASF014 page (formula 3 power series, d-code 788474).",
+  },
+  /* ────── Phase 18 additions (May 2026) ──────
+   * Nikon six-digit missing-Sellmeier audit. Hikari entries use the vendor's
+   * published formula-3 power-series coefficients from the 2023 Nikon/Hikari
+   * catalog; N-SSK5 uses Schott's published Sellmeier constants.
+   */
+  {
+    name: "J-FK5",
+    vendor: "Hikari",
+    powerSeries: [
+      [2.18826855, 0],
+      [-0.00919044724, 2],
+      [-0.000111621071, 4],
+      [0.00926372815, -2],
+      [0.0000734900733, -4],
+      [0.00000419724242, -6],
+      [-0.000000115412203, -8],
+    ],
+    nd: 1.48749,
+    vd: 70.31,
+    PgF: 0.5291,
+    code6: "487703",
+    source: "Hikari Optical Glass Catalog 2023, J-FK5 page (formula 3 power series, d-code 487703).",
+  },
+  {
+    name: "J-FK01A",
+    vendor: "Hikari",
+    powerSeries: [
+      [2.21785004, 0],
+      [-0.00552619544, 2],
+      [-0.0000404219098, 4],
+      [0.00839820345, -2],
+      [0.000088019088, -4],
+      [0.000000115723877, -6],
+      [0.0000000538178618, -8],
+    ],
+    nd: 1.497,
+    vd: 81.65,
+    PgF: 0.5372,
+    code6: "497817",
+    source: "Hikari Optical Glass Catalog 2023, J-FK01A page (formula 3 power series, d-code 497817).",
+  },
+  {
+    name: "J-FKH1",
+    vendor: "Hikari",
+    powerSeries: [
+      [2.22016073, 0],
+      [-0.00500725473, 2],
+      [-0.0000355507111, 4],
+      [0.00842088796, -2],
+      [0.0000702327459, -4],
+      [0.000002470079, -6],
+      [-0.0000000650002003, -8],
+    ],
+    nd: 1.49782,
+    vd: 82.57,
+    PgF: 0.5386,
+    code6: "498826",
+    source: "Hikari Optical Glass Catalog 2023, J-FKH1 page (formula 3 power series, d-code 498826).",
+  },
+  {
+    name: "J-BASF2",
+    vendor: "Hikari",
+    powerSeries: [
+      [2.69527127, 0],
+      [-0.0109541476, 2],
+      [0, 4],
+      [0.0234490053, -2],
+      [0.0019016356, -4],
+      [-0.000368276327, -6],
+      [0.0000749770823, -8],
+      [-0.0000070943249, -10],
+      [0.000000301058679, -12],
+    ],
+    nd: 1.66446,
+    vd: 35.87,
+    PgF: 0.589,
+    code6: "664359",
+    source: "Hikari Optical Glass Catalog 2023, J-BASF2 page (formula 3 power series, d-code 664359).",
+  },
+  {
+    name: "J-BASF6",
+    vendor: "Hikari",
+    powerSeries: [
+      [2.71408053, 0],
+      [-0.011443869, 2],
+      [-0.000185062065, 4],
+      [0.0232439131, -2],
+      [0.000130291556, -4],
+      [0.00011814346, -6],
+      [-0.0000118586652, -8],
+      [0.000000768022789, -10],
+      [0, -12],
+    ],
+    nd: 1.66755,
+    vd: 41.87,
+    PgF: 0.5751,
+    code6: "668419",
+    source: "Hikari Optical Glass Catalog 2023, J-BASF6 page (formula 3 power series, d-code 668419).",
+  },
+  {
+    name: "J-SFH1",
+    vendor: "Hikari",
+    powerSeries: [
+      [3.11637039, 0],
+      [-0.0140103252, 2],
+      [0.0455295459, -2],
+      [0.00379129507, -4],
+      [-0.000576203793, -6],
+      [0.000153793977, -8],
+      [-0.0000165780029, -10],
+      [0.0000008814805, -12],
+    ],
+    nd: 1.80809,
+    vd: 22.74,
+    PgF: 0.6288,
+    code6: "808227",
+    source: "Hikari Optical Glass Catalog 2023, J-SFH1 page (formula 3 power series, d-code 808227).",
+  },
+  {
+    name: "J-SFH5",
+    vendor: "Hikari",
+    powerSeries: [
+      [2.95136928, 0],
+      [-0.012221704, 2],
+      [-0.000229797089, 4],
+      [0.0439763055, -2],
+      [-0.000594321969, -4],
+      [0.000766493369, -6],
+      [-0.000098263351, -8],
+      [0.00000700260276, -10],
+    ],
+    nd: 1.75575,
+    vd: 24.71,
+    PgF: 0.6291,
+    code6: "756247",
+    source: "Hikari Optical Glass Catalog 2023, J-SFH5 page (formula 3 power series, d-code 756247).",
+  },
+  {
+    name: "J-KZFH6",
+    vendor: "Hikari",
+    powerSeries: [
+      [2.76011543, 0],
+      [-0.0123775063, 2],
+      [0.0244484473, -2],
+      [0.00115656997, -4],
+      [-0.0000974000173, -6],
+      [0.0000185226731, -8],
+      [-0.00000141521524, -10],
+      [0.0000000592951085, -12],
+    ],
+    nd: 1.68376,
+    vd: 37.64,
+    PgF: 0.5782,
+    code6: "684376",
+    source: "Hikari Optical Glass Catalog 2023, J-KZFH6 page (formula 3 power series, d-code 684376).",
+  },
+  {
+    name: "J-LAF01",
+    vendor: "Hikari",
+    powerSeries: [
+      [2.82684251, 0],
+      [-0.0103793494, 2],
+      [-0.0000759583803, 4],
+      [0.021662046, -2],
+      [0.000472548975, -4],
+      [-0.00000258637501, -6],
+      [0.00000143317138, -8],
+    ],
+    nd: 1.7,
+    vd: 48.11,
+    PgF: 0.5604,
+    code6: "700481",
+    source: "Hikari Optical Glass Catalog 2023, J-LAF01 page (formula 3 power series, d-code 700481).",
+  },
+  {
+    name: "Q-LASFPH2S",
+    vendor: "Hikari",
+    powerSeries: [
+      [3.0430437, 0],
+      [-0.0130169861, 2],
+      [-0.000175012562, 4],
+      [0.025603733, -2],
+      [0.000448879136, -4],
+      [0.00001131389, -6],
+      [0.000000469755261, -8],
+    ],
+    nd: 1.765437,
+    vd: 46.75,
+    PgF: 0.5565,
+    code6: "765468",
+    source: "Nikon Q-series optical glass catalog, Q-LASFPH2S page (formula 3 power series, d-code 765468).",
+  },
+  {
+    name: "J-LAF016",
+    vendor: "Hikari",
+    powerSeries: [
+      [3.13977744, 0],
+      [-0.0136971027, 2],
+      [-0.000162026634, 4],
+      [0.0348771551, -2],
+      [0.000587991861, -4],
+      [0.000135087453, -6],
+      [-0.0000123042975, -8],
+      [0.000000955959268, -10],
+    ],
+    nd: 1.801,
+    vd: 34.92,
+    PgF: 0.5853,
+    code6: "801349",
+    source: "Hikari Optical Glass Catalog 2023, J-LAF016 page (formula 3 power series, d-code 801349).",
+  },
+  {
+    name: "J-LASF017",
+    vendor: "Hikari",
+    powerSeries: [
+      [3.14264424, 0],
+      [-0.0144687256, 2],
+      [-0.000159589924, 4],
+      [0.0273342436, -2],
+      [0.00060806842, -4],
+      [0.00000018005447, -6],
+      [0.00000158176253, -8],
+    ],
+    nd: 1.795,
+    vd: 45.31,
+    PgF: 0.5598,
+    code6: "795453",
+    source: "Hikari Optical Glass Catalog 2023, J-LASF017 page (formula 3 power series, d-code 795453).",
+  },
+  {
+    name: "Q-LASFPH3S",
+    vendor: "Hikari",
+    powerSeries: [
+      [3.14195165, 0],
+      [-0.0129889243, 2],
+      [-0.000159736617, 4],
+      [0.0279148154, -2],
+      [0.000509887221, -4],
+      [0.0000132658761, -6],
+      [0.00000051483904, -8],
+    ],
+    nd: 1.795256,
+    vd: 45.25,
+    PgF: 0.5589,
+    code6: "795453",
+    source: "Hikari Optical Glass Catalog 2023, Q-LASFPH3S page (formula 3 power series, d-code 795453).",
+  },
+  {
+    name: "J-LASF021",
+    vendor: "Hikari",
+    powerSeries: [
+      [3.30477446, 0],
+      [-0.0145925795, 2],
+      [0.0359916632, -2],
+      [0.00349591839, -4],
+      [-0.000743189237, -6],
+      [0.000155265866, -8],
+      [-0.0000152961037, -10],
+      [0.000000659677491, -12],
+    ],
+    nd: 1.85026,
+    vd: 32.35,
+    PgF: 0.5947,
+    code6: "850324",
+    source: "Hikari Optical Glass Catalog 2023, J-LASF021 page (formula 3 power series, d-code 850324).",
+  },
+  {
+    name: "J-LASFH23",
+    vendor: "Hikari",
+    powerSeries: [
+      [3.27710657, 0],
+      [-0.0120480934, 2],
+      [-0.000105066933, 4],
+      [0.0470173769, -2],
+      [0.00095791266, -4],
+      [0.000282409939, -6],
+      [-0.0000289931135, -8],
+      [0.00000251862401, -10],
+    ],
+    nd: 1.85,
+    vd: 27.03,
+    PgF: 0.6093,
+    code6: "850270",
+    source: "Hikari Optical Glass Catalog 2023, J-LASFH23 page (formula 3 power series, d-code 850270).",
+  },
+  /* ────── Phase 20 six-digit Sweep 2 additions (June 2026) ──────
+   * Exact public coefficient-backed matches for six-digit code rows in the
+   * missing-Sellmeier report. Entries below are vendor Zemax formulas mirrored
+   * by refractiveindex.info and selected only when the six-digit code and
+   * stored patent nd/vd values stayed inside the runtime nd guard.
+   */
+  {
+    name: "SF9",
+    vendor: "Hikari",
+    polynomial: [2.661342, -0.01187903, 0.02328061, 0.001777729, -0.0001298549, 0.00001138748],
+    nd: 1.65446,
+    vd: 33.855439,
+    code6: "654339",
+    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; SF9 page (formula 3 polynomial).",
+  },
+  {
+    name: "J-BK7",
+    vendor: "Hikari",
+    powerSeries: [
+      [2.27109726, 0],
+      [-0.00947304881, 2],
+      [-0.000089187152, 4],
+      [0.0109352525, -2],
+      [0.000136527555, -4],
+      [0.00000168617824, -6],
+      [0.0000000585391298, -8],
+    ],
+    nd: 1.5168,
+    vd: 63.880667,
+    code6: "517639",
+    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; J-BK7 page (formula 3 power series).",
+  },
+  {
+    name: "SK9",
+    vendor: "Hikari",
+    polynomial: [2.564248, -0.01337038, 0.01398205, 0.0007030773, -0.00004206054, 0.000002098576],
+    nd: 1.61405,
+    vd: 55.0612,
+    code6: "614551",
+    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; SK9 page (formula 3 polynomial).",
+  },
+  {
+    name: "E-BAF8",
+    vendor: "Hikari",
+    polynomial: [2.58072694, -0.00826682499, 0.0192773695, 0.000287391647, 0.0000154399962, 0.000000463747358],
+    nd: 1.62374,
+    vd: 47.044346,
+    code6: "624470",
+    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; E-BAF8 page (formula 3 polynomial).",
+  },
+  {
+    name: "LAK011",
+    vendor: "Hikari",
+    polynomial: [2.971993, -0.01524352, 0.02020612, 0.0008752522, -0.00007671807, 0.000004848072],
+    nd: 1.741,
+    vd: 52.672067,
+    code6: "741527",
+    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; LAK011 page (formula 3 polynomial).",
+  },
+  {
+    name: "E-LAF11",
+    vendor: "Hikari",
+    polynomial: [2.98504551, -0.012861773, 0.0312839577, 0.00217089493, -0.000162580251, 0.0000182003929],
+    nd: 1.75692,
+    vd: 31.591329,
+    code6: "757316",
+    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; E-LAF11 page (formula 3 polynomial).",
+  },
+  {
+    name: "E-LASF016",
+    vendor: "Hikari",
+    polynomial: [3.07241378, -0.0146849804, 0.0244285857, 0.000398183255, 0.0000122802438, 0.000000123709929],
+    nd: 1.772499,
+    vd: 49.60759,
+    code6: "772496",
+    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; E-LASF016 page (formula 3 polynomial).",
+  },
+  {
+    name: "E-LAFH2",
+    vendor: "Hikari",
+    polynomial: [3.1493362, -0.0148382096, 0.0335324542, 0.00159644183, -0.0000706244061, 0.0000115814159],
+    nd: 1.80384,
+    vd: 33.889217,
+    code6: "804339",
+    source: "Nikon/Hikari Zemax catalog 2017-11 via refractiveindex.info; E-LAFH2 page (formula 3 polynomial).",
+  },
+];
