+++ src/optics/glassCatalogEntries/hoya.ts
@@ -1,0 +1,665 @@
+/**
+ * Vendor-published Hoya optical glass dispersion entries.
+ */
+
+import type { GlassEntry } from "../glassCatalogTypes.js";
+
+export const HOYA_GLASS_ENTRIES: readonly GlassEntry[] = [
+  /* ────── Hoya — ED crowns and dense lanthanum flints ──────
+   * Hoya AGF entries use the polynomial Schott formula (form 1), not
+   * Sellmeier-1. The B/C below are a least-squares Sellmeier-1 fit to
+   * the polynomial sampled densely (80 points across 365–1014 nm). Max
+   * absolute index error across the spectrum is below 1.4e-5.
+   */
+  {
+    name: "FCD1",
+    vendor: "Hoya",
+    B: [0.708299951, 0.5095943307, 0.5656963028],
+    C: [0.002809999603, 0.01255522353, 101.5038807],
+    nd: 1.497,
+    vd: 81.61,
+    PgF: 0.544,
+    code6: "497816",
+    source:
+      "Hoya AGF (hoya_2017-04-01.agf) form-1 polynomial; Sellmeier-1 fit, max abs index error 1.4e-5 across 365–1014 nm. ED, S-FPL51-equivalent.",
+  },
+  {
+    name: "FCD505",
+    vendor: "Hoya",
+    B: [1.034776218, 0.4665786919, 0.6985808477],
+    C: [0.004826471098, 0.01646358686, 105.4578263],
+    nd: 1.59282,
+    vd: 68.63,
+    PgF: 0.5478,
+    code6: "593686",
+    source:
+      "Hoya AGF (hoya_2017-04-01.agf) form-1 polynomial; Sellmeier-1 fit, max abs index error 4.3e-6 across 365–1014 nm.",
+  },
+  {
+    name: "TAFD30",
+    vendor: "Hoya",
+    B: [1.660453863, 0.7793549404, 1.03002609],
+    C: [0.007957073706, 0.02889879311, 79.74654506],
+    nd: 1.883,
+    vd: 40.8,
+    PgF: 0.5658,
+    code6: "883408",
+    source:
+      "Hoya AGF (hoya_2017-04-01.agf) form-1 polynomial; Sellmeier-1 fit, max abs index error 7.1e-6 across 365–1014 nm.",
+  },
+  {
+    name: "TAFD37A",
+    vendor: "Hoya",
+    B: [2.188879041, 0.3049704899, 1.232253827],
+    C: [0.01151042519, 0.04599567769, 93.14898114],
+    nd: 1.90043,
+    vd: 37.37,
+    PgF: 0.5767,
+    code6: "900374",
+    source:
+      "Hoya Zemax catalog (hoya_2017-04-01, formula 3) via refractiveindex.info; Sellmeier-1 fit, max abs index error 5.1e-6 across 365–1014 nm. Ultra-high-index lanthanum; used in Voigtländer Nokton X 50mm f/1.2.",
+  },
+  {
+    name: "TAFD37",
+    vendor: "Hoya",
+    B: [2.259360214, 0.2343652859, 1.481042614],
+    C: [0.01221915793, 0.04989586099, 111.9367914],
+    nd: 1.90043,
+    vd: 37.37,
+    PgF: 0.5772,
+    source:
+      "Hoya Zemax catalog (hoya_2017-04-01, formula 3) via refractiveindex.info; Sellmeier-1 fit, max abs index error 3.7e-6 across 365–1014 nm. Distinct Sellmeier from TAFD37A.",
+  },
+  {
+    name: "TAFD33",
+    vendor: "Hoya",
+    B: [2.212737897, 0.2183360248, 1.27909893],
+    C: [0.01179423107, 0.04560326145, 100.6882193],
+    nd: 1.881,
+    vd: 40.14,
+    PgF: 0.5704,
+    code6: "881401",
+    source:
+      "Hoya Zemax catalog (hoya_2017-04-01, formula 3) via refractiveindex.info; Sellmeier-1 fit, max abs index error 1.0e-6 across 365–1014 nm. Dense lanthanum flint used in Ricoh GR III 28mm.",
+  },
+  /* Hoya priority additions */
+  {
+    name: "E-FDS1",
+    vendor: "Hoya",
+    polynomial: [3.5044639, -0.020333833, 0.0538027, 0.0062874192, -0.00056922176, 0.000073559387],
+    nd: 1.92286,
+    vd: 20.88,
+    PgF: 0.6369,
+    code6: "923209",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FDS1 page (formula 3 polynomial).",
+  },
+  {
+    name: "FCD100",
+    vendor: "Hoya",
+    polynomial: [2.0482157, -0.0043211222, 0.0061826755, 0.00003141148, 0.0000035370793, -0.00000016694497],
+    nd: 1.437,
+    vd: 95.1,
+    PgF: 0.54,
+    code6: "437951",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FCD100 page (formula 3 polynomial).",
+  },
+  {
+    name: "FCD515",
+    vendor: "Hoya",
+    polynomial: [2.502355, -0.0072080042, 0.012012743, 0.00034920842, -0.000024729676, 0.0000013801567],
+    nd: 1.59282,
+    vd: 68.63,
+    PgF: 0.5478,
+    code6: "593686",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FCD515 page (formula 3 polynomial).",
+  },
+  {
+    name: "FCD705",
+    vendor: "Hoya",
+    polynomial: [2.3742832, -0.0062014418, 0.010582967, 0.000063100703, 0.000008196715, -0.00000034424446],
+    nd: 1.55032,
+    vd: 75.5,
+    PgF: 0.5445,
+    code6: "550755",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FCD705 page (formula 3 polynomial).",
+  },
+  {
+    name: "TAFD45",
+    vendor: "Hoya",
+    polynomial: [3.672962, -0.0151805, 0.0457187, 0.002186221, -0.0000993742, 0.00001490466],
+    nd: 1.95375,
+    vd: 32.32,
+    code6: "954323",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD45 page (formula 3 polynomial).",
+  },
+  {
+    name: "TAFD55",
+    vendor: "Hoya",
+    polynomial: [3.8317578, -0.015266397, 0.054573843, 0.0023109346, -0.000044476201, 0.000016143887],
+    nd: 2.001,
+    vd: 29.13,
+    PgF: 0.5986,
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD55 page (formula 3 polynomial).",
+  },
+  {
+    name: "E-CF6",
+    vendor: "Hoya",
+    polynomial: [2.2668912, -0.0093866212, 0.012174195, 0.0005076903, -0.000033573414, 0.0000029059013],
+    nd: 1.51742,
+    vd: 52.15,
+    PgF: 0.5561,
+    code6: "517522",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-CF6 page (formula 3 polynomial).",
+  },
+  {
+    name: "MC-TAF101-100",
+    vendor: "Hoya",
+    polynomial: [3.0591566, -0.013604717, 0.024624272, 0.00041834242, 0.0000025361908, 0.0000007998082],
+    nd: 1.76902,
+    vd: 49.29,
+    PgF: 0.5627,
+    code6: "769493",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; MC-TAF101-100 page (formula 3 polynomial).",
+  },
+  {
+    name: "NBFD15",
+    vendor: "Hoya",
+    polynomial: [3.1540011, -0.013256753, 0.034103312, 0.0018472167, -0.00010807332, 0.000012975901],
+    nd: 1.8061,
+    vd: 33.27,
+    code6: "806333",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; NBFD15 page (formula 3 polynomial).",
+  },
+  {
+    name: "NBFD25",
+    vendor: "Hoya",
+    polynomial: [3.2879389, -0.015856356, 0.045638807, 0.0033027661, -0.00021690686, 0.000029625863],
+    nd: 1.85451,
+    vd: 25.15,
+    PgF: 0.6103,
+    code6: "855252",
+    source: "HOYA Optical Glass Catalogue NBFD25 PDF, HOYA20250701M.xlsm; formula-3 A0-A5 constants, d-code 855-252.",
+  },
+  {
+    name: "NBFD29",
+    vendor: "Hoya",
+    polynomial: [3.023484, -0.014446333, 0.034673303, 0.0020637437, -0.00011126222, 0.000014468973],
+    nd: 1.77047,
+    vd: 29.74,
+    PgF: 0.5951,
+    code6: "770297",
+    source: "HOYA Optical Glass Catalogue NBFD29 PDF, HOYA20250701M.xlsm; formula-3 A0-A5 constants, d-code 770-297.",
+  },
+  {
+    name: "FDS24",
+    vendor: "Hoya",
+    polynomial: [3.5126327, -0.0153015, 0.053936061, 0.0034949369, -0.00017950295, 0.000034403181],
+    nd: 1.92119,
+    vd: 23.96,
+    PgF: 0.6465,
+    code6: "921240",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FDS24 page (formula 3 polynomial).",
+  },
+  {
+    name: "BACD15",
+    vendor: "Hoya",
+    polynomial: [2.5864832, -0.0080357829, 0.018080573, -0.00045974172, 0.000088484321, -0.0000037544765],
+    nd: 1.62299,
+    vd: 58.12,
+    PgF: 0.5439,
+    code6: "623581",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; BACD15 page (formula 3 polynomial).",
+  },
+  {
+    name: "TAFD40",
+    vendor: "Hoya",
+    polynomial: [3.8106232, -0.016351565, 0.058801097, 0.0034919026, -0.0001694811, 0.000032569333],
+    nd: 2.00069,
+    vd: 25.46,
+    PgF: 0.6121,
+    code6: "001255",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD40 page (formula 3 polynomial).",
+  },
+  /* Hoya unresolved-token additions */
+  {
+    name: "FC5",
+    vendor: "Hoya",
+    polynomial: [2.1894054, -0.0099044908, 0.008640337, 0.00022263067, -0.000012291942, 0.00000059386349],
+    nd: 1.48749,
+    vd: 70.45,
+    PgF: 0.5345,
+    code6: "487704",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FC5 page (formula 3 polynomial).",
+  },
+  {
+    name: "E-FD15",
+    vendor: "Hoya",
+    polynomial: [2.7919943, -0.012311206, 0.028802741, 0.0021430059, -0.00016932119, 0.000019741766],
+    nd: 1.69895,
+    vd: 30.05,
+    PgF: 0.6019,
+    code6: "699301",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FD15 page (formula 3 polynomial).",
+  },
+  {
+    name: "TAFD5F",
+    vendor: "Hoya",
+    polynomial: [3.275546, -0.014470354, 0.030922062, 0.00073289981, 0.0000046573307, 0.0000018242746],
+    nd: 1.83481,
+    vd: 42.72,
+    PgF: 0.5657,
+    code6: "835427",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD5F page (formula 3 polynomial).",
+  },
+  /* Hoya Phase 9 additions */
+  {
+    name: "E-FD2",
+    vendor: "Hoya",
+    polynomial: [2.6399644, -0.011645718, 0.023061733, 0.0017348219, -0.00013986059, 0.00001391667],
+    nd: 1.64769,
+    vd: 33.84,
+    PgF: 0.5918,
+    code6: "648338",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FD2 page (formula 3 polynomial).",
+  },
+  {
+    name: "E-FD8",
+    vendor: "Hoya",
+    polynomial: [2.7617586, -0.011254039, 0.028221737, 0.0017540315, -0.00012127053, 0.000015522384],
+    nd: 1.68893,
+    vd: 31.16,
+    PgF: 0.5981,
+    code6: "689312",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FD8 page (formula 3 polynomial).",
+  },
+  {
+    name: "FF8",
+    vendor: "Hoya",
+    polynomial: [2.9564361, -0.019762195, 0.029878823, 0.0051830852, -0.00054202389, 0.000048463245],
+    nd: 1.75211,
+    vd: 25.05,
+    code6: "752251",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FF8 page (formula 3 polynomial).",
+  },
+  {
+    name: "TAFD35",
+    vendor: "Hoya",
+    polynomial: [3.5274747, -0.014839948, 0.039745402, 0.001803472, -0.000089279334, 0.000011377783],
+    nd: 1.91082,
+    vd: 35.25,
+    PgF: 0.5819,
+    code6: "911353",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD35 page (formula 3 polynomial).",
+  },
+  {
+    name: "LAC13",
+    vendor: "Hoya",
+    polynomial: [2.8125316, -0.012915271, 0.019960339, 0.00019877888, 0.000017647061, -0.00000044767632],
+    nd: 1.6935,
+    vd: 53.34,
+    PgF: 0.5475,
+    code6: "694533",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; LAC13 page (formula 3 polynomial).",
+  },
+  /* ────── Phase 10 research additions (May 2026) ──────
+   * Batch sourced from the post-Phase-9 unresolved and relabel reports. Hoya
+   * formula-3 entries and Hikari power-series entries keep the vendor
+   * published form; Ohara entries use standard Sellmeier-1 coefficients.
+   */
+
+  /* Hoya formula-3 additions */
+  {
+    name: "FDS18",
+    vendor: "Hoya",
+    polynomial: [3.563784, -0.025449644, 0.058401659, 0.0092310983, -0.00095499727, 0.00011777456],
+    nd: 1.94595,
+    vd: 17.98,
+    PgF: 0.6522,
+    code6: "946180",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FDS18 page (formula 3 polynomial).",
+  },
+  {
+    name: "NBFD3",
+    vendor: "Hoya",
+    polynomial: [3.1633746, -0.01327628, 0.031133309, 0.0008352338, 0.0000029980499, 0.0000025568853],
+    nd: 1.8045,
+    vd: 39.63,
+    PgF: 0.5715,
+    code6: "805396",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; NBFD3 page (formula 3 polynomial).",
+  },
+  {
+    name: "NBFD11",
+    vendor: "Hoya",
+    polynomial: [3.1010267, -0.010022108, 0.032631687, -0.00081029268, 0.0001935011, -0.0000080279799],
+    nd: 1.7859,
+    vd: 43.93,
+    PgF: 0.5619,
+    code6: "786439",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; NBFD11 page (formula 3 polynomial).",
+  },
+  {
+    name: "TAFD25",
+    vendor: "Hoya",
+    polynomial: [3.4905528, -0.016325771, 0.041146322, 0.0027169124, -0.00018260518, 0.000020641862],
+    nd: 1.90366,
+    vd: 31.32,
+    PgF: 0.5939,
+    code6: "904313",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD25 page (formula 3 polynomial).",
+  },
+  {
+    name: "TAF5",
+    vendor: "Hoya",
+    polynomial: [3.2191571, -0.015310562, 0.02636257, 0.0010534452, -0.000061660412, 0.0000039733897],
+    nd: 1.816,
+    vd: 46.57,
+    PgF: 0.5572,
+    code6: "816466",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAF5 page (formula 3 polynomial).",
+  },
+  {
+    name: "E-FD4",
+    vendor: "Hoya",
+    polynomial: [2.9673173, -0.013989565, 0.03345668, 0.0030717804, -0.00026682435, 0.000028950241],
+    nd: 1.7552,
+    vd: 27.53,
+    PgF: 0.6078,
+    code6: "755275",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FD4 page (formula 3 polynomial).",
+  },
+  {
+    name: "E-FD5",
+    vendor: "Hoya",
+    polynomial: [2.7135839, -0.011419719, 0.026133708, 0.0017576479, -0.00013307737, 0.000015203351],
+    nd: 1.6727,
+    vd: 32.17,
+    PgF: 0.5956,
+    code6: "673322",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FD5 page (formula 3 polynomial).",
+  },
+  {
+    name: "E-FL5",
+    vendor: "Hoya",
+    polynomial: [2.4465754, -0.010107041, 0.017759132, 0.00087970259, -0.000057014929, 0.0000060924663],
+    nd: 1.58144,
+    vd: 40.89,
+    PgF: 0.577,
+    code6: "581409",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FL5 page (formula 3 polynomial).",
+  },
+  /* ────── Phase 11 code-family additions (May 2026) ──────
+   * Exact six-digit patent-code targets from the remaining unresolved report.
+   * These entries turn code-only labels into catalog-backed Sellmeier surfaces
+   * without requiring speculative vendor-name relabeling in the lens data.
+   */
+  {
+    name: "TAFD65",
+    vendor: "Hoya",
+    polynomial: [4.0113202, -0.017752009, 0.05856933, 0.0043200932, -0.00031070842, 0.000037485638],
+    nd: 2.0509,
+    vd: 26.94,
+    PgF: 0.604,
+    code6: "051269",
+    source:
+      "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD65 page (formula 3 polynomial, d-code 051269 from AGF).",
+  },
+  {
+    name: "E-FDS2",
+    vendor: "Hoya",
+    polynomial: [3.7672603, -0.017001227, 0.069849634, 0.0063862892, -0.00050812539, 0.000084455529],
+    nd: 2.00272,
+    vd: 19.32,
+    PgF: 0.6428,
+    code6: "003193",
+    source:
+      "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FDS2 page (formula 3 polynomial, d-code 003193 from AGF).",
+  },
+  {
+    name: "M-FCD500",
+    vendor: "Hoya",
+    polynomial: [2.3841554, -0.0078084423, 0.0097130883, 0.00049959889, -0.000047303224, 0.0000023205441],
+    nd: 1.55332,
+    vd: 71.68,
+    PgF: 0.5443,
+    code6: "553717",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-FCD500 page (formula 3 polynomial).",
+  },
+  {
+    name: "E-F1",
+    vendor: "Hoya",
+    polynomial: [2.5730565, -0.010203605, 0.023095784, 0.00087114252, -0.000030730079, 0.0000069457986],
+    nd: 1.62588,
+    vd: 35.74,
+    PgF: 0.5875,
+    code6: "626357",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-F1 page (formula 3 polynomial).",
+  },
+  {
+    name: "LAC12",
+    vendor: "Hoya",
+    polynomial: [2.7634844, -0.011068339, 0.018246442, 0.00037697356, -0.000017788655, 0.0000015314262],
+    nd: 1.6779,
+    vd: 55.52,
+    PgF: 0.5473,
+    code6: "678555",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; LAC12 page (formula 3 polynomial).",
+  },
+  {
+    name: "TAFD32",
+    vendor: "Hoya",
+    polynomial: [3.396108, -0.013087065, 0.034853516, 0.00079083177, 0.0000076995212, 0.0000023516742],
+    nd: 1.8707,
+    vd: 40.73,
+    code6: "871407",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAFD32 page (formula 3 polynomial).",
+  },
+  {
+    name: "E-F2",
+    vendor: "Hoya",
+    polynomial: [2.5561822, -0.010237038, 0.022470348, 0.00084580679, -0.000033087089, 0.0000071054919],
+    nd: 1.62004,
+    vd: 36.3,
+    code6: "620363",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; formula 3 polynomial.",
+  },
+  {
+    name: "NBFD13",
+    vendor: "Hoya",
+    polynomial: [3.1729223, -0.014248534, 0.029168114, 0.001342876, -0.000096178753, 0.0000084042556],
+    nd: 1.8061,
+    vd: 40.73,
+    code6: "806407",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; formula 3 polynomial.",
+  },
+  {
+    name: "E-F3",
+    vendor: "Hoya",
+    polynomial: [2.5362697, -0.010610424, 0.021233821, 0.00099012105, -0.000059984348, 0.0000082354366],
+    nd: 1.61293,
+    vd: 36.96,
+    code6: "613370",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; formula 3 polynomial.",
+  },
+  /* ────── Phase 19 Sigma patent-audit additions (May 2026) ──────
+   * Coefficient-backed HOYA rows needed by the Sigma 105/1.4, 20/1.4,
+   * and 14-24/2.8 patent audit. All are Zemax formula-3 polynomial rows
+   * from HOYA's 2017 public AGF as mirrored by refractiveindex.info.
+   */
+  {
+    name: "PCD40",
+    vendor: "Hoya",
+    polynomial: [2.5844463, -0.0090521858, 0.014301987, 0.0001705077, 0.000004964536, -0.00000012927479],
+    nd: 1.61997,
+    vd: 63.88,
+    PgF: 0.5457,
+    code6: "620639",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; PCD40 page (formula 3 polynomial).",
+  },
+  {
+    name: "M-LAC130",
+    vendor: "Hoya",
+    polynomial: [2.813412, -0.01332448, 0.01927865, 0.0004124452, -0.000009940432, 0.0000009622024],
+    nd: 1.6935,
+    vd: 53.2,
+    PgF: 0.5484,
+    code6: "694532",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-LAC130 page (formula 3 polynomial).",
+  },
+  {
+    name: "M-BACD5N",
+    vendor: "Hoya",
+    polynomial: [2.48951, -0.01224496, 0.01294262, 0.0003490449, -0.00001777228, 0.0000009113719],
+    nd: 1.58913,
+    vd: 61.25,
+    PgF: 0.5401,
+    code6: "589613",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-BACD5N page (formula 3 polynomial).",
+  },
+  {
+    name: "M-TAFD305",
+    vendor: "Hoya",
+    polynomial: [3.327184, -0.01343242, 0.033398, 0.001006047, -0.00002136908, 0.000003972561],
+    nd: 1.85135,
+    vd: 40.1,
+    PgF: 0.5697,
+    code6: "851401",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-TAFD305 page (formula 3 polynomial).",
+  },
+  {
+    name: "BAFD7",
+    vendor: "Hoya",
+    polynomial: [2.8217211, -0.0097265483, 0.024345979, 0.00078406305, -0.000022431497, 0.0000046726029],
+    nd: 1.70154,
+    vd: 41.15,
+    PgF: 0.5774,
+    code6: "702412",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; BAFD7 page (formula 3 polynomial).",
+  },
+  {
+    name: "MC-NBFD135",
+    vendor: "Hoya",
+    polynomial: [3.1819698, -0.014926522, 0.028987649, 0.0012223654, -0.000052916688, 0.0000047807227],
+    nd: 1.80834,
+    vd: 40.92,
+    code6: "808409",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; MC-NBFD135 page (formula 3 polynomial).",
+  },
+  {
+    name: "M-PCD51",
+    vendor: "Hoya",
+    polynomial: [2.502026, -0.01108292, 0.0112772, 0.0005715408, -0.00005587455, 0.000002905292],
+    nd: 1.59201,
+    vd: 67.02,
+    code6: "592670",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-PCD51 page (formula 3 polynomial).",
+  },
+  {
+    name: "M-TAC80",
+    vendor: "Hoya",
+    polynomial: [2.9351948, -0.016193458, 0.018392287, 0.0010086167, -0.000087565474, 0.0000046921333],
+    nd: 1.72903,
+    vd: 54.04,
+    code6: "729540",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-TAC80 page (formula 3 polynomial).",
+  },
+  {
+    name: "M-LAF81",
+    vendor: "Hoya",
+    polynomial: [2.9167886, -0.012398199, 0.026675237, 0.00065190504, 0.0000046956175, 0.000002691054],
+    nd: 1.73077,
+    vd: 40.5,
+    code6: "731405",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-LAF81 page (formula 3 polynomial).",
+  },
+  {
+    name: "TAC2",
+    vendor: "Hoya",
+    polynomial: [2.9717137, -0.014952593, 0.020162868, 0.00094072283, -0.000088614104, 0.0000053191242],
+    nd: 1.741,
+    vd: 52.6,
+    code6: "741526",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; TAC2 page (formula 3 polynomial).",
+  },
+  {
+    name: "M-TAF401",
+    vendor: "Hoya",
+    polynomial: [3.07279, -0.01366385, 0.02527346, 0.000612182, -0.00001147045, 0.000001543],
+    nd: 1.77377,
+    vd: 47.17,
+    code6: "774472",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-TAF401 page (formula 3 polynomial).",
+  },
+  {
+    name: "M-TAF31",
+    vendor: "Hoya",
+    polynomial: [3.164358, -0.01403321, 0.02767034, 0.0006403025, -0.000006349234, 0.000001615744],
+    nd: 1.80139,
+    vd: 45.45,
+    code6: "801455",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-TAF31 page (formula 3 polynomial).",
+  },
+  {
+    name: "M-NBFD10",
+    vendor: "Hoya",
+    polynomial: [3.264623, -0.01581977, 0.03225894, 0.001709312, -0.0001081957, 0.00001058323],
+    nd: 1.83441,
+    vd: 37.28,
+    code6: "834373",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; M-NBFD10 page (formula 3 polynomial).",
+  },
+  {
+    name: "FDS16-W",
+    vendor: "Hoya",
+    polynomial: [3.6906098, -0.027849024, 0.063139051, 0.012443698, -0.0014774085, 0.00017478701],
+    nd: 1.98612,
+    vd: 16.48,
+    code6: "986165",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; FDS16-W page (formula 3 polynomial).",
+  },
+  /* Phase 23 named-token additions (June 2026)
+   * Direct records from Hoya/Ohara Zemax data mirrored by refractiveindex.info.
+   * BSC3 is resolved through the legacy Hoya E-C3 optical constants; incompatible
+   * 583/594 annotations must use their own matching catalog row instead.
+   */
+  {
+    name: "E-FD13",
+    vendor: "Hoya",
+    polynomial: [2.9205906, -0.013681618, 0.03256776, 0.0028725519, -0.00024277107, 0.000026614514],
+    nd: 1.74077,
+    vd: 27.76,
+    PgF: 0.6077,
+    code6: "741278",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FD13 page (formula 3 polynomial).",
+  },
+  {
+    name: "E-FD10",
+    vendor: "Hoya",
+    polynomial: [2.881518, -0.013228312, 0.03145559, 0.0026851666, -0.00022577544, 0.000024693268],
+    nd: 1.72825,
+    vd: 28.32,
+    PgF: 0.6059,
+    code6: "728283",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-FD10 page (formula 3 polynomial).",
+  },
+  {
+    name: "BACD4",
+    vendor: "Hoya",
+    polynomial: [2.5584621, -0.0097457283, 0.015110495, 0.00023273333, 0.0000011137249, 0.00000016424767],
+    nd: 1.61272,
+    vd: 58.58,
+    PgF: 0.5448,
+    code6: "613586",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; BACD4 page (formula 3 polynomial).",
+  },
+  {
+    name: "E-C3",
+    vendor: "Hoya",
+    polynomial: [2.2725452, -0.008410491, 0.011306843, 0.00037782297, -0.000026903423, 0.0000015976489],
+    nd: 1.51823,
+    vd: 58.96,
+    PgF: 0.5442,
+    code6: "518590",
+    source: "Hoya Zemax catalog 2017-04-01 via refractiveindex.info; E-C3 page (formula 3 polynomial).",
+  },
+];
