+++ src/optics/glassCatalogEntries/ohara.ts
@@ -1,0 +1,1487 @@
+/**
+ * Vendor-published Ohara optical glass dispersion entries.
+ */
+
+import type { GlassEntry } from "../glassCatalogTypes.js";
+
+export const OHARA_GLASS_ENTRIES: readonly GlassEntry[] = [
+  {
+    name: "S-BSL7",
+    vendor: "Ohara",
+    B: [1.03773418, 0.231473863, 1.01072958],
+    C: [0.00611873959, 0.0202423191, 103.677302],
+    nd: 1.51633,
+    vd: 64.14,
+    PgF: 0.5349,
+    code6: "517641",
+    source: "Ohara optical glass catalog (S-BSL7 datasheet).",
+  },
+  /* ────── Ohara — primary catalog vendor for this codebase ──────
+   * S-FPL51 alone accounts for 74 declarations across the lens library; the
+   * remaining Ohara entries together cover several hundred more. Coefficients
+   * transcribed verbatim from the vendor-published Sellmeier-1 form (AGF
+   * dispersion form 2). Round-trip at d-line is well inside 1e-4.
+   */
+  {
+    name: "S-FPL51",
+    vendor: "Ohara",
+    B: [1.17010505, 0.0475710783, 0.763832445],
+    C: [0.00616203924, 0.0263372876, 141.882642],
+    nd: 1.496999,
+    vd: 81.5459,
+    PgF: 0.5346,
+    code6: "497816",
+    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
+  },
+  {
+    name: "S-FPL53",
+    vendor: "Ohara",
+    B: [0.983532327, 0.069568814, 1.11409238],
+    C: [0.00492234955, 0.0193581091, 264.275294],
+    nd: 1.43875,
+    vd: 94.946,
+    PgF: 0.5302,
+    code6: "439950",
+    source:
+      "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data. Super-ED defining glass for many APO triplets.",
+  },
+  {
+    name: "S-FSL5",
+    vendor: "Ohara",
+    B: [1.17447043, 0.0140056154, 1.19272435],
+    C: [0.00841855181, -0.0581790767, 129.599726],
+    nd: 1.48749,
+    vd: 70.2363,
+    PgF: 0.5279,
+    code6: "487702",
+    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
+  },
+  {
+    name: "S-LAL18",
+    vendor: "Ohara",
+    B: [1.50276318, 0.430224497, 1.3472606],
+    C: [0.0145462356, -0.00332784153, 93.3508342],
+    nd: 1.729157,
+    vd: 54.68,
+    PgF: 0.5432,
+    code6: "729547",
+    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
+  },
+  {
+    name: "S-LAH58",
+    vendor: "Ohara",
+    B: [1.78764964, 0.6526356, 1.79914564],
+    C: [0.00847378536, 0.0313126408, 132.788001],
+    nd: 1.882997,
+    vd: 40.7651,
+    PgF: 0.5664,
+    code6: "883408",
+    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
+  },
+  {
+    name: "S-LAH55V",
+    vendor: "Ohara",
+    B: [1.97025325, 0.30489414, 1.39214665],
+    C: [0.00991088134, 0.0383202295, 97.7785249],
+    nd: 1.834807,
+    vd: 42.7253,
+    PgF: 0.5644,
+    source:
+      "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data. Vacuum-melt variant; distinct from non-V.",
+  },
+  {
+    name: "S-LAH65V",
+    vendor: "Ohara",
+    B: [1.81419034, 0.361376301, 1.32729484],
+    C: [0.00874935029, 0.0318352836, 91.3406898],
+    nd: 1.803999,
+    vd: 46.5834,
+    PgF: 0.5566,
+    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
+  },
+  {
+    name: "S-LAH79",
+    vendor: "Ohara",
+    B: [2.32557148, 0.507967133, 2.43087198],
+    C: [0.0132895208, 0.0528335449, 161.122408],
+    nd: 2.0033,
+    vd: 28.2733,
+    PgF: 0.5985,
+    code6: "003283",
+    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
+  },
+  {
+    name: "S-NBH5",
+    vendor: "Ohara",
+    B: [1.47544521, 0.193060095, 1.5093901],
+    C: [0.0095583674, 0.0460430483, 126.422746],
+    nd: 1.654115,
+    vd: 39.6828,
+    PgF: 0.5735,
+    code6: "654397",
+    source:
+      "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data. KZFS-class with negative ΔPgF; APO-relevant.",
+  },
+  {
+    name: "S-NPH2",
+    vendor: "Ohara",
+    B: [2.0386951, 0.437269641, 2.96711461],
+    C: [0.0170796224, 0.0749254813, 174.155354],
+    nd: 1.92286,
+    vd: 18.8969,
+    PgF: 0.6506,
+    code6: "923189",
+    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
+  },
+  {
+    name: "S-TIH3",
+    vendor: "Ohara",
+    B: [1.64797648, 0.267261917, 2.19772845],
+    C: [0.0121917693, 0.0597893039, 192.15834],
+    nd: 1.739998,
+    vd: 28.296005,
+    code6: "740283",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIH3 page.",
+  },
+  {
+    name: "S-TIH6",
+    vendor: "Ohara",
+    B: [1.77227611, 0.34569125, 2.40788501],
+    C: [0.0131182633, 0.0614479619, 200.753254],
+    nd: 1.805181,
+    vd: 25.4254,
+    PgF: 0.6168,
+    code6: "805254",
+    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
+  },
+  {
+    name: "S-TIH53",
+    vendor: "Ohara",
+    B: [1.87904886, 0.369719775, 2.33730863],
+    C: [0.014412177, 0.063881799, 182.66818],
+    nd: 1.84666,
+    vd: 23.7779,
+    PgF: 0.6213,
+    code6: "847238",
+    source: "Ohara AGF (ohara_2017-11-30.agf), vendor-published Zemax catalog data.",
+  },
+  /* ────── Ohara expansion (Phase 1 buildout, Apr 2026) ──────
+   * Eight high-frequency Ohara glasses each with ≥10 lens-element occurrences
+   * across the library. Sourced from the same OHARA Zemax catalog
+   * (ohara_2017-11-30.agf) as the entries above; coefficients pulled via
+   * refractiveindex.info's mirror of that AGF (form 2 = Sellmeier-1).
+   * PgF computed from Schott's normal-line baseline plus ΔPgF.
+   */
+  {
+    name: "S-PHM52",
+    vendor: "Ohara",
+    B: [1.0996655, 0.478125422, 1.13214074],
+    C: [0.0132718559, -0.000601649685, 130.595472],
+    nd: 1.618,
+    vd: 63.3335,
+    PgF: 0.5424,
+    code6: "618634",
+    source:
+      "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Phosphate crown with notable +ΔPgF; APO-relevant.",
+  },
+  {
+    name: "S-FPM2",
+    vendor: "Ohara",
+    B: [0.761242785, 0.747033375, 0.938928947],
+    C: [0.00321174095, 0.0140234423, 139.52353],
+    nd: 1.59522,
+    vd: 67.7357,
+    PgF: 0.5422,
+    code6: "595677",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Fluorophosphate ED with significant +ΔPgF.",
+  },
+  {
+    name: "S-LAL14",
+    vendor: "Ohara",
+    B: [1.2372097, 0.589722623, 1.3192188],
+    C: [0.015355132, -0.00030789625, 93.7202947],
+    nd: 1.696797,
+    vd: 55.5322,
+    PgF: 0.5422,
+    code6: "697555",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Lanthanum crown.",
+  },
+  {
+    name: "S-LAH66",
+    vendor: "Ohara",
+    B: [1.39280586, 0.679577094, 1.38702069],
+    C: [0.00608475118, 0.0233925351, 95.8354094],
+    nd: 1.772499,
+    vd: 49.5984,
+    PgF: 0.5512,
+    code6: "773496",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum.",
+  },
+  {
+    name: "S-BAL42",
+    vendor: "Ohara",
+    B: [1.39570615, 0.071850507, 1.27129267],
+    C: [0.0112218843, -0.0252117422, 134.49786],
+    nd: 1.583126,
+    vd: 59.3747,
+    PgF: 0.5419,
+    code6: "583594",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Barium crown.",
+  },
+  {
+    name: "S-BSM22",
+    vendor: "Ohara",
+    B: [1.44305741, 0.140786358, 1.26093951],
+    C: [0.0081920891, 0.0356911455, 131.959337],
+    nd: 1.622296,
+    vd: 53.169067,
+    code6: "622532",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BSM22 page.",
+  },
+  {
+    name: "S-LAH63",
+    vendor: "Ohara",
+    B: [1.89458276, 0.268702978, 1.45705526],
+    C: [0.0102277048, 0.0442801243, 104.874927],
+    nd: 1.804398,
+    vd: 39.5862,
+    PgF: 0.5727,
+    code6: "804396",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum.",
+  },
+  {
+    name: "S-LAM66",
+    vendor: "Ohara",
+    B: [1.92094221, 0.219901208, 1.72705231],
+    C: [0.0115075241, 0.0547993543, 120.133674],
+    nd: 1.800999,
+    vd: 34.9674,
+    PgF: 0.5865,
+    code6: "801350",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Lanthanum medium-dispersion.",
+  },
+  {
+    name: "S-LAH55",
+    vendor: "Ohara",
+    B: [1.95615766, 0.319216215, 1.39173189],
+    C: [0.00979338965, 0.0376836296, 94.8775271],
+    nd: 1.834807,
+    vd: 42.7137,
+    PgF: 0.5638,
+    code6: "835427",
+    source:
+      "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Standard melt; distinct from S-LAH55V (vacuum-melt variant) — different Sellmeier coefficients despite shared nominal nd.",
+  },
+  /* ────── Ohara expansion (Phase 2 buildout, Apr 2026) ──────
+   * Six more OHARA glasses targeted at the largest "no candidate" clusters
+   * surfaced by the verified mismatch sweep (see agent_docs/glass-relabel-followup.md).
+   * Each fills a real gap: S-LAH95 covers the ~10 surfaces of "S-LAH79 type (903/313)"
+   * mislabels; S-LAH59 covers the (816/466) cluster; S-PHM53 / S-LAH53 / S-LAH92 /
+   * S-LAH88 fill smaller but distinct clusters. Same AGF source as the entries above.
+   */
+  {
+    name: "S-LAH95",
+    vendor: "Ohara",
+    B: [2.15636617, 0.329558178, 1.72178935],
+    C: [0.012288051, 0.0555507835, 124.43934],
+    nd: 1.903659,
+    vd: 31.3431,
+    PgF: 0.5966,
+    code6: "904313",
+    source:
+      "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum (904/313); resolves the largest cluster of mislabeled S-LAH79 surfaces in the lens library.",
+  },
+  {
+    name: "S-LAH59",
+    vendor: "Ohara",
+    B: [1.51372967, 0.702462343, 1.33600982],
+    C: [0.00705246901, 0.0249488689, 100.085908],
+    nd: 1.816,
+    vd: 46.6206,
+    PgF: 0.5654,
+    code6: "816466",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Lanthanum dense crown.",
+  },
+  {
+    name: "S-LAH53",
+    vendor: "Ohara",
+    B: [1.91811619, 0.253724399, 1.39473885],
+    C: [0.0102147684, 0.0433176011, 101.938021],
+    nd: 1.806098,
+    vd: 40.926,
+    PgF: 0.575,
+    code6: "806409",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum.",
+  },
+  {
+    name: "S-LAH92",
+    vendor: "Ohara",
+    B: [2.10440311, 0.358346161, 1.63010064],
+    C: [0.0108531811, 0.044340592, 123.2498],
+    nd: 1.891899,
+    vd: 37.134,
+    PgF: 0.578,
+    code6: "892371",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum (892/371).",
+  },
+  {
+    name: "S-LAH88",
+    vendor: "Ohara",
+    B: [2.1284434, 0.405082139, 1.67918461],
+    C: [0.0117309815, 0.0508706599, 107.091456],
+    nd: 1.916499,
+    vd: 31.6041,
+    PgF: 0.5915,
+    code6: "917316",
+    source:
+      "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info; Ohara GmbH S-LAH88 datasheet gives d-code 917316.",
+  },
+  {
+    name: "S-PHM53",
+    vendor: "Ohara",
+    B: [1.09775423, 0.434816432, 1.13894976],
+    C: [0.01233694, -0.000372522903, 124.276984],
+    nd: 1.603001,
+    vd: 65.4436,
+    PgF: 0.5337,
+    code6: "603655",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Phosphate crown (603/655).",
+  },
+  /* ────── Ohara expansion (Phase 3 buildout, Apr 2026) ──────
+   * Fifteen additional Ohara glasses covering the next tier of frequency:
+   * dense TIM/TIH flints (S-TIH14, S-TIM35, S-TIM25, S-TIM22, S-TIM2),
+   * ED and fluorophosphate crowns (S-FPL52, S-FPM3),
+   * barium and borosilicate crowns (S-BAL35, S-NSL3),
+   * high-index lanthanum variants (S-LAH51, S-LAH52, S-LAH65, S-LAM54),
+   * and ultra-high-index flints (S-NPH1, S-NPH53).
+   * All sourced from ohara_2017-11-30.agf via refractiveindex.info (form 2).
+   */
+  {
+    name: "S-TIH14",
+    vendor: "Ohara",
+    B: [1.68915108, 0.290462024, 2.37971516],
+    C: [0.0128202514, 0.0618090841, 201.094352],
+    nd: 1.761821,
+    vd: 26.517885,
+    PgF: 0.6142,
+    code6: "762265",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Dense flint (762/265).",
+  },
+  {
+    name: "S-TIM35",
+    vendor: "Ohara",
+    B: [1.55849775, 0.230767007, 1.84436099],
+    C: [0.0115367235, 0.0586095947, 162.981888],
+    nd: 1.698947,
+    vd: 30.127865,
+    PgF: 0.6034,
+    code6: "699301",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Dense flint (699/301).",
+  },
+  {
+    name: "S-FPL52",
+    vendor: "Ohara",
+    B: [1.06785857, 0.0335857718, 1.10219763],
+    C: [0.00699227302, -0.0207608925, 226.496541],
+    nd: 1.455999,
+    vd: 90.288026,
+    PgF: 0.5305,
+    code6: "456903",
+    source:
+      "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. ED fluorophosphate crown (456/903); sister to S-FPL51. C2 is negative — this is valid for fluorophosphate glasses with anomalous UV resonances.",
+  },
+  {
+    name: "S-TIM25",
+    vendor: "Ohara",
+    B: [1.50659233, 0.204786135, 1.92036668],
+    C: [0.0109501562, 0.0574980285, 178.128535],
+    nd: 1.6727,
+    vd: 32.099206,
+    PgF: 0.5991,
+    code6: "673321",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Dense flint (673/321).",
+  },
+  {
+    name: "S-FPM3",
+    vendor: "Ohara",
+    B: [0.809407286, 0.527007033, 0.909127704],
+    C: [0.00376072389, 0.0135654895, 142.503612],
+    nd: 1.53775,
+    vd: 74.703119,
+    PgF: 0.5368,
+    code6: "538747",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Fluorophosphate crown with +ΔPgF.",
+  },
+  {
+    name: "S-BAL35",
+    vendor: "Ohara",
+    B: [0.941357273, 0.546174895, 1.16168917],
+    C: [0.0140333996, 0.000906635683, 114.163758],
+    nd: 1.58913,
+    vd: 61.135024,
+    PgF: 0.5392,
+    code6: "589612",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Barium crown (589/612).",
+  },
+  {
+    name: "S-TIM22",
+    vendor: "Ohara",
+    B: [1.44222294, 0.194432265, 1.74092482],
+    C: [0.0104249404, 0.0550235257, 169.710769],
+    nd: 1.647689,
+    vd: 33.792803,
+    PgF: 0.594,
+    code6: "648338",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Dense flint (648/338).",
+  },
+  {
+    name: "S-NPH1",
+    vendor: "Ohara",
+    B: [1.75156623, 0.364006304, 2.47874141],
+    C: [0.0135004681, 0.0668245147, 170.756006],
+    nd: 1.808095,
+    vd: 22.760817,
+    PgF: 0.6316,
+    code6: "808228",
+    source:
+      "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Ultra-high-index niobophosphate flint (808/228).",
+  },
+  {
+    name: "S-TIM2",
+    vendor: "Ohara",
+    B: [1.42193846, 0.133827968, 1.45060574],
+    C: [0.0107291511, 0.0572587546, 145.381805],
+    nd: 1.620041,
+    vd: 36.263378,
+    PgF: 0.5879,
+    code6: "620363",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Dense flint (620/363).",
+  },
+  {
+    name: "S-TIM3",
+    vendor: "Ohara",
+    B: [1.40691144, 0.128369745, 1.51826191],
+    C: [0.0105633641, 0.0568483105, 152.107924],
+    nd: 1.612929,
+    vd: 37.005342,
+    PgF: 0.586,
+    code6: "613370",
+    source: "Ohara S-TIM/S-FTM table and Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIM3 page.",
+  },
+  {
+    name: "S-LAH65",
+    vendor: "Ohara",
+    B: [1.68191258, 0.493779818, 1.45682822],
+    C: [0.0077668425, 0.0288916181, 99.2574356],
+    nd: 1.804,
+    vd: 46.570373,
+    PgF: 0.5565,
+    code6: "804466",
+    source:
+      "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum (804/466); standard melt — distinct from S-LAH65V (vacuum-melt variant).",
+  },
+  {
+    name: "S-LAH52",
+    vendor: "Ohara",
+    B: [1.85390925, 0.297925555, 1.39382086],
+    C: [0.00955320687, 0.039381685, 102.706848],
+    nd: 1.799516,
+    vd: 42.225007,
+    PgF: 0.5668,
+    code6: "800422",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum (800/422).",
+  },
+  {
+    name: "S-NSL3",
+    vendor: "Ohara",
+    B: [0.882514764, 0.389271907, 1.10693448],
+    C: [0.00464504582, 0.0200551397, 136.234339],
+    nd: 1.518229,
+    vd: 58.902057,
+    PgF: 0.5442,
+    code6: "518590",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Borosilicate crown (518/590).",
+  },
+  {
+    name: "S-LAM54",
+    vendor: "Ohara",
+    B: [1.84213306, 0.175468631, 1.25750878],
+    C: [0.0094399322, 0.0395281122, 86.5463013],
+    nd: 1.756998,
+    vd: 47.82317,
+    PgF: 0.5558,
+    code6: "757478",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Lanthanum medium-dispersion (757/478).",
+  },
+  {
+    name: "S-NPH53",
+    vendor: "Ohara",
+    B: [1.85484904, 0.396194484, 2.43512461],
+    C: [0.0134621486, 0.0631945361, 170.864886],
+    nd: 1.84666,
+    vd: 23.88395,
+    PgF: 0.6226,
+    code6: "847239",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Niobophosphate flint (847/239).",
+  },
+  {
+    name: "S-LAH51",
+    vendor: "Ohara",
+    B: [1.82586991, 0.283023349, 1.35964319],
+    C: [0.00935297152, 0.0373803057, 100.655798],
+    nd: 1.785896,
+    vd: 44.202637,
+    PgF: 0.5626,
+    code6: "786442",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. High-index lanthanum (786/442).",
+  },
+  /* ────── Ohara expansion (Phase 4 buildout, Apr 2026) ──────
+   * S-TIM28: dense TIM flint used in Canon RF, Nikon L35AF/PCE19, Fuji XF zooms.
+   * S-BSM14: borosilicate medium crown used in Canon EF/RF and other normal lenses.
+   * Coefficients from OHARA Zemax catalog (ohara_2017-11-30.agf) via refractiveindex.info.
+   */
+  {
+    name: "S-TIM28",
+    vendor: "Ohara",
+    B: [1.5427081, 0.217113891, 1.81904459],
+    C: [0.0113925005, 0.0579224572, 167.697189],
+    nd: 1.688931,
+    vd: 31.07545,
+    PgF: 0.6008,
+    code6: "689311",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Dense flint (689/311).",
+  },
+  {
+    name: "S-BSM14",
+    vendor: "Ohara",
+    B: [1.2828627, 0.247647429, 1.10383999],
+    C: [0.0122902399, -0.00613142361, 106.883378],
+    nd: 1.603112,
+    vd: 60.64108,
+    PgF: 0.5399,
+    code6: "603607",
+    source: "Ohara AGF (ohara_2017-11-30.agf) via refractiveindex.info. Borosilicate medium crown (603/607).",
+  },
+  /* ────── Ohara Phase 5 additions (May 2026) ──────
+   * Sourced from OHARA Zemax catalog (ohara_2017-11-30.agf) via
+   * refractiveindex.info. All entries use formula 2 (standard Sellmeier-1,
+   * K=0). Covers dense flints, NBH niobophosphate family, LAH/LAM/LAL/BAL/BSM
+   * lanthanum and barium families, and NPH ultra-high-index flints not in the
+   * earlier phases.
+   */
+
+  /* Dense flints — TIH / TIM additions */
+  {
+    name: "S-TIH18",
+    vendor: "Ohara",
+    B: [1.59921608, 0.259532164, 2.12454543],
+    C: [0.0116469304, 0.0584824883, 186.927779],
+    nd: 1.721507,
+    vd: 29.23,
+    code6: "722292",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  {
+    name: "S-TIH4",
+    vendor: "Ohara",
+    B: [1.66755531, 0.294411865, 2.49422119],
+    C: [0.0122052137, 0.0597775329, 214.869618],
+    nd: 1.755199,
+    vd: 27.51,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  {
+    name: "S-TIM5",
+    vendor: "Ohara",
+    B: [1.38531342, 0.122372945, 1.40508326],
+    C: [0.0104074567, 0.0557440088, 144.878733],
+    nd: 1.60342,
+    vd: 38.03,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  {
+    name: "S-TIM27",
+    vendor: "Ohara",
+    B: [1.4168047, 0.196785057, 1.68001322],
+    C: [0.0100732158, 0.0537616908, 164.672436],
+    nd: 1.639799,
+    vd: 34.47,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  {
+    name: "S-NBH8",
+    vendor: "Ohara",
+    B: [1.61344136, 0.257295888, 1.98364455],
+    C: [0.0106386752, 0.0487071624, 159.784404],
+    nd: 1.720467,
+    vd: 34.71,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  /* NBH niobophosphate family — anomalous partial dispersion dense flints */
+  {
+    name: "S-NBH51",
+    vendor: "Ohara",
+    B: [1.71203689, 0.255989588, 1.81456998],
+    C: [0.0107724134, 0.0488593504, 136.359013],
+    nd: 1.749504,
+    vd: 35.33,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  {
+    name: "S-NBH52",
+    vendor: "Ohara",
+    B: [1.50305799, 0.221715926, 1.84496391],
+    C: [0.00999021738, 0.0450327698, 163.722302],
+    nd: 1.672999,
+    vd: 38.15,
+    code6: "673382",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  {
+    name: "S-NBH55",
+    vendor: "Ohara",
+    B: [1.83145156, 0.287818024, 2.152083],
+    C: [0.0122443139, 0.057387731, 186.099124],
+    nd: 1.799999,
+    vd: 29.84,
+    code6: "800298",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  {
+    name: "S-NBH56",
+    vendor: "Ohara",
+    B: [1.85191438, 0.431102852, 3.45278284],
+    C: [0.013273262, 0.0585944644, 239.357089],
+    nd: 1.854779,
+    vd: 24.8,
+    code6: "855248",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  /* LAH lanthanum dense flint additions */
+  {
+    name: "S-LAH55VS",
+    vendor: "Ohara",
+    B: [1.92591095, 0.34895346, 1.42230744],
+    C: [0.0096115249, 0.036513298, 103.36409],
+    nd: 1.834809,
+    vd: 42.74,
+    source:
+      "Ohara Zemax catalog 2017-11-30 via refractiveindex.info. Vacuum-melt stabilized variant of S-LAH55V; distinct Sellmeier.",
+  },
+  {
+    name: "S-LAH60",
+    vendor: "Ohara",
+    B: [1.95243469, 0.30710021, 1.56578094],
+    C: [0.0106442437, 0.0456735302, 110.28141],
+    nd: 1.834,
+    vd: 37.16,
+    code6: "834372",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  {
+    name: "S-LAH60V",
+    vendor: "Ohara",
+    B: [2.05081962, 0.208475257, 1.31486394],
+    C: [0.0116035991, 0.0526489359, 99.38065],
+    nd: 1.833999,
+    vd: 37.21,
+    source:
+      "Ohara Zemax catalog 2017-11-30 via refractiveindex.info. Vacuum-melt variant; distinct Sellmeier from S-LAH60.",
+  },
+  {
+    name: "S-LAH64",
+    vendor: "Ohara",
+    B: [1.83021453, 0.29156359, 1.28544024],
+    C: [0.0090482329, 0.0330756689, 89.3675501],
+    nd: 1.788001,
+    vd: 47.37,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  {
+    name: "S-LAH89",
+    vendor: "Ohara",
+    B: [1.95118827, 0.377607223, 1.47757262],
+    C: [0.00976560799, 0.0382232043, 112.23672],
+    nd: 1.8515,
+    vd: 40.78,
+    code6: "852408",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  {
+    name: "S-LAH97",
+    vendor: "Ohara",
+    B: [1.0273018, 0.989293564, 1.25781057],
+    C: [0.0183406129, 0.00371264195, 87.85105],
+    nd: 1.755,
+    vd: 52.32,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  /* LAM / LAL lanthanum crown additions */
+  {
+    name: "S-LAM2",
+    vendor: "Ohara",
+    B: [1.7713, 0.19581423, 1.19487834],
+    C: [0.00976652444, 0.0412718628, 110.458122],
+    nd: 1.743997,
+    vd: 44.79,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  {
+    name: "S-LAM51",
+    vendor: "Ohara",
+    B: [1.638472, 0.188330533, 1.47502357],
+    C: [0.00904853452, 0.0372740173, 137.77005],
+    nd: 1.699998,
+    vd: 48.08,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  {
+    name: "S-LAL8",
+    vendor: "Ohara",
+    B: [1.30663291, 0.571377253, 1.24303605],
+    C: [0.00611862448, 0.021272147, 90.6285686],
+    nd: 1.712995,
+    vd: 53.87,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  {
+    name: "S-LAL9",
+    vendor: "Ohara",
+    B: [1.16195687, 0.644860099, 1.25062221],
+    C: [0.0159659509, 0.000505502467, 93.8284169],
+    nd: 1.691002,
+    vd: 54.82,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  /* BAM / BAL barium additions */
+  {
+    name: "S-BAM4",
+    vendor: "Ohara",
+    B: [1.41059317, 0.111201306, 1.34148939],
+    C: [0.00963312192, 0.049877821, 152.237696],
+    nd: 1.60562,
+    vd: 43.71,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  {
+    name: "S-BAL14",
+    vendor: "Ohara",
+    B: [1.27553696, 0.146083393, 1.16754699],
+    C: [0.00749692359, 0.031042153, 128.947092],
+    nd: 1.568832,
+    vd: 56.36,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  /* BSM barium crown additions */
+  {
+    name: "S-BSM18",
+    vendor: "Ohara",
+    B: [0.927886025, 0.708858526, 1.18610897],
+    C: [0.00417549199, 0.0184691838, 122.210416],
+    nd: 1.638539,
+    vd: 55.38,
+    code6: "639554",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  {
+    name: "S-BSM81",
+    vendor: "Ohara",
+    B: [0.996356844, 0.651392837, 1.22432622],
+    C: [0.0144821587, 0.00154826389, 89.9818604],
+    nd: 1.639999,
+    vd: 60.08,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  /* NSL light crown addition */
+  {
+    name: "S-NSL5",
+    vendor: "Ohara",
+    B: [1.04574577, 0.239613026, 1.1590685],
+    C: [0.0058523228, 0.0236858752, 131.329061],
+    nd: 1.522494,
+    vd: 59.84,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  /* NPH ultra-high-index dense flint addition */
+  {
+    name: "S-NPH4",
+    vendor: "Ohara",
+    B: [1.93563931, 0.449596478, 2.71828573],
+    C: [0.0152585289, 0.0696815778, 170.327149],
+    nd: 1.89286,
+    vd: 20.36,
+    code6: "893204",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info.",
+  },
+  /* ────── Phase 6 priority additions (May 2026) ──────
+   * High-impact unresolved glasses from the catalog-priority sweep. Formula-2
+   * entries are standard Sellmeier. Formula-3 Hoya/Sumita entries store the
+   * vendor polynomial exactly rather than using a fitted Sellmeier surrogate.
+   */
+
+  /* Ohara priority additions */
+  {
+    name: "S-FPL55",
+    vendor: "Ohara",
+    B: [0.839067682, 0.214083503, 1.47914677],
+    C: [0.00795286639, -0.00232581717, 340.0437],
+    nd: 1.43875,
+    vd: 94.661059,
+    PgF: 0.5303,
+    code6: "439948",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-FPL55 page.",
+  },
+  {
+    name: "S-LAH98",
+    vendor: "Ohara",
+    B: [2.28510629, 0.385532264, 2.0655112],
+    C: [0.0122178962, 0.0514752342, 145.92087],
+    nd: 1.95375,
+    vd: 32.32,
+    PgF: 0.5907,
+    code6: "954323",
+    source: "Ohara S-LAH98 datasheet (Ohara public PDF, OHARA 24-01).",
+  },
+  {
+    name: "S-LAH99",
+    vendor: "Ohara",
+    B: [2.39140662, 0.439219228, 2.38358467],
+    C: [0.01314675, 0.0553226042, 161.2599],
+    nd: 2.001,
+    vd: 29.14,
+    PgF: 0.5997,
+    code6: "001291",
+    source: "Ohara S-LAH99 datasheet (Ohara public PDF, 2026 page crawl).",
+  },
+  {
+    name: "L-LAH85V",
+    vendor: "Ohara",
+    B: [1.97460503, 0.361903592, 1.38433241],
+    C: [0.0101854258, 0.0384829118, 103.713627],
+    nd: 1.854,
+    vd: 40.378368,
+    PgF: 0.5688,
+    code6: "854404",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; L-LAH85V page.",
+  },
+  {
+    name: "S-LAM60",
+    vendor: "Ohara",
+    B: [1.60673056, 0.36641564, 1.31761804],
+    C: [0.0077504614, 0.0289967611, 93.0720709],
+    nd: 1.743198,
+    vd: 49.339444,
+    PgF: 0.5523,
+    code6: "743493",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAM60 page.",
+  },
+  {
+    name: "S-NPH3",
+    vendor: "Ohara",
+    B: [2.09834903, 0.489088388, 2.94009268],
+    C: [0.0179123869, 0.0776653353, 160.930428],
+    nd: 1.95906,
+    vd: 17.471252,
+    PgF: 0.661,
+    code6: "959175",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-NPH3 page.",
+  },
+  {
+    name: "S-NPH5",
+    vendor: "Ohara",
+    B: [1.89108996, 0.395220126, 2.20492127],
+    C: [0.0141164499, 0.0662834445, 148.6807],
+    nd: 1.858955,
+    vd: 22.728736,
+    PgF: 0.6293,
+    code6: "859227",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-NPH5 page.",
+  },
+  {
+    name: "S-LAH63Q",
+    vendor: "Ohara",
+    B: [1.96723017, 0.194953915, 1.25386282],
+    C: [0.0110456086, 0.0497137061, 104.84352],
+    nd: 1.8044,
+    vd: 39.580489,
+    PgF: 0.576,
+    code6: "804396",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAH63Q page.",
+  },
+  {
+    name: "S-LAH65VS",
+    vendor: "Ohara",
+    B: [1.76068422, 0.414128906, 1.33415439],
+    C: [0.00853607198, 0.0301826383, 98.09421],
+    nd: 1.804,
+    vd: 46.527532,
+    PgF: 0.557,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAH65VS page.",
+  },
+  {
+    name: "S-NBM51",
+    vendor: "Ohara",
+    B: [1.37023101, 0.177665568, 1.30515471],
+    C: [0.00871920342, 0.0405725552, 112.703058],
+    nd: 1.613397,
+    vd: 44.267706,
+    PgF: 0.5628,
+    code6: "613443",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-NBM51 page.",
+  },
+  {
+    name: "S-FTM16",
+    vendor: "Ohara",
+    B: [1.32940907, 0.141512125, 1.44299068],
+    C: [0.0102377287, 0.0578081956, 150.597139],
+    nd: 1.592701,
+    vd: 35.310075,
+    PgF: 0.5934,
+    code6: "593353",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-FTM16 page.",
+  },
+  {
+    name: "S-NSL36",
+    vendor: "Ohara",
+    B: [1.09666153, 0.168990073, 1.20580827],
+    C: [0.00667491123, 0.033609545, 141.668738],
+    nd: 1.517417,
+    vd: 52.430903,
+    PgF: 0.5554,
+    code6: "517524",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-NSL36 page.",
+  },
+  /* ────── Phase 8 unresolved-token additions (May 2026) ──────
+   * Clean, sourceable names from the unresolved-glass report. Ambiguous
+   * high-frequency labels such as NBFD3, S-NPH7, and TAFD25 are intentionally
+   * left for per-lens patent relabeling rather than broad catalog resolution.
+   */
+
+  /* Ohara unresolved-token additions */
+  {
+    name: "S-LAH93",
+    vendor: "Ohara",
+    B: [2.17884378, 0.325508683, 1.70608646],
+    C: [0.0115760879, 0.0495982503, 128.91358],
+    nd: 1.90525,
+    vd: 35.036826,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAH93 page.",
+  },
+  {
+    name: "S-TIH11",
+    vendor: "Ohara",
+    B: [1.72677471, 0.324568628, 2.65816809],
+    C: [0.0129369958, 0.0618255245, 221.904637],
+    nd: 1.784723,
+    vd: 25.683446,
+    PgF: 0.6168,
+    code6: "785257",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIH11 page.",
+  },
+  {
+    name: "S-TIL25",
+    vendor: "Ohara",
+    B: [1.32122534, 0.123824976, 1.43685254],
+    C: [0.00952091436, 0.0516062665, 149.064883],
+    nd: 1.581439,
+    vd: 40.747566,
+    PgF: 0.5772,
+    code6: "581407",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIL25 page.",
+  },
+  {
+    name: "S-BAH27",
+    vendor: "Ohara",
+    B: [1.68939052, 0.133081013, 1.41165515],
+    C: [0.0103598193, 0.0533982239, 126.515503],
+    nd: 1.701536,
+    vd: 41.23944,
+    PgF: 0.5762,
+    code6: "702412",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BAH27 page.",
+  },
+  {
+    name: "S-NBH52V",
+    vendor: "Ohara",
+    B: [1.51336868, 0.212341478, 1.54149143],
+    C: [0.00987077827, 0.0462843662, 126.97851],
+    nd: 1.673,
+    vd: 38.256077,
+    PgF: 0.5756,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-NBH52V page.",
+  },
+  {
+    name: "S-TIH53W",
+    vendor: "Ohara",
+    B: [1.87904886, 0.369719775, 2.33730863],
+    C: [0.014412177, 0.063881799, 182.66818],
+    nd: 1.84666,
+    vd: 23.77794,
+    PgF: 0.6213,
+    code6: "847238",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIH53W page. Improved-transmittance variant.",
+  },
+  {
+    name: "S-BAH28",
+    vendor: "Ohara",
+    B: [1.69493484, 0.192890298, 1.56385948],
+    C: [0.010272319, 0.052118764, 137.818035],
+    nd: 1.72342,
+    vd: 37.955602,
+    PgF: 0.5835,
+    code6: "723380",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BAH28 page.",
+  },
+  {
+    name: "S-LAL59",
+    vendor: "Ohara",
+    B: [1.13962742, 0.805227838, 1.29488061],
+    C: [0.00493294862, 0.020247996, 93.4746507],
+    nd: 1.733997,
+    vd: 51.47057,
+    PgF: 0.5476,
+    code6: "734515",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAL59 page.",
+  },
+  {
+    name: "S-TIL27",
+    vendor: "Ohara",
+    B: [1.31433154, 0.112300168, 1.413901],
+    C: [0.00950404477, 0.0524112772, 148.429972],
+    nd: 1.575006,
+    vd: 41.505207,
+    PgF: 0.5764,
+    code6: "575415",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIL27 page.",
+  },
+  {
+    name: "S-TIL6",
+    vendor: "Ohara",
+    B: [1.17701777, 0.12795803, 1.34740124],
+    C: [0.00771087686, 0.0411325328, 154.531692],
+    nd: 1.531717,
+    vd: 48.840718,
+    PgF: 0.5623,
+    code6: "532489",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIL6 page.",
+  },
+  /* ────── Phase 9 named-token additions (May 2026) ──────
+   * Directly sourceable 2+ occurrence names whose catalog constants match at
+   * least one current lens-data declaration. Ambiguous names from the same
+   * survey remain deferred to per-lens patent relabeling.
+   */
+
+  /* Ohara Phase 9 additions */
+  {
+    name: "S-BSM16",
+    vendor: "Ohara",
+    B: [1.14490383, 0.439563911, 1.27688079],
+    C: [0.0137034916, -0.00186514205, 119.535585],
+    nd: 1.620411,
+    vd: 60.289582,
+    PgF: 0.5412,
+    code6: "620603",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BSM16 page.",
+  },
+  {
+    name: "S-TIH13",
+    vendor: "Ohara",
+    B: [1.62224674, 0.293844589, 1.99225164],
+    C: [0.0118368386, 0.0590208025, 171.959976],
+    nd: 1.740769,
+    vd: 27.789359,
+    PgF: 0.6101,
+    code6: "741278",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIH13 page.",
+  },
+  {
+    name: "S-NBH53V",
+    vendor: "Ohara",
+    B: [1.65444141, 0.267453927, 2.14530347],
+    C: [0.0112485533, 0.052027274, 167.3661],
+    nd: 1.737999,
+    vd: 32.32659,
+    PgF: 0.5902,
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-NBH53V page.",
+  },
+  {
+    name: "S-TIH23",
+    vendor: "Ohara",
+    B: [1.73986485, 0.313894918, 2.31093206],
+    C: [0.01294413, 0.0612116868, 197.420482],
+    nd: 1.784696,
+    vd: 26.291221,
+    PgF: 0.6142,
+    code6: "785263",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIH23 page.",
+  },
+  /* Phase 9 six-digit code-family additions */
+  {
+    name: "S-LAH96",
+    vendor: "Ohara",
+    B: [1.85078519, 0.189204854, 1.19763137],
+    C: [0.00940657541, 0.0380345187, 101.426835],
+    nd: 1.76385,
+    vd: 48.488692,
+    PgF: 0.5581,
+    code6: "764485",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAH96 page.",
+  },
+  /* Ohara unresolved-token additions */
+  {
+    name: "S-BAL2",
+    vendor: "Ohara",
+    B: [1.30923813, 0.114137353, 1.17882259],
+    C: [0.00838873953, 0.0399436485, 140.257892],
+    nd: 1.570989,
+    vd: 50.799858,
+    PgF: 0.5579,
+    code6: "571508",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BAL2 page.",
+  },
+  {
+    name: "S-LAM55",
+    vendor: "Ohara",
+    B: [1.85412979, 0.165450323, 1.27255422],
+    C: [0.0108438152, 0.051405098, 109.986837],
+    nd: 1.762001,
+    vd: 40.100176,
+    PgF: 0.5763,
+    code6: "762401",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAM55 page.",
+  },
+  {
+    name: "S-TIH10",
+    vendor: "Ohara",
+    B: [1.61549392, 0.262433239, 2.09426189],
+    C: [0.0119830897, 0.059651024, 181.657554],
+    nd: 1.72825,
+    vd: 28.460599,
+    PgF: 0.6082,
+    code6: "728285",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIH10 page.",
+  },
+  {
+    name: "S-BSM28",
+    vendor: "Ohara",
+    B: [1.43822841, 0.128100017, 1.3435553],
+    C: [0.0085977975, 0.0408617854, 143.70989],
+    nd: 1.617722,
+    vd: 49.815052,
+    PgF: 0.5594,
+    code6: "618498",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BSM28 page.",
+  },
+  {
+    name: "S-BSM71",
+    vendor: "Ohara",
+    B: [1.50847885, 0.158099826, 1.36815368],
+    C: [0.00812769076, 0.0354200898, 136.110038],
+    nd: 1.648498,
+    vd: 53.023881,
+    PgF: 0.5536,
+    code6: "649530",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BSM71 page.",
+  },
+  {
+    name: "S-LAL13",
+    vendor: "Ohara",
+    B: [0.980071267, 0.832904776, 1.28111995],
+    C: [0.00389123698, 0.0189164592, 98.9052676],
+    nd: 1.693501,
+    vd: 53.205327,
+    PgF: 0.5462,
+    code6: "694532",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAL13 page.",
+  },
+  {
+    name: "S-LAM3",
+    vendor: "Ohara",
+    B: [1.64258713, 0.23963461, 1.22483026],
+    C: [0.0086824602, 0.0351226242, 116.604369],
+    nd: 1.717004,
+    vd: 47.927969,
+    PgF: 0.5598,
+    code6: "717479",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAM3 page.",
+  },
+  {
+    name: "S-LAM52",
+    vendor: "Ohara",
+    B: [1.73442942, 0.15155391, 1.46225433],
+    C: [0.0100690928, 0.0470634701, 140.084396],
+    nd: 1.72,
+    vd: 43.689812,
+    PgF: 0.5694,
+    code6: "720437",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAM52 page.",
+  },
+  {
+    name: "PBL25",
+    vendor: "Ohara",
+    polynomial: [2.4465509, -0.0097768464, 0.01753112, 0.00095250889, -0.000054837761, 0.0000047303899],
+    nd: 1.581439,
+    vd: 40.753351,
+    PgF: 0.5748,
+    code6: "581408",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; PBL25 page (formula 3 polynomial).",
+  },
+  {
+    name: "S-LAM59",
+    vendor: "Ohara",
+    B: [1.63056133, 0.186994897, 1.30014289],
+    C: [0.00899690705, 0.0368011993, 122.239544],
+    nd: 1.697002,
+    vd: 48.516066,
+    PgF: 0.5581,
+    code6: "697485",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAM59 page.",
+  },
+  {
+    name: "S-LAL10",
+    vendor: "Ohara",
+    B: [1.52812575, 0.367965267, 1.11751784],
+    C: [0.00776817644, 0.0272026548, 88.86974],
+    nd: 1.719995,
+    vd: 50.230564,
+    PgF: 0.5512,
+    code6: "720502",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAL10 page.",
+  },
+  {
+    name: "S-TIL2",
+    vendor: "Ohara",
+    B: [1.23401499, 0.0959796833, 1.20503991],
+    C: [0.00869507801, 0.0465611429, 137.953301],
+    nd: 1.54072,
+    vd: 47.227127,
+    code6: "541472",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIL2 page.",
+  },
+  {
+    name: "S-TIL26",
+    vendor: "Ohara",
+    B: [1.31066488, 0.0941903094, 1.23292644],
+    C: [0.00968897812, 0.0527763106, 133.296422],
+    nd: 1.567322,
+    vd: 42.817076,
+    code6: "567428",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-TIL26 page.",
+  },
+  {
+    name: "S-PHM52Q",
+    vendor: "Ohara",
+    B: [1.2696875, 0.308418136, 1.04078976],
+    C: [0.00621622949, 0.0207785265, 112.05125],
+    nd: 1.617999,
+    vd: 63.321449,
+    code6: "618633",
+    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
+  },
+  {
+    name: "S-FPM4",
+    vendor: "Ohara",
+    B: [0.685585084, 0.623380215, 0.914178386],
+    C: [0.0028817201, 0.0124701707, 153.5772],
+    nd: 1.52841,
+    vd: 76.452824,
+    code6: "528765",
+    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
+  },
+  {
+    name: "S-NBH58",
+    vendor: "Ohara",
+    B: [1.71408219, 0.362966167, 2.56486266],
+    C: [0.0116812775, 0.0540847347, 188.7857],
+    nd: 1.7888,
+    vd: 28.42865,
+    code6: "789284",
+    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
+  },
+  {
+    name: "S-BAH11",
+    vendor: "Ohara",
+    B: [1.5713886, 0.147869313, 1.28092846],
+    C: [0.00910807936, 0.0402401684, 130.399367],
+    nd: 1.666718,
+    vd: 48.325247,
+    code6: "667483",
+    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
+  },
+  {
+    name: "S-BAL41",
+    vendor: "Ohara",
+    B: [1.243442, 0.166301104, 1.10586114],
+    C: [0.0116396708, -0.00890464938, 114.11122],
+    nd: 1.563839,
+    vd: 60.666575,
+    code6: "564607",
+    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
+  },
+  {
+    name: "S-LAH52Q",
+    vendor: "Ohara",
+    B: [1.91082318, 0.239854589, 1.16159733],
+    C: [0.0103565352, 0.0413805081, 96.60373],
+    nd: 1.79952,
+    vd: 42.239506,
+    code6: "800422",
+    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
+  },
+  {
+    name: "S-LAH71",
+    vendor: "Ohara",
+    B: [1.98280031, 0.31675845, 2.44472646],
+    C: [0.0118987456, 0.0527156001, 213.220697],
+    nd: 1.850259,
+    vd: 32.27012,
+    code6: "850323",
+    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
+  },
+  {
+    name: "S-LAL7",
+    vendor: "Ohara",
+    B: [0.916121247, 0.765948319, 1.27745023],
+    C: [0.00395889743, 0.0167547425, 110.762706],
+    nd: 1.651597,
+    vd: 58.550655,
+    code6: "652585",
+    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
+  },
+  {
+    name: "S-LAM61",
+    vendor: "Ohara",
+    B: [1.7388333, 0.15093743, 1.12118445],
+    C: [0.00980244105, 0.0433179685, 101.214625],
+    nd: 1.720002,
+    vd: 46.024571,
+    code6: "720460",
+    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
+  },
+  {
+    name: "S-TIH1",
+    vendor: "Ohara",
+    B: [1.60326759, 0.242980935, 1.81313592],
+    C: [0.0118019139, 0.0591363658, 161.218747],
+    nd: 1.717362,
+    vd: 29.518091,
+    code6: "717295",
+    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
+  },
+  {
+    name: "S-TIL1",
+    vendor: "Ohara",
+    B: [1.25088944, 0.0997973327, 1.20583504],
+    C: [0.00883921279, 0.0482685052, 137.414953],
+    nd: 1.548141,
+    vd: 45.784277,
+    code6: "548458",
+    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
+  },
+  {
+    name: "S-TIM8",
+    vendor: "Ohara",
+    B: [1.37262713, 0.112636276, 1.39786421],
+    C: [0.0103220068, 0.0550195044, 147.735609],
+    nd: 1.595509,
+    vd: 39.242273,
+    code6: "596392",
+    source: "Ohara Zemax OpticStudio DATA 2025-06-27 (OHARA_250627.AGF), vendor-published catalog data.",
+  },
+  /* ────── Phase 17 additions (May 2026) ──────
+   * Six-digit missing-Sellmeier audit over Hasselblad, Laowa, Leica, Minolta,
+   * and Nikon rows. Entries below are coefficient-backed public matches or
+   * documented catalog equivalents that round-trip through the d-line guard.
+   */
+  {
+    name: "BPH50",
+    vendor: "Ohara",
+    polynomial: [2.9271459, -0.011970157, 0.031922538, 0.0015457537, -0.000063507309, 0.0000096835139],
+    nd: 1.739997,
+    vd: 31.710001,
+    code6: "740317",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; BPH50 page (formula 3 polynomial).",
+  },
+  {
+    name: "S-LAM58",
+    vendor: "Ohara",
+    B: [1.70984856, 0.173342897, 1.64833565],
+    C: [0.0100852127, 0.0470890831, 157.46852],
+    nd: 1.72,
+    vd: 41.98,
+    PgF: 0.5729,
+    code6: "720420",
+    source: "Ohara GmbH S-LAM58 datasheet, April 2018; Sellmeier constants A1-A3/B1-B3.",
+  },
+  {
+    name: "S-LAL58",
+    vendor: "Ohara",
+    B: [1.06368789, 0.744939067, 1.59178942],
+    C: [0.018519964, 0.00116295862, 156.636025],
+    nd: 1.693495,
+    vd: 50.810652,
+    code6: "694508",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-LAL58 page.",
+  },
+  /* ────── Phase 21 chromatic-coverage additions (June 2026) ──────
+   * Source-backed entries from the generated named-token opportunity queue.
+   * Each row round-trips through assertCatalogConsistent at the existing 1e-4
+   * tolerance. Broad E-LAF7 and S-NPH7 aliases remain intentionally omitted
+   * because current annotations span incompatible nd/vd regions.
+   */
+  {
+    name: "S-LAL12",
+    vendor: "Ohara",
+    B: [0.992053895, 0.771377731, 1.18296264],
+    C: [0.0167095063, 0.00236750156, 105.90108],
+    nd: 1.6779,
+    vd: 55.34,
+    PgF: 0.5472,
+    code6: "678553",
+    source: "Ohara S-LAL12 datasheet, OHARA 25-04 official optical-glass PDF.",
+  },
+  {
+    name: "S-BSM10",
+    vendor: "Ohara",
+    B: [0.945443081, 0.643237376, 1.17752968],
+    C: [0.0157263798, 0.00161924066, 121.361748],
+    nd: 1.6228,
+    vd: 57.05,
+    PgF: 0.5464,
+    code6: "623570",
+    source: "Ohara S-BSM10 datasheet, OHARA 25-04 official optical-glass PDF.",
+  },
+  {
+    name: "S-LAM7",
+    vendor: "Ohara",
+    B: [1.71014712, 0.256943292, 1.63986271],
+    C: [0.010516108, 0.0502809636, 146.181217],
+    nd: 1.7495,
+    vd: 35.28,
+    PgF: 0.5869,
+    code6: "750353",
+    source: "Ohara S-LAM7 datasheet, OHARA 25-04 official optical-glass PDF.",
+  },
+  {
+    name: "L-LAM69",
+    vendor: "Ohara",
+    B: [1.7403896, 0.176996917, 1.76775413],
+    C: [0.010339887, 0.0484822765, 136.671996],
+    nd: 1.73077,
+    vd: 40.51,
+    PgF: 0.5728,
+    code6: "731405",
+    source: "Ohara L-LAM69 datasheet, OHARA 25-04 official optical-glass PDF.",
+  },
+  {
+    name: "S-BSM15",
+    vendor: "Ohara",
+    B: [0.953128328, 0.637613977, 1.65245647],
+    C: [0.00387638985, 0.0185094632, 159.442367],
+    nd: 1.622992,
+    vd: 58.165763,
+    PgF: 0.5459,
+    code6: "623582",
+    source: "Ohara Zemax catalog 2017-11-30 via refractiveindex.info; S-BSM15 page.",
+  },
+];
