+++ src/optics/glassCatalogEntries/schott.ts
@@ -1,0 +1,551 @@
+/**
+ * Vendor-published Schott optical glass dispersion entries.
+ */
+
+import type { GlassEntry } from "../glassCatalogTypes.js";
+
+export const SCHOTT_GLASS_ENTRIES: readonly GlassEntry[] = [
+  /* ────── Universal crowns (BK7-class) ──────
+   * N-BK7 is the universal reference crown; S-BSL7 is Ohara's near-equivalent.
+   * Both are present in nearly every lens design somewhere in the rear group.
+   */
+  {
+    name: "N-BK7",
+    vendor: "Schott",
+    B: [1.03961212, 0.231792344, 1.01046945],
+    C: [0.00600069867, 0.0200179144, 103.560653],
+    nd: 1.5168,
+    vd: 64.17,
+    PgF: 0.5349,
+    code6: "517642",
+    source: "Schott TIE-29 / N-BK7 datasheet (Schott AG, public).",
+  },
+  /* ────── Schott — high-dispersion flints and lanthanum crowns ──────
+   * SF6/SF4/SF1 are the legacy lead flints used widely in vintage designs.
+   * N-LAK8 is a high-frequency lanthanum crown. N-KZFS5 carries notable
+   * negative ΔPgF and is APO-relevant. Coefficients transcribed verbatim
+   * from the vendor-published Sellmeier-1 form (AGF dispersion form 2).
+   */
+  {
+    name: "SF6",
+    vendor: "Schott",
+    B: [1.72448482, 0.390104889, 1.04572858],
+    C: [0.0134871947, 0.0569318095, 118.557185],
+    nd: 1.80518,
+    vd: 25.43,
+    PgF: 0.6102,
+    code6: "805254",
+    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data.",
+  },
+  {
+    name: "SF4",
+    vendor: "Schott",
+    B: [1.61957826, 0.339493189, 1.02566931],
+    C: [0.0125502104, 0.0544559822, 117.652222],
+    nd: 1.7552,
+    vd: 27.58,
+    PgF: 0.6036,
+    code6: "755276",
+    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data.",
+  },
+  {
+    name: "SF1",
+    vendor: "Schott",
+    B: [1.55912923, 0.284246288, 0.968842926],
+    C: [0.0121481001, 0.0534549042, 112.174809],
+    nd: 1.71736,
+    vd: 29.51,
+    PgF: 0.5984,
+    code6: "717295",
+    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data.",
+  },
+  {
+    name: "N-LAK8",
+    vendor: "Schott",
+    B: [1.33183167, 0.546623206, 1.19084015],
+    C: [0.00620023871, 0.0216465439, 82.5827736],
+    nd: 1.713,
+    vd: 53.83,
+    PgF: 0.545,
+    code6: "713538",
+    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data.",
+  },
+  {
+    name: "N-KZFS5",
+    vendor: "Schott",
+    B: [1.47460789, 0.193584488, 1.26589974],
+    C: [0.00986143816, 0.0445477583, 106.436258],
+    nd: 1.65412,
+    vd: 39.7,
+    PgF: 0.571,
+    code6: "654397",
+    source:
+      "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data. KZFS family — negative ΔPgF, APO-relevant.",
+  },
+  /* ────── Schott expansion (Phase 4 buildout, Apr 2026) ──────
+   * SF2: legacy dense flint (648/339), fills the gap between SF1 and N-BK7 in
+   *   pre-1990 Schott-dominant double-Gauss and Sonnar designs.
+   * N-SK16: dense lanthanum crown (620/603) ubiquitous in Zeiss, classic Nikon,
+   *   Leica Elmarit, and other pre-1990 normal/standard designs (also sold as
+   *   Hoya BACD5 — same glass, different vendor name).
+   * N-KZFS4: KZFS family with negative ΔPgF; APO paired glass for ED crowns.
+   * N-FK51A: fluorcrown with strong positive ΔPgF (+0.034); APO-relevant in
+   *   Zeiss and Nikon telephoto designs.
+   * N-SK11: barium crown used in Carl Zeiss zoom rear groups.
+   * F2: canonical Schott flint (620/364); dense lead flint in vintage designs.
+   * N-LAK22: lanthanum crown (651/559); common in compact zoom rear elements.
+   * All coefficients from SCHOTT Zemax catalog (schott_2017-01-20b.agf) via
+   * refractiveindex.info.
+   */
+  {
+    name: "SF2",
+    vendor: "Schott",
+    B: [1.40301821, 0.231767504, 0.939056586],
+    C: [0.0105795466, 0.0493226978, 112.405955],
+    nd: 1.64769,
+    vd: 33.85,
+    PgF: 0.5886,
+    code6: "648339",
+    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data.",
+  },
+  {
+    name: "N-SK16",
+    vendor: "Schott",
+    B: [1.34317774, 0.241144399, 0.994317969],
+    C: [0.00704687339, 0.0229005, 92.7508526],
+    nd: 1.62041,
+    vd: 60.32,
+    PgF: 0.5412,
+    code6: "620603",
+    source:
+      "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data. Dense lanthanum crown; Hoya BACD5-equivalent.",
+  },
+  {
+    name: "N-KZFS4",
+    vendor: "Schott",
+    B: [1.35055424, 0.197575506, 1.09962992],
+    C: [0.0087628207, 0.0371767201, 90.3866994],
+    nd: 1.61336,
+    vd: 44.49,
+    PgF: 0.559,
+    code6: "613445",
+    source:
+      "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data. KZFS family — negative ΔPgF, APO-relevant.",
+  },
+  {
+    name: "N-FK51A",
+    vendor: "Schott",
+    B: [0.971247817, 0.216901417, 0.904651666],
+    C: [0.00472301995, 0.0153575612, 168.68133],
+    nd: 1.48656,
+    vd: 84.47,
+    PgF: 0.5359,
+    code6: "487845",
+    source:
+      "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data. Fluorcrown with strong +ΔPgF (+0.034); APO-relevant.",
+  },
+  {
+    name: "N-SK11",
+    vendor: "Schott",
+    B: [1.17963631, 0.229817295, 0.935789652],
+    C: [0.00680282081, 0.0219737205, 101.513232],
+    nd: 1.56384,
+    vd: 60.8,
+    PgF: 0.5412,
+    code6: "564608",
+    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data. Barium crown.",
+  },
+  {
+    name: "F2",
+    vendor: "Schott",
+    B: [1.34533359, 0.209073176, 0.937357162],
+    C: [0.00997743871, 0.0470450767, 111.886764],
+    nd: 1.62004,
+    vd: 36.37,
+    PgF: 0.5829,
+    code6: "620364",
+    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data. Legacy lead flint.",
+  },
+  {
+    name: "N-LAK22",
+    vendor: "Schott",
+    B: [1.14229781, 0.535138441, 1.04088385],
+    C: [0.00585778594, 0.0198546147, 100.834017],
+    nd: 1.65113,
+    vd: 55.89,
+    PgF: 0.5467,
+    code6: "651559",
+    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data. Lanthanum crown.",
+  },
+  /* ────── Schott Phase 5 additions (May 2026) ──────
+   * N-SK10 and N-SK14 are barium / lanthanum crowns found in legacy Schott-
+   * dominant designs; SK10 and SK14 are the deprecated pre-2000 names.
+   * Coefficients from Schott Zemax catalog (schott_2017-01-20b.agf) via
+   * refractiveindex.info.
+   */
+  {
+    name: "N-SK10",
+    vendor: "Schott",
+    B: [1.34972093, 0.238587973, 0.9667336],
+    C: [0.00736272269, 0.0253765327, 103.502909],
+    nd: 1.62278,
+    vd: 56.98,
+    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data.",
+  },
+  {
+    name: "N-SK14",
+    vendor: "Schott",
+    B: [0.936155374, 0.594052018, 1.04374583],
+    C: [0.00461716525, 0.016885927, 103.736265],
+    nd: 1.60311,
+    vd: 60.6,
+    code6: "603606",
+    source: "Schott AGF (schott_2017-01-20b.agf), vendor-published Zemax catalog data.",
+  },
+  /* Schott priority additions */
+  {
+    name: "N-SF66",
+    vendor: "Schott",
+    B: [2.0245976, 0.470187196, 2.59970433],
+    C: [0.0147053225, 0.0692998276, 161.817601],
+    nd: 1.92286,
+    vd: 20.88,
+    PgF: 0.6394,
+    code6: "923209",
+    source: "Schott AGF (schott_2017-01-20b.agf) via refractiveindex.info; N-SF66 page.",
+  },
+  {
+    name: "N-LAF21",
+    vendor: "Schott",
+    B: [1.87134529, 0.25078301, 1.22048639],
+    C: [0.0093332228, 0.0345637762, 83.2404866],
+    nd: 1.788,
+    vd: 47.49,
+    PgF: 0.5555,
+    code6: "788475",
+    source: "Schott AGF (schott_2017-01-20b.agf) via refractiveindex.info; N-LAF21 page.",
+  },
+  {
+    name: "SF11",
+    vendor: "Schott",
+    B: [1.73848403, 0.311168974, 1.17490871],
+    C: [0.0136068604, 0.0615960463, 121.922711],
+    nd: 1.78472,
+    vd: 25.76,
+    PgF: 0.6147,
+    code6: "785258",
+    source: "Schott AGF (schott_2017-01-20b.agf) via refractiveindex.info; SF11 page.",
+  },
+  {
+    name: "N-SK2",
+    vendor: "Schott",
+    B: [1.28189012, 0.257738258, 0.96818604],
+    C: [0.0072719164, 0.0242823527, 110.377773],
+    nd: 1.60738,
+    vd: 56.65,
+    PgF: 0.5477,
+    code6: "607567",
+    source: "Schott AGF (schott_2017-01-20b.agf) via refractiveindex.info; N-SK2 page.",
+  },
+  /* Schott unresolved-token addition */
+  {
+    name: "N-FK5",
+    vendor: "Schott",
+    B: [0.844309338, 0.344147824, 0.910790213],
+    C: [0.00475111955, 0.0149814849, 97.8600293],
+    nd: 1.48749,
+    vd: 70.41,
+    PgF: 0.529,
+    code6: "487704",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-FK5 page.",
+  },
+  /* Schott current datasheet additions */
+  {
+    name: "SF56A",
+    vendor: "Schott",
+    B: [1.70579259, 0.344223052, 1.09601828],
+    C: [0.01338747, 0.0579561608, 121.616024],
+    nd: 1.7847,
+    vd: 26.08,
+    PgF: 0.6098,
+    code6: "785261",
+    source: "Schott optical glass datasheet SF56A, public PDF (as of 2014-02-01; media.schott.com content 7428154c).",
+  },
+  {
+    name: "N-BALF4",
+    vendor: "Schott",
+    B: [1.31004128, 0.142038259, 0.964929351],
+    C: [0.007965965, 0.0330672072, 109.19732],
+    nd: 1.57956,
+    vd: 53.87,
+    PgF: 0.552,
+    code6: "580539",
+    source: "Schott optical glass datasheet N-BALF4, public PDF (as of 2014-02-01; media.schott.com content 710baa54).",
+  },
+  /* Schott Phase 9 additions */
+  {
+    name: "N-SK5",
+    vendor: "Schott",
+    B: [0.991463823, 0.495982121, 0.987393925],
+    C: [0.00522730467, 0.0172733646, 98.3594579],
+    nd: 1.58913,
+    vd: 61.27,
+    PgF: 0.54,
+    code6: "589613",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-SK5 page.",
+  },
+  {
+    name: "N-BAF10",
+    vendor: "Schott",
+    B: [1.5851495, 0.143559385, 1.08521269],
+    C: [0.00926681282, 0.0424489805, 105.613573],
+    nd: 1.67003,
+    vd: 47.11,
+    PgF: 0.563,
+    code6: "670471",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-BAF10 page.",
+  },
+  {
+    name: "N-LAF34",
+    vendor: "Schott",
+    B: [1.75836958, 0.313537785, 1.18925231],
+    C: [0.00872810026, 0.0293020832, 85.1780644],
+    nd: 1.7725,
+    vd: 49.62,
+    PgF: 0.5518,
+    code6: "773496",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-LAF34 page.",
+  },
+  {
+    name: "N-LAK14",
+    vendor: "Schott",
+    B: [1.50781212, 0.318866829, 1.14287213],
+    C: [0.00746098727, 0.0242024834, 80.9565165],
+    nd: 1.6968,
+    vd: 55.41,
+    PgF: 0.5427,
+    code6: "697554",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-LAK14 page.",
+  },
+  {
+    name: "F5",
+    vendor: "Schott",
+    B: [1.3104463, 0.19603426, 0.96612977],
+    C: [0.00958633048, 0.0457627627, 115.011883],
+    nd: 1.60342,
+    vd: 38.03,
+    PgF: 0.5795,
+    code6: "603380",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; F5 page.",
+  },
+  {
+    name: "N-KZFS8",
+    vendor: "Schott",
+    B: [1.62693651, 0.24369876, 1.62007141],
+    C: [0.010880863, 0.0494207753, 131.009163],
+    nd: 1.72047,
+    vd: 34.7,
+    PgF: 0.5833,
+    code6: "720347",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-KZFS8 page.",
+  },
+  {
+    name: "SF10",
+    vendor: "Schott",
+    B: [1.61625977, 0.259229334, 1.07762317],
+    C: [0.0127534559, 0.0581983954, 116.60768],
+    nd: 1.72825,
+    vd: 28.41,
+    PgF: 0.6045,
+    code6: "728284",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; SF10 page.",
+  },
+  {
+    name: "N-LAK33B",
+    vendor: "Schott",
+    B: [1.42288601, 0.593661336, 1.1613526],
+    C: [0.00670283452, 0.021941621, 80.7407701],
+    nd: 1.755,
+    vd: 52.3,
+    code6: "755523",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-LAK33B page.",
+  },
+  {
+    name: "N-BALF5",
+    vendor: "Schott",
+    B: [1.28385965, 0.0719300942, 1.05048927],
+    C: [0.00825815975, 0.0441920027, 107.097324],
+    nd: 1.54739,
+    vd: 53.63,
+    code6: "547536",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info.",
+  },
+  {
+    name: "FK3",
+    vendor: "Schott",
+    B: [0.973346627, 0.146642231, 0.679304225],
+    C: [0.00640795469, 0.020565293, 80.4965389],
+    nd: 1.4645,
+    vd: 65.77,
+    code6: "464658",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info.",
+  },
+  {
+    name: "N-SSK5",
+    vendor: "Schott",
+    B: [1.59222659, 0.103520774, 1.05174016],
+    C: [0.00920284626, 0.0423530072, 106.927374],
+    nd: 1.65844,
+    vd: 50.88,
+    PgF: 0.5575,
+    code6: "658509",
+    source: "Schott N-SSK5 datasheet / optical glass collection, Sellmeier constants B1-B3/C1-C3, code 658509.371.",
+  },
+  {
+    name: "BK7G18",
+    vendor: "Schott",
+    B: [1.26538542, 0.0144191073, 1.00323028],
+    C: [0.00813104078, 0.0543303226, 102.821166],
+    nd: 1.51975,
+    vd: 63.58,
+    code6: "520636",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; BK7G18 page.",
+  },
+  {
+    name: "P-SK57Q1",
+    vendor: "Schott",
+    B: [1.30536483, 0.171434328, 1.10117219],
+    C: [0.00736408831, 0.0255786047, 106.72606],
+    nd: 1.586,
+    vd: 59.5,
+    code6: "586595",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; P-SK57Q1 page.",
+  },
+  {
+    name: "N-SF8",
+    vendor: "Schott",
+    B: [1.55075812, 0.209816918, 1.46205491],
+    C: [0.011433834, 0.0582725652, 133.24165],
+    nd: 1.68894,
+    vd: 31.31,
+    PgF: 0.5999,
+    code6: "689313",
+    source: "SCHOTT N-SF8 optical-glass datasheet, official SCHOTT Advanced Optics PDF.",
+  },
+  /* Phase 24 named-token additions (June 2026)
+   * Repeated Schott-name gaps from the coverage queue with source-backed
+   * formula-2 coefficients. E-line patent annotations that cite these families
+   * but preserve n_e values remain explicitly unmatched in lens data so the
+   * d-line Sellmeier catalog is not applied to a different reference line.
+   */
+  {
+    name: "SF5",
+    vendor: "Schott",
+    B: [1.46141885, 0.247713019, 0.949995832],
+    C: [0.0111826126, 0.0508594669, 112.041888],
+    nd: 1.6727,
+    vd: 32.21,
+    PgF: 0.5901,
+    code6: "673322",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; SF5 page.",
+  },
+  {
+    name: "N-SF5",
+    vendor: "Schott",
+    B: [1.52481889, 0.187085527, 1.42729015],
+    C: [0.011254756, 0.0588995392, 129.141675],
+    nd: 1.67271,
+    vd: 32.25,
+    PgF: 0.5969,
+    code6: "673323",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-SF5 page.",
+  },
+  {
+    name: "N-LASF44",
+    vendor: "Schott",
+    B: [1.78897105, 0.38675867, 1.30506243],
+    C: [0.00872506277, 0.0308085023, 92.7743824],
+    nd: 1.8042,
+    vd: 46.5,
+    PgF: 0.5573,
+    code6: "804465",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-LASF44 page.",
+  },
+  {
+    name: "N-LAK9",
+    vendor: "Schott",
+    B: [1.46231905, 0.344399589, 1.15508372],
+    C: [0.00724270156, 0.0243353131, 85.4686868],
+    nd: 1.691,
+    vd: 54.71,
+    PgF: 0.5368,
+    code6: "691547",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-LAK9 page.",
+  },
+  {
+    name: "N-PSK53A",
+    vendor: "Schott",
+    B: [1.38121836, 0.196745645, 0.886089205],
+    C: [0.00706416337, 0.0233251345, 97.4847345],
+    nd: 1.618,
+    vd: 63.39,
+    PgF: 0.5366,
+    code6: "618634",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-PSK53A page.",
+  },
+  {
+    name: "N-LAF2",
+    vendor: "Schott",
+    B: [1.80984227, 0.15729555, 1.0930037],
+    C: [0.0101711622, 0.0442431765, 100.687748],
+    nd: 1.74397,
+    vd: 44.85,
+    PgF: 0.5588,
+    code6: "744449",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-LAF2 page.",
+  },
+  {
+    name: "N-BAK4",
+    vendor: "Schott",
+    B: [1.28834642, 0.132817724, 0.945395373],
+    C: [0.00779980626, 0.0315631177, 105.965875],
+    nd: 1.56883,
+    vd: 55.98,
+    PgF: 0.5431,
+    code6: "569560",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-BAK4 page.",
+  },
+  {
+    name: "N-LAK7",
+    vendor: "Schott",
+    B: [1.23679889, 0.445051837, 1.01745888],
+    C: [0.00610105538, 0.0201388334, 90.638038],
+    nd: 1.6516,
+    vd: 58.52,
+    PgF: 0.5417,
+    code6: "652585",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-LAK7 page.",
+  },
+  {
+    name: "N-BAF4",
+    vendor: "Schott",
+    B: [1.42056328, 0.102721269, 1.14380976],
+    C: [0.00942015382, 0.0531087291, 110.278856],
+    nd: 1.60568,
+    vd: 43.72,
+    PgF: 0.5584,
+    code6: "606437",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-BAF4 page.",
+  },
+  {
+    name: "N-SSK2",
+    vendor: "Schott",
+    B: [1.4306027, 0.153150554, 1.01390904],
+    C: [0.00823982975, 0.0333736841, 106.870822],
+    nd: 1.62229,
+    vd: 53.27,
+    PgF: 0.5542,
+    code6: "622533",
+    source: "Schott Zemax catalog 2017-01-20b via refractiveindex.info; N-SSK2 page.",
+  },
+];
